Moved stuff to mathics.core.character and removed duplicate code
