+++ mathics/core/characters.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
+import re
 
 # Character ranges of letters
 letters = "a-zA-Z\u00c0-\u00d6\u00d8-\u00f6\u00f8-\u0103\u0106\u0107\
@@ -30,1784 +31,2232 @@ letterlikes = "\u0024\u00A1\u00A2\u00A3\u00A5\u00A7\u00A9\u00AB\u00AE\
 \uF82A\uF82B\uF82C\uF82D\uF82E\uF82F\uF830\uF831\uF832\uF833\uFE35\uFE36\
 \uFE37\uFE38"
 
-# All supported longname characters
-named_characters = {
-    "AAcute": "\u00E1",
-    "ABar": "\u0101",
-    "ACup": "\u0103",
-    "ADoubleDot": "\u00E4",
-    "AE": "\u00E6",
-    "AGrave": "\u00E0",
-    "AHat": "\u00E2",
-    "ARing": "\u00E5",
-    "ATilde": "\u00E3",
-    "Akuz": "\uF728",
-    "Aleph": "\u2135",
-    "AliasDelimiter": "\uF764",
-    "AliasIndicator": "\uF768",
-    "AlignmentMarker": "\uF760",
-    "Alpha": "\u03B1",
-    "AltKey": "\uF7D1",
-    "And": "\u2227",
-    "Andy": "\uF729",
-    "Angle": "\u2220",
-    "Angstrom": "\u212B",
-    "AquariusSign": "\u2652",
-    "AriesSign": "\u2648",
-    "AscendingEllipsis": "\u22F0",
-    "AutoLeftMatch": "\uF3A8",
-    "AutoOperand": "\uF3AE",
-    "AutoPlaceholder": "\uF3A4",
-    "AutoRightMatch": "\uF3A9",
-    "AutoSpace": "\uF3AD",
-    "Backslash": "\u2216",
-    "BeamedEighthNote": "\u266B",
-    "BeamedSixteenthNote": "\u266C",
-    "Because": "\u2235",
-    "Bet": "\u2136",
-    "Beta": "\u03B2",
-    "BlackBishop": "\u265D",
-    "BlackKing": "\u265A",
-    "BlackKnight": "\u265E",
-    "BlackPawn": "\u265F",
-    "BlackQueen": "\u265B",
-    "BlackRook": "\u265C",
-    "Breve": "\u02D8",
-    "Bullet": "\u2022",
-    "CAcute": "\u0107",
-    "CCedilla": "\u00E7",
-    "CHacek": "\u010D",
-    "COMPATIBILITYKanjiSpace": "\u3000",
-    "COMPATIBILITYNoBreak": "\uF3A2",
-    "CancerSign": "\u264B",
-    "Cap": "\u2322",
-    "CapitalAAcute": "\u00C1",
-    "CapitalABar": "\u0100",
-    "CapitalACup": "\u0102",
-    "CapitalADoubleDot": "\u00C4",
-    "CapitalAE": "\u00C6",
-    "CapitalAGrave": "\u00C0",
-    "CapitalAHat": "\u00C2",
-    "CapitalARing": "\u00C5",
-    "CapitalATilde": "\u00C3",
-    "CapitalAlpha": "\u0391",
-    "CapitalBeta": "\u0392",
-    "CapitalCAcute": "\u0106",
-    "CapitalCCedilla": "\u00C7",
-    "CapitalCHacek": "\u010C",
-    "CapitalChi": "\u03A7",
-    "CapitalDHacek": "\u010E",
-    "CapitalDelta": "\u0394",
-    "CapitalDifferentialD": "\uF74B",
-    "CapitalDigamma": "\u03DC",
-    "CapitalEAcute": "\u00C9",
-    "CapitalEBar": "\u0112",
-    "CapitalECup": "\u0114",
-    "CapitalEDoubleDot": "\u00CB",
-    "CapitalEGrave": "\u00C8",
-    "CapitalEHacek": "\u011A",
-    "CapitalEHat": "\u00CA",
-    "CapitalEpsilon": "\u0395",
-    "CapitalEta": "\u0397",
-    "CapitalEth": "\u00D0",
-    "CapitalGamma": "\u0393",
-    "CapitalIAcute": "\u00CD",
-    "CapitalICup": "\u012C",
-    "CapitalIDoubleDot": "\u00CF",
-    "CapitalIGrave": "\u00CC",
-    "CapitalIHat": "\u00CE",
-    "CapitalIota": "\u0399",
-    "CapitalKappa": "\u039A",
-    "CapitalKoppa": "\u03DE",
-    "CapitalLSlash": "\u0141",
-    "CapitalLambda": "\u039B",
-    "CapitalMu": "\u039C",
-    "CapitalNHacek": "\u0147",
-    "CapitalNTilde": "\u00D1",
-    "CapitalNu": "\u039D",
-    "CapitalOAcute": "\u00D3",
-    "CapitalODoubleAcute": "\u0150",
-    "CapitalODoubleDot": "\u00D6",
-    "CapitalOE": "\u0152",
-    "CapitalOGrave": "\u00D2",
-    "CapitalOHat": "\u00D4",
-    "CapitalOSlash": "\u00D8",
-    "CapitalOTilde": "\u00D5",
-    "CapitalOmega": "\u03A9",
-    "CapitalOmicron": "\u039F",
-    "CapitalPhi": "\u03A6",
-    "CapitalPi": "\u03A0",
-    "CapitalPsi": "\u03A8",
-    "CapitalRHacek": "\u0158",
-    "CapitalRho": "\u03A1",
-    "CapitalSHacek": "\u0160",
-    "CapitalSampi": "\u03E0",
-    "CapitalSigma": "\u03A3",
-    "CapitalStigma": "\u03DA",
-    "CapitalTHacek": "\u0164",
-    "CapitalTau": "\u03A4",
-    "CapitalTheta": "\u0398",
-    "CapitalThorn": "\u00DE",
-    "CapitalUAcute": "\u00DA",
-    "CapitalUDoubleAcute": "\u0170",
-    "CapitalUDoubleDot": "\u00DC",
-    "CapitalUGrave": "\u00D9",
-    "CapitalUHat": "\u00DB",
-    "CapitalURing": "\u016E",
-    "CapitalUpsilon": "\u03A5",
-    "CapitalXi": "\u039E",
-    "CapitalYAcute": "\u00DD",
-    "CapitalZHacek": "\u017D",
-    "CapitalZeta": "\u0396",
-    "CapricornSign": "\u2651",
-    "Cedilla": "\u00B8",
-    "Cent": "\u00A2",
-    "CenterDot": "\u00B7",
-    "CenterEllipsis": "\u22EF",
-    "CheckedBox": "\u2612",
-    "Checkmark": "\u2713",
-    "CheckedBox": "\u2612",
-    "Chi": "\u03C7",
-    "CircleDot": "\u2299",
-    "CircleMinus": "\u2296",
-    "CirclePlus": "\u2295",
-    "CircleTimes": "\u2297",
-    "ClockwiseContourIntegral": "\u2232",
-    "CloseCurlyDoubleQuote": "\u201D",
-    "CloseCurlyQuote": "\u2019",
-    "CloverLeaf": "\u2318",
-    "ClubSuit": "\u2663",
-    "Colon": "\u2236",
-    "CommandKey": "\uF76A",
-    "Conditioned": "\uF3D3",
-    "Congruent": "\u2261",
-    "Conjugate": "\uF3C8",
-    "ConjugateTranspose": "\uF3C9",
-    "ConstantC": "\uF7DA",
-    "Continuation": "\uF3B1",
-    "ContinuedFractionK": "\uF3D9",
-    "ContourIntegral": "\u222E",
-    "ControlKey": "\uF763",
-    "Coproduct": "\u2210",
-    "Copyright": "\u00A9",
-    "CounterClockwiseContourIntegral": "\u2233",
-    "Cross": "\uF4A0",
-    "CubeRoot": "\u221B",
-    "CupCap": "\u224D",
-    "Cup": "\u2323",
-    "Curl": "\uF3D8",
-    "CurlyCapitalUpsilon": "\u03D2",
-    "CurlyEpsilon": "\u03B5",
-    "CurlyKappa": "\u03F0",
-    "CurlyPhi": "\u03C6",
-    "CurlyPi": "\u03D6",
-    "CurlyRho": "\u03F1",
-    "CurlyTheta": "\u03D1",
-    "Currency": "\u00A4",
-    "DHacek": "\u010F",
-    "Dagger": "\u2020",
-    "Dalet": "\u2138",
-    "Dash": "\u2013",
-    "Degree": "\u00B0",
-    "Del": "\u2207",
-    "DeleteKey": "\uF7D0",
-    "Delta": "\u03B4",
-    "DescendingEllipsis": "\u22F1",
-    "Diameter": "\u2300",
-    "Diamond": "\u22C4",
-    "DiamondSuit": "\u2662",
-    "DifferenceDelta": "\u2206",
-    "DifferentialD": "\uF74C",
-    "Digamma": "\u03DD",
-    "DirectedEdge": "\uF3D5",
-    "DiscreteRatio": "\uF4A4",
-    "DiscreteShift": "\uF4A3",
-    "DiscretionaryHyphen": "\u00AD",
-    "DiscretionaryLineSeparator": "\uF76E",
-    "DiscretionaryPageBreakAbove": "\uF3BF",
-    "DiscretionaryPageBreakBelow": "\uF3C6",
-    "DiscretionaryParagraphSeparator": "\uF76F",
-    "Distributed": "\uF3D2",
-    "Divergence": "\uF3D7",
-    "Divide": "\u00F7",
-    "Divides": "\u2223",
-    "DivisionSlash": "\u2215",
-    "DotEqual": "\u2250",
-    "DotlessI": "\u0131",
-    "DotlessJ": "\uF700",
-    "DottedSquare": "\uF751",
-    "DoubleContourIntegral": "\u222F",
-    "DoubleDagger": "\u2021",
-    "DoubleDot": "\u00A8",
-    "DoubleDownArrow": "\u21D3",
-    "DoubleLeftArrow": "\u21D0",
-    "DoubleLeftRightArrow": "\u21D4",
-    "DoubleLeftTee": "\u2AE4",
-    "DoubleLongLeftArrow": "\u27F8",
-    "DoubleLongLeftRightArrow": "\u27FA",
-    "DoubleLongRightArrow": "\u27F9",
-    "DoublePrime": "\u2033",
-    "DoubleRightArrow": "\u21D2",
-    "DoubleRightTee": "\u22A8",
-    "DoubleStruckA": "\uF6E6",
-    "DoubleStruckB": "\uF6E7",
-    "DoubleStruckC": "\uF6E8",
-    "DoubleStruckCapitalA": "\uF7A4",
-    "DoubleStruckCapitalB": "\uF7A5",
-    "DoubleStruckCapitalC": "\uF7A6",
-    "DoubleStruckCapitalD": "\uF7A7",
-    "DoubleStruckCapitalE": "\uF7A8",
-    "DoubleStruckCapitalF": "\uF7A9",
-    "DoubleStruckCapitalG": "\uF7AA",
-    "DoubleStruckCapitalH": "\uF7AB",
-    "DoubleStruckCapitalI": "\uF7AC",
-    "DoubleStruckCapitalJ": "\uF7AD",
-    "DoubleStruckCapitalK": "\uF7AE",
-    "DoubleStruckCapitalL": "\uF7AF",
-    "DoubleStruckCapitalM": "\uF7B0",
-    "DoubleStruckCapitalN": "\uF7B1",
-    "DoubleStruckCapitalO": "\uF7B2",
-    "DoubleStruckCapitalP": "\uF7B3",
-    "DoubleStruckCapitalQ": "\uF7B4",
-    "DoubleStruckCapitalR": "\uF7B5",
-    "DoubleStruckCapitalS": "\uF7B6",
-    "DoubleStruckCapitalT": "\uF7B7",
-    "DoubleStruckCapitalU": "\uF7B8",
-    "DoubleStruckCapitalV": "\uF7B9",
-    "DoubleStruckCapitalW": "\uF7BA",
-    "DoubleStruckCapitalX": "\uF7BB",
-    "DoubleStruckCapitalY": "\uF7BC",
-    "DoubleStruckCapitalZ": "\uF7BD",
-    "DoubleStruckD": "\uF6E9",
-    "DoubleStruckE": "\uF6EA",
-    "DoubleStruckEight": "\uF7E3",
-    "DoubleStruckF": "\uF6EB",
-    "DoubleStruckFive": "\uF7E0",
-    "DoubleStruckFour": "\uF7DF",
-    "DoubleStruckG": "\uF6EC",
-    "DoubleStruckH": "\uF6ED",
-    "DoubleStruckI": "\uF6EE",
-    "DoubleStruckJ": "\uF6EF",
-    "DoubleStruckK": "\uF6F0",
-    "DoubleStruckL": "\uF6F1",
-    "DoubleStruckM": "\uF6F2",
-    "DoubleStruckN": "\uF6F3",
-    "DoubleStruckNine": "\uF7E4",
-    "DoubleStruckO": "\uF6F4",
-    "DoubleStruckOne": "\uF7DC",
-    "DoubleStruckP": "\uF6F5",
-    "DoubleStruckQ": "\uF6F6",
-    "DoubleStruckR": "\uF6F7",
-    "DoubleStruckS": "\uF6F8",
-    "DoubleStruckSeven": "\uF7E2",
-    "DoubleStruckSix": "\uF7E1",
-    "DoubleStruckT": "\uF6F9",
-    "DoubleStruckThree": "\uF7DE",
-    "DoubleStruckTwo": "\uF7DD",
-    "DoubleStruckU": "\uF6FA",
-    "DoubleStruckV": "\uF6FB",
-    "DoubleStruckW": "\uF6FC",
-    "DoubleStruckX": "\uF6FD",
-    "DoubleStruckY": "\uF6FE",
-    "DoubleStruckZ": "\uF6FF",
-    "DoubleStruckZero": "\uF7DB",
-    "DoubleUpArrow": "\u21D1",
-    "DoubleUpDownArrow": "\u21D5",
-    "DoubleVerticalBar": "\u2225",
-    "DoubledGamma": "\uF74A",
-    "DoubledPi": "\uF749",
-    "DownArrow": "\u2193",
-    "DownArrowBar": "\u2913",
-    "DownArrowUpArrow": "\u21F5",
-    "DownBreve": "\uF755",
-    "DownExclamation": "\u00A1",
-    "DownLeftRightVector": "\u2950",
-    "DownLeftTeeVector": "\u295E",
-    "DownLeftVector": "\u21BD",
-    "DownLeftVectorBar": "\u2956",
-    "DownPointer": "\u25BE",
-    "DownQuestion": "\u00BF",
-    "DownRightTeeVector": "\u295F",
-    "DownRightVector": "\u21C1",
-    "DownRightVectorBar": "\u2957",
-    "DownTee": "\u22A4",
-    "DownTeeArrow": "\u21A7",
-    "EAcute": "\u00E9",
-    "EBar": "\u0113",
-    "ECup": "\u0115",
-    "EDoubleDot": "\u00EB",
-    "EGrave": "\u00E8",
-    "EHacek": "\u011B",
-    "EHat": "\u00EA",
-    "Earth": "\u2641",
-    "EighthNote": "\u266A",
-    "Element": "\u2208",
-    "Ellipsis": "\u2026",
-    "EmptyCircle": "\u25CB",
-    "EmptyDiamond": "\u25C7",
-    "EmptyDownTriangle": "\u25BD",
-    "EmptyRectangle": "\u25AF",
-    "EmptySet": "\u2205",
-    "EmptySmallCircle": "\u25E6",
-    "EmptySmallSquare": "\u25FB",
-    "EmptySquare": "\u25A1",
-    "EmptyUpTriangle": "\u25B3",
-    "EmptyVerySmallSquare": "\u25AB",
-    "EnterKey": "\uF7D4",
-    "EntityEnd": "\uF3B9",
-    "EntityStart": "\uF3B8",
-    "Epsilon": "\u03F5",
-    "Equal": "\uF431",
-    "EqualTilde": "\u2242",
-    "Equilibrium": "\u21CC",
-    "Equivalent": "\u29E6",
-    "ErrorIndicator": "\uF767",
-    "EscapeKey": "\uF769",
-    "Eta": "\u03B7",
-    "Eth": "\u00F0",
-    "Euro": "\u20AC",
-    "Exists": "\u2203",
-    "ExpectationE": "\uF3DD",
-    "ExponentialE": "\uF74D",
-    "FiLigature": "\uFB01",
-    "FilledCircle": "\u25CF",
-    "FilledDiamond": "\u25C6",
-    "FilledDownTriangle": "\u25BC",
-    "FilledLeftTriangle": "\u25C0",
-    "FilledRectangle": "\u25AE",
-    "FilledRightTriangle": "\u25B6",
-    "FilledSmallCircle": "\uF750",
-    "FilledSmallSquare": "\u25FC",
-    "FilledSquare": "\u25A0",
-    "FilledUpTriangle": "\u25B2",
-    "FilledVerySmallSquare": "\u25AA",
-    "FinalSigma": "\u03C2",
-    "FirstPage": "\uF7FA",
-    "FivePointedStar": "\u2605",
-    "FlLigature": "\uFB02",
-    "Flat": "\u266D",
-    "Florin": "\u0192",
-    "ForAll": "\u2200",
-    "FormalA": "\uF800",
-    "FormalAlpha": "\uF854",
-    "FormalB": "\uF801",
-    "FormalBeta": "\uF855",
-    "FormalC": "\uF802",
-    "FormalCapitalA": "\uF81A",
-    "FormalCapitalAlpha": "\uF834",
-    "FormalCapitalB": "\uF81B",
-    "FormalCapitalBeta": "\uF835",
-    "FormalCapitalC": "\uF81C",
-    "FormalCapitalChi": "\uF84A",
-    "FormalCapitalD": "\uF81D",
-    "FormalCapitalDelta": "\uF837",
-    "FormalCapitalDigamma": "\uF87F",
-    "FormalCapitalE": "\uF81E",
-    "FormalCapitalEpsilon": "\uF838",
-    "FormalCapitalEta": "\uF83A",
-    "FormalCapitalF": "\uF81F",
-    "FormalCapitalG": "\uF820",
-    "FormalCapitalGamma": "\uF836",
-    "FormalCapitalH": "\uF821",
-    "FormalCapitalI": "\uF822",
-    "FormalCapitalIota": "\uF83C",
-    "FormalCapitalJ": "\uF823",
-    "FormalCapitalK": "\uF824",
-    "FormalCapitalKappa": "\uF83D",
-    "FormalCapitalKoppa": "\uF881",
-    "FormalCapitalL": "\uF825",
-    "FormalCapitalLambda": "\uF83E",
-    "FormalCapitalM": "\uF826",
-    "FormalCapitalMu": "\uF83F",
-    "FormalCapitalN": "\uF827",
-    "FormalCapitalNu": "\uF840",
-    "FormalCapitalO": "\uF828",
-    "FormalCapitalOmega": "\uF84C",
-    "FormalCapitalOmicron": "\uF842",
-    "FormalCapitalP": "\uF829",
-    "FormalCapitalPhi": "\uF849",
-    "FormalCapitalPi": "\uF843",
-    "FormalCapitalPsi": "\uF84B",
-    "FormalCapitalQ": "\uF82A",
-    "FormalCapitalR": "\uF82B",
-    "FormalCapitalRho": "\uF844",
-    "FormalCapitalS": "\uF82C",
-    "FormalCapitalSampi": "\uF883",
-    "FormalCapitalSigma,": "\uF846",
-    "FormalCapitalStigma": "\uF87D",
-    "FormalCapitalT": "\uF82D",
-    "FormalCapitalTau": "\uF847",
-    "FormalCapitalTheta": "\uF83B",
-    "FormalCapitalU": "\uF82E",
-    "FormalCapitalUpsilon": "\uF848",
-    "FormalCapitalV": "\uF82F",
-    "FormalCapitalW": "\uF830",
-    "FormalCapitalX": "\uF831",
-    "FormalCapitalXi": "\uF841",
-    "FormalCapitalY": "\uF832",
-    "FormalCapitalZ": "\uF833",
-    "FormalCapitalZeta": "\uF839",
-    "FormalChi": "\uF86A",
-    "FormalCurlyCapitalUpsilon": "\uF875",
-    "FormalCurlyEpsilon": "\uF858",
-    "FormalCurlyKappa": "\uF885",
-    "FormalCurlyPhi": "\uF869",
-    "FormalCurlyPi": "\uF879",
-    "FormalCurlyRho": "\uF886",
-    "FormalCurlyTheta": "\uF874",
-    "FormalD": "\uF803",
-    "FormalDelta": "\uF857",
-    "FormalDigamma": "\uF880",
-    "FormalE": "\uF804",
-    "FormalEpsilon": "\uF88A",
-    "FormalEta": "\uF85A",
-    "FormalF": "\uF805",
-    "FormalFinalSigma": "\uF865",
-    "FormalG": "\uF806",
-    "FormalGamma": "\uF856",
-    "FormalH": "\uF807",
-    "FormalI": "\uF808",
-    "FormalIota": "\uF85C",
-    "FormalJ": "\uF809",
-    "FormalK": "\uF80A",
-    "FormalKappa": "\uF85D",
-    "FormalKoppa": "\uF882",
-    "FormalL": "\uF80B",
-    "FormalLambda": "\uF85E",
-    "FormalM": "\uF80C",
-    "FormalMu": "\uF85F",
-    "FormalN": "\uF80D",
-    "FormalNu": "\uF860",
-    "FormalO": "\uF80E",
-    "FormalOmega": "\uF86C",
-    "FormalOmicron": "\uF862",
-    "FormalP": "\uF80F",
-    "FormalPhi": "\uF878",
-    "FormalPi": "\uF863",
-    "FormalPsi": "\uF86B",
-    "FormalQ": "\uF810",
-    "FormalR": "\uF811",
-    "FormalRho": "\uF864",
-    "FormalS": "\uF812",
-    "FormalSampi": "\uF884",
-    "FormalSigma": "\uF866",
-    "FormalStigma": "\uF87E",
-    "FormalT": "\uF813",
-    "FormalTau": "\uF867",
-    "FormalTheta": "\uF85B",
-    "FormalU": "\uF814",
-    "FormalUpsilon": "\uF868",
-    "FormalV": "\uF815",
-    "FormalW": "\uF816",
-    "FormalX": "\uF817",
-    "FormalXi": "\uF861",
-    "FormalY": "\uF818",
-    "FormalZ": "\uF819",
-    "FormalZeta": "\uF859",
-    "FreakedSmiley": "\uF721",
-    "FreeformPrompt": "\uF351",
-    "Function": "\uF4A1",
-    "Gamma": "\u03B3",
-    "GeminiSign": "\u264A",
-    "Gimel": "\u2137",
-    "GothicA": "\uF6CC",
-    "GothicB": "\uF6CD",
-    "GothicC": "\uF6CE",
-    "GothicCapitalA": "\uF78A",
-    "GothicCapitalB": "\uF78B",
-    "GothicCapitalC": "\u212D",
-    "GothicCapitalD": "\uF78D",
-    "GothicCapitalE": "\uF78E",
-    "GothicCapitalF": "\uF78F",
-    "GothicCapitalG": "\uF790",
-    "GothicCapitalH": "\u210C",
-    "GothicCapitalI": "\u2111",
-    "GothicCapitalJ": "\uF793",
-    "GothicCapitalK": "\uF794",
-    "GothicCapitalL": "\uF795",
-    "GothicCapitalM": "\uF796",
-    "GothicCapitalN": "\uF797",
-    "GothicCapitalO": "\uF798",
-    "GothicCapitalP": "\uF799",
-    "GothicCapitalQ": "\uF79A",
-    "GothicCapitalR": "\u211C",
-    "GothicCapitalS": "\uF79C",
-    "GothicCapitalT": "\uF79D",
-    "GothicCapitalU": "\uF79E",
-    "GothicCapitalV": "\uF79F",
-    "GothicCapitalW": "\uF7A0",
-    "GothicCapitalX": "\uF7A1",
-    "GothicCapitalY": "\uF7A2",
-    "GothicCapitalZ": "\u2128",
-    "GothicD": "\uF6CF",
-    "GothicE": "\uF6D0",
-    "GothicEight": "\uF7ED",
-    "GothicF": "\uF6D1",
-    "GothicFive": "\uF7EA",
-    "GothicFour": "\uF7E9",
-    "GothicG": "\uF6D2",
-    "GothicH": "\uF6D3",
-    "GothicI": "\uF6D4",
-    "GothicJ": "\uF6D5",
-    "GothicK": "\uF6D6",
-    "GothicL": "\uF6D7",
-    "GothicM": "\uF6D8",
-    "GothicN": "\uF6D9",
-    "GothicNine": "\uF7EF",
-    "GothicO": "\uF6DA",
-    "GothicOne": "\uF7E6",
-    "GothicP": "\uF6DB",
-    "GothicQ": "\uF6DC",
-    "GothicR": "\uF6DD",
-    "GothicS": "\uF6DE",
-    "GothicSeven": "\uF7EC",
-    "GothicSix": "\uF7EB",
-    "GothicT": "\uF6DF",
-    "GothicThree": "\uF7E8",
-    "GothicTwo": "\uF7E7",
-    "GothicU": "\uF6E0",
-    "GothicV": "\uF6E1",
-    "GothicW": "\uF6E2",
-    "GothicX": "\uF6E3",
-    "GothicY": "\uF6E4",
-    "GothicZ": "\uF6E5",
-    "GothicZero": "\uF7E5",
-    "Gradient": "\uF3D6",
-    "GrayCircle": "\uF753",
-    "GraySquare": "\uF752",
-    "GreaterEqual": "\u2265",
-    "GreaterEqualLess": "\u22DB",
-    "GreaterFullEqual": "\u2267",
-    "GreaterGreater": "\u226B",
-    "GreaterLess": "\u2277",
-    "GreaterSlantEqual": "\u2A7E",
-    "GreaterTilde": "\u2273",
-    "HBar": "\u210F",
-    "Hacek": "\u02C7",
-    "HappySmiley": "\u263A",
-    "HeartSuit": "\u2661",
-    "HermitianConjugate": "\uF3CE",
-    "HorizontalLine": "\u2500",
-    "HumpDownHump": "\u224E",
-    "HumpEqual": "\u224F",
-    "Hyphen": "\u2010",
-    "IAcute": "\u00ED",
-    "ICup": "\u012D",
-    "IDoubleDot": "\u00EF",
-    "IGrave": "\u00EC",
-    "IHat": "\u00EE",
-    "ImaginaryI": "\uF74E",
-    "ImaginaryJ": "\uF74F",
-    "ImplicitPlus": "\uF39E",
-    "Implies": "\uF523",
-    "IndentingNewLine": "\uF3A3",
-    "Infinity": "\u221E",
-    "InlinePart": "\uF51E",
-    "Integral": "\u222B",
-    "Intersection": "\u22C2",
-    "InvisibleApplication": "\uF76D",
-    "InvisibleComma": "\uF765",
-    "InvisiblePostfixScriptBase": "\uF3B4",
-    "InvisiblePrefixScriptBase": "\uF3B3",
-    "InvisibleSpace": "\uF360",
-    "InvisibleTimes": "\u2062",
-    "Iota": "\u03B9",
-    "Jupiter": "\u2643",
-    "Kappa": "\u03BA",
-    "KernelIcon": "\uF756",
-    "KeyBar": "\uF7D3",
-    "Koppa": "\u03DF",
-    "LSlash": "\u0142",
-    "Lambda": "\u03BB",
-    "Laplacian": "\u2206",
-    "LastPage": "\uF7FB",
-    "LeftAngleBracket": "\u2329",
-    "LeftArrow": "\u2190",
-    "LeftArrowBar": "\u21E4",
-    "LeftArrowRightArrow": "\u21C6",
-    "LeftAssociation": "\uF113",
-    "LeftBracketingBar": "\uF603",
-    "LeftCeiling": "\u2308",
-    "LeftDoubleBracket": "\u301A",
-    "LeftDoubleBracketingBar": "\uF605",
-    "LeftDownTeeVector": "\u2961",
-    "LeftDownVector": "\u21C3",
-    "LeftDownVectorBar": "\u2959",
-    "LeftFloor": "\u230A",
-    "LeftGuillemet": "\u00AB",
-    "LeftModified": "\uF76B",
-    "LeftPointer": "\u25C2",
-    "LeftRightArrow": "\u2194",
-    "LeftRightVector": "\u294E",
-    "LeftSkeleton": "\uF761",
-    "LeftTee": "\u22A3",
-    "LeftTeeArrow": "\u21A4",
-    "LeftTeeVector": "\u295A",
-    "LeftTriangle": "\u22B2",
-    "LeftTriangleBar": "\u29CF",
-    "LeftTriangleEqual": "\u22B4",
-    "LeftUpDownVector": "\u2951",
-    "LeftUpTeeVector": "\u2960",
-    "LeftUpVector": "\u21BF",
-    "LeftUpVectorBar": "\u2958",
-    "LeftVector": "\u21BC",
-    "LeftVectorBar": "\u2952",
-    "LeoSign": "\u264C",
-    "LessEqual": "\u2264",
-    "LessEqualGreater": "\u22DA",
-    "LessFullEqual": "\u2266",
-    "LessGreater": "\u2276",
-    "LessLess": "\u226A",
-    "LessSlantEqual": "\u2A7D",
-    "LessTilde": "\u2272",
-    "LetterSpace": "\uF754",
-    "LibraSign": "\u264E",
-    "LightBulb": "\uF723",
-    "Limit": "\uF438",
-    "LineSeparator": "\u2028",
-    "LongDash": "\u2014",
-    "LongEqual": "\uF7D9",
-    "LongLeftArrow": "\u27F5",
-    "LongLeftRightArrow": "\u27F7",
-    "LongRightArrow": "\u27F6",
-    "LowerLeftArrow": "\u2199",
-    "LowerRightArrow": "\u2198",
-    "Mars": "\u2642",
-    "MathematicaIcon": "\uF757",
-    "MaxLimit": "\uF439",
-    "MeasuredAngle": "\u2221",
-    "MediumSpace": "\u205F",
-    "Mercury": "\u263F",
-    "Mho": "\u2127",
-    "Micro": "\u00B5",
-    "MinLimit": "\uF43A",
-    "Minus": "\u2212",
-    "MinusPlus": "\u2213",
-    "Mod1Key": "\uF7D6",
-    "Mod2Key": "\uF7D7",
-    "Moon": "\u263E",
-    "Mu": "\u03BC",
-    "NHacek": "\u0148",
-    "NTilde": "\u00F1",
-    "Nand": "\u22BC",
-    "Natural": "\u266E",
-    "NegativeMediumSpace": "\uF383",
-    "NegativeThickSpace": "\uF384",
-    "NegativeThinSpace": "\uF382",
-    "NegativeVeryThinSpace": "\uF380",
-    "Neptune": "\u2646",
-    "NestedGreaterGreater": "\u2AA2",
-    "NestedLessLess": "\u2AA1",
-    "NeutralSmiley": "\uF722",
-    "NewLine": "\u000A",
-    "NoBreak": "\u2060",
-    "NonBreakingSpace": "\u00A0",
-    "Nor": "\u22BD",
-    "Not": "\u00AC",
-    "NotCongruent": "\u2262",
-    "NotCupCap": "\u226D",
-    "NotDoubleVerticalBar": "\u2226",
-    "NotElement": "\u2209",
-    "NotEqual": "\u2260",
-    "NotEqualTilde": "\uF400",
-    "NotExists": "\u2204",
-    "NotGreater": "\u226F",
-    "NotGreaterEqual": "\u2271",
-    "NotGreaterFullEqual": "\u2269",
-    "NotGreaterGreater": "\uF427",
-    "NotGreaterLess": "\u2279",
-    "NotGreaterSlantEqual": "\uF429",
-    "NotGreaterTilde": "\u2275",
-    "NotHumpDownHump": "\uF402",
-    "NotHumpEqual": "\uF401",
-    "NotLeftTriangle": "\u22EA",
-    "NotLeftTriangleBar": "\uF412",
-    "NotLeftTriangleEqual": "\u22EC",
-    "NotLess": "\u226E",
-    "NotLessEqual": "\u2270",
-    "NotLessFullEqual": "\u2268",
-    "NotLessGreater": "\u2278",
-    "NotLessLess": "\uF422",
-    "NotLessSlantEqual": "\uF424",
-    "NotLessTilde": "\u2274",
-    "NotNestedGreaterGreater": "\uF428",
-    "NotNestedLessLess": "\uF423",
-    "NotPrecedes": "\u2280",
-    "NotPrecedesEqual": "\uF42B",
-    "NotPrecedesSlantEqual": "\u22E0",
-    "NotPrecedesTilde": "\u22E8",
-    "NotReverseElement": "\u220C",
-    "NotRightTriangle": "\u22EB",
-    "NotRightTriangleBar": "\uF413",
-    "NotRightTriangleEqual": "\u22ED",
-    "NotSquareSubset": "\uF42E",
-    "NotSquareSubsetEqual": "\u22E2",
-    "NotSquareSuperset": "\uF42F",
-    "NotSquareSupersetEqual": "\u22E3",
-    "NotSubset": "\u2284",
-    "NotSubsetEqual": "\u2288",
-    "NotSucceeds": "\u2281",
-    "NotSucceedsEqual": "\uF42D",
-    "NotSucceedsSlantEqual": "\u22E1",
-    "NotSucceedsTilde": "\u22E9",
-    "NotSuperset": "\u2285",
-    "NotSupersetEqual": "\u2289",
-    "NotTilde": "\u2241",
-    "NotTildeEqual": "\u2244",
-    "NotTildeFullEqual": "\u2247",
-    "NotTildeTilde": "\u2249",
-    "NotVerticalBar": "\u2224",
-    "Nu": "\u03BD",
-    "Null": "\uF3A0",
-    "NumberSign": "\uF724",
-    "OAcute": "\u00F3",
-    "ODoubleAcute": "\u0151",
-    "ODoubleDot": "\u00F6",
-    "OE": "\u0153",
-    "OGrave": "\u00F2",
-    "OHat": "\u00F4",
-    "OSlash": "\u00F8",
-    "OTilde": "\u00F5",
-    "Omega": "\u03C9",
-    "Omicron": "\u03BF",
-    "OpenCurlyDoubleQuote": "\u201C",
-    "OpenCurlyQuote": "\u2018",
-    "OptionKey": "\uF7D2",
-    "Or": "\u2228",
-    "OverBrace": "\uFE37",
-    "OverBracket": "\u23B4",
-    "OverParenthesis": "\uFE35",
-    "PageBreakAbove": "\uF3BD",
-    "PageBreakBelow": "\uF3BE",
-    "Paragraph": "\u00B6",
-    "ParagraphSeparator": "\u2029",
-    "PartialD": "\u2202",
-    "PermutationProduct": "\uF3DE",
-    "Perpendicular": "\u27C2",
-    "Phi": "\u03D5",
-    "Pi": "\u03C0",
-    "Piecewise": "\uF361",
-    "PiscesSign": "\u2653",
-    "Placeholder": "\uF528",
-    "PlusMinus": "\u00B1",
-    "Pluto": "\u2647",
-    "Precedes": "\u227A",
-    "PrecedesEqual": "\u2AAF",
-    "PrecedesSlantEqual": "\u227C",
-    "PrecedesTilde": "\u227E",
-    "Prime": "\u2032",
-    "ProbabilityPr": "\uF3DC",
-    "Product": "\u220F",
-    "Proportion": "\u2237",
-    "Proportional": "\u221D",
-    "Psi": "\u03C8",
-    "QuarterNote": "\u2669",
-    "RHacek": "\u0159",
-    "RawAmpersand": "\u0026",
-    "RawAt": "\u0040",
-    "RawBackquote": "\u0060",
-    "RawBackslash": "\u005C",
-    "RawColon": "\u003A",
-    "RawComma": "\u002C",
-    "RawDash": "\u002D",
-    "RawDollar": "\u0024",
-    "RawDot": "\u002E",
-    "RawDoubleQuote": "\u0022",
-    "RawEqual": "\u003D",
-    "RawEscape": "\u001B",
-    "RawExclamation": "\u0021",
-    "RawGreater": "\u003E",
-    "RawLeftBrace": "\u007B",
-    "RawLeftBracket": "\u005B",
-    "RawLeftParenthesis": "\u0028",
-    "RawLess": "\u003C",
-    "RawNumberSign": "\u0023",
-    "RawPercent": "\u0025",
-    "RawPlus": "\u002B",
-    "RawQuestion": "\u003F",
-    "RawQuote": "\u0027",
-    "RawReturn": "\u000D",
-    "RawRightBrace": "\u007D",
-    "RawRightBracket": "\u005D",
-    "RawRightParenthesis": "\u0029",
-    "RawSemicolon": "\u003B",
-    "RawSlash": "\u002F",
-    "RawSpace": "\u0020",
-    "RawStar": "\u002A",
-    "RawTab": "\u0009",
-    "RawTilde": "\u007E",
-    "RawUnderscore": "\u005F",
-    "RawVerticalBar": "\u007C",
-    "RawWedge": "\u005E",
-    "RegisteredTrademark": "\u00AE",
-    "ReturnIndicator": "\u21B5",
-    "ReturnKey": "\uF766",
-    "ReverseDoublePrime": "\u2036",
-    "ReverseElement": "\u220B",
-    "ReverseEquilibrium": "\u21CB",
-    "ReversePrime": "\u2035",
-    "ReverseUpEquilibrium": "\u296F",
-    "Rho": "\u03C1",
-    "RightAngle": "\u221F",
-    "RightAngleBracket": "\u232A",
-    "RightArrow": "\u2192",
-    "RightArrowBar": "\u21E5",
-    "RightArrowLeftArrow": "\u21C4",
-    "RightAssociation": "\uF114",
-    "RightBracketingBar": "\uF604",
-    "RightCeiling": "\u2309",
-    "RightDoubleBracket": "\u301B",
-    "RightDoubleBracketingBar": "\uF606",
-    "RightDownTeeVector": "\u295D",
-    "RightDownVector": "\u21C2",
-    "RightDownVectorBar": "\u2955",
-    "RightFloor": "\u230B",
-    "RightGuillemet": "\u00BB",
-    "RightModified": "\uF76C",
-    "RightPointer": "\u25B8",
-    "RightSkeleton": "\uF762",
-    "RightTee": "\u22A2",
-    "RightTeeArrow": "\u21A6",
-    "RightTeeVector": "\u295B",
-    "RightTriangle": "\u22B3",
-    "RightTriangleBar": "\u29D0",
-    "RightTriangleEqual": "\u22B5",
-    "RightUpDownVector": "\u294F",
-    "RightUpTeeVector": "\u295C",
-    "RightUpVector": "\u21BE",
-    "RightUpVectorBar": "\u2954",
-    "RightVector": "\u21C0",
-    "RightVectorBar": "\u2953",
-    "RoundImplies": "\u2970",
-    "RoundSpaceIndicator": "\uF3B2",
-    "Rule": "\uF522",
-    "RuleDelayed": "\uF51F",
-    "SHacek": "\u0161",
-    "SZ": "\u00DF",
-    "SadSmiley": "\u2639",
-    "SagittariusSign": "\u2650",
-    "Sampi": "\u03E0",
-    "Saturn": "\u2644",
-    "ScorpioSign": "\u264F",
-    "ScriptA": "\uF6B2",
-    "ScriptB": "\uF6B3",
-    "ScriptC": "\uF6B4",
-    "ScriptCapitalA": "\uF770",
-    "ScriptCapitalB": "\u212C",
-    "ScriptCapitalC": "\uF772",
-    "ScriptCapitalD": "\uF773",
-    "ScriptCapitalE": "\u2130",
-    "ScriptCapitalF": "\u2131",
-    "ScriptCapitalG": "\uF776",
-    "ScriptCapitalH": "\u210B",
-    "ScriptCapitalI": "\u2110",
-    "ScriptCapitalJ": "\uF779",
-    "ScriptCapitalK": "\uF77A",
-    "ScriptCapitalL": "\u2112",
-    "ScriptCapitalM": "\u2133",
-    "ScriptCapitalN": "\uF77D",
-    "ScriptCapitalO": "\uF77E",
-    "ScriptCapitalP": "\u2118",
-    "ScriptCapitalQ": "\uF780",
-    "ScriptCapitalR": "\u211B",
-    "ScriptCapitalS": "\uF782",
-    "ScriptCapitalT": "\uF783",
-    "ScriptCapitalU": "\uF784",
-    "ScriptCapitalV": "\uF785",
-    "ScriptCapitalW": "\uF786",
-    "ScriptCapitalX": "\uF787",
-    "ScriptCapitalY": "\uF788",
-    "ScriptCapitalZ": "\uF789",
-    "ScriptD": "\uF6B5",
-    "ScriptDotlessI": "\uF730",
-    "ScriptDotlessJ": "\uF731",
-    "ScriptE": "\u212F",
-    "ScriptEight": "\uF7F8",
-    "ScriptF": "\uF6B7",
-    "ScriptFive": "\uF7F5",
-    "ScriptFour": "\uF7F4",
-    "ScriptG": "\u210A",
-    "ScriptH": "\uF6B9",
-    "ScriptI": "\uF6BA",
-    "ScriptJ": "\uF6BB",
-    "ScriptK": "\uF6BC",
-    "ScriptL": "\u2113",
-    "ScriptM": "\uF6BE",
-    "ScriptN": "\uF6BF",
-    "ScriptNine": "\uF7F9",
-    "ScriptO": "\u2134",
-    "ScriptOne": "\uF7F1",
-    "ScriptP": "\uF6C1",
-    "ScriptQ": "\uF6C2",
-    "ScriptR": "\uF6C3",
-    "ScriptS": "\uF6C4",
-    "ScriptSeven": "\uF7F7",
-    "ScriptSix": "\uF7F6",
-    "ScriptT": "\uF6C5",
-    "ScriptThree": "\uF7F3",
-    "ScriptTwo": "\uF7F2",
-    "ScriptU": "\uF6C6",
-    "ScriptV": "\uF6C7",
-    "ScriptW": "\uF6C8",
-    "ScriptX": "\uF6C9",
-    "ScriptY": "\uF6CA",
-    "ScriptZ": "\uF6CB",
-    "ScriptZero": "\uF7F0",
-    "Section": "\u00A7",
-    "SelectionPlaceholder": "\uF527",
-    "Shah": "\uF11D",
-    "Sharp": "\u266F",
-    "ShiftKey": "\uF7D5",
-    "ShortDownArrow": "\uF52B",
-    "ShortLeftArrow": "\uF526",
-    "ShortRightArrow": "\uF525",
-    "Sigma": "\u03C3",
-    "SixPointedStar": "\u2736",
-    "SkeletonIndicator": "\u2043",
-    "SmallCircle": "\u2218",
-    "SpaceIndicator": "\u2423",
-    "SpaceKey": "\uF7BF",
-    "SpadeSuit": "\u2660",
-    "SpanFromAbove": "\uF3BB",
-    "SpanFromBoth": "\uF3BC",
-    "SpanFromLeft": "\uF3BA",
-    "SphericalAngle": "\u2222",
-    "Sqrt": "\u221A",
-    "Square": "\uF520",
-    "SquareIntersection": "\u2293",
-    "SquareSubset": "\u228F",
-    "SquareSubsetEqual": "\u2291",
-    "SquareSuperset": "\u2290",
-    "SquareSupersetEqual": "\u2292",
-    "SquareUnion": "\u2294",
-    "Star": "\u22C6",
-    "StepperDown": "\uF3CD",
-    "StepperLeft": "\uF3CB",
-    "StepperRight": "\uF3CA",
-    "StepperUp": "\uF3CC",
-    "Sterling": "\u00A3",
-    "Stigma": "\u03DB",
-    "Subset": "\u2282",
-    "SubsetEqual": "\u2286",
-    "Succeeds": "\u227B",
-    "SucceedsEqual": "\u2AB0",
-    "SucceedsSlantEqual": "\u227D",
-    "SucceedsTilde": "\u227F",
-    "SuchThat": "\u220D",
-    "Sum": "\u2211",
-    "Sun": "\u2609",
-    "Superset": "\u2283",
-    "SupersetEqual": "\u2287",
-    "SystemEnterKey": "\uF75F",
-    "SystemsModelDelay": "\uF3AF",
-    "THacek": "\u0165",
-    "TabKey": "\uF7BE",
-    "Tau": "\u03C4",
-    "TaurusSign": "\u2649",
-    "TensorProduct": "\uF3DA",
-    "TensorWedge": "\uF3DB",
-    "Therefore": "\u2234",
-    "Theta": "\u03B8",
-    "ThickSpace": "\u2005",
-    "ThinSpace": "\u2009",
-    "Thorn": "\u00FE",
-    "Tilde": "\u223C",
-    "TildeEqual": "\u2243",
-    "TildeFullEqual": "\u2245",
-    "TildeTilde": "\u2248",
-    "Times": "\u00D7",
-    "Trademark": "\u2122",
-    "Transpose": "\uF3C7",
-    "TripleDot": "\uF758",
-    "TwoWayRule": "\uF120",
-    "UAcute": "\u00FA",
-    "UDoubleAcute": "\u0171",
-    "UDoubleDot": "\u00FC",
-    "UGrave": "\u00F9",
-    "UHat": "\u00FB",
-    "URing": "\u016F",
-    "UnderBrace": "\uFE38",
-    "UnderBracket": "\u23B5",
-    "UnderParenthesis": "\uFE36",
-    "UndirectedEdge": "\uF3D4",
-    "Union": "\u22C3",
-    "UnionPlus": "\u228E",
-    "UnknownGlyph": "\uFFFD",
-    "UpArrow": "\u2191",
-    "UpArrowBar": "\u2912",
-    "UpArrowDownArrow": "\u21C5",
-    "UpDownArrow": "\u2195",
-    "UpEquilibrium": "\u296E",
-    "UpPointer": "\u25B4",
-    "UpTee": "\u22A5",
-    "UpTeeArrow": "\u21A5",
-    "UpperLeftArrow": "\u2196",
-    "UpperRightArrow": "\u2197",
-    "Upsilon": "\u03C5",
-    "Uranus": "\u2645",
-    "VectorGreaterEqual": "\uF435",
-    "VectorGreater": "\uF434",
-    "VectorLessEqual": "\uF437",
-    "VectorLess": "\uF436",
-    "Vee": "\u22C1",
-    "Venus": "\u2640",
-    "VerticalBar": "\u2223",
-    "VerticalEllipsis": "\u22EE",
-    "VerticalLine": "\u2502",
-    "VerticalSeparator": "\uF432",
-    "VerticalTilde": "\u2240",
-    "VeryThinSpace": "\u200A",
-    "Villa": "\uF727",
-    "VirgoSign": "\u264D",
-    "WarningSign": "\uF725",
-    "WatchIcon": "\u231A",
-    "Wedge": "\u22C0",
-    "WeierstrassP": "\u2118",
-    "WhiteBishop": "\u2657",
-    "WhiteKing": "\u2654",
-    "WhiteKnight": "\u2658",
-    "WhitePawn": "\u2659",
-    "WhiteQueen": "\u2655",
-    "WhiteRook": "\u2656",
-    "Wolf": "\uF720",
-    "WolframAlphaPrompt": "\uF352",
-    "WolframLanguageLogoCircle": "\uF11F",
-    "WolframLanguageLogo": "\uF11E",
-    "Xi": "\u03BE",
-    "Xnor": "\uF4A2",
-    "Xor": "\u22BB",
-    "YAcute": "\u00FD",
-    "YDoubleDot": "\u00FF",
-    "Yen": "\u00A5",
-    "ZHacek": "\u017E",
-    "Zeta": "\u03B6",
+# Conversion from WL to unicode
+# IMPORTANT: The entries in this dictionary should be kept in here even when 
+# the key maps to itself, or otherwise they will be overwritten by the entries 
+# of WL_TO_NAMED in WL_TO_UNICODE_DICT
+WL_TO_UNICODE = {
+    'á': 'á', # \[AAcute] (LATIN SMALL LETTER A WITH ACUTE) -> LATIN SMALL LETTER A WITH ACUTE
+    'ā': 'ā', # \[ABar] (LATIN SMALL LETTER A WITH MACRON) -> LATIN SMALL LETTER A WITH MACRON
+    'ă': 'ă', # \[ACup] (LATIN SMALL LETTER A WITH BREVE) -> LATIN SMALL LETTER A WITH BREVE
+    'ä': 'ä', # \[ADoubleDot] (LATIN SMALL LETTER A WITH DIAERESIS) -> LATIN SMALL LETTER A WITH DIAERESIS
+    'æ': 'æ', # \[AE] (LATIN SMALL LETTER AE) -> LATIN SMALL LETTER AE
+    'à': 'à', # \[AGrave] (LATIN SMALL LETTER A WITH GRAVE) -> LATIN SMALL LETTER A WITH GRAVE
+    'â': 'â', # \[AHat] (LATIN SMALL LETTER A WITH CIRCUMFLEX) -> LATIN SMALL LETTER A WITH CIRCUMFLEX
+    'ℵ': 'ℵ', # \[Aleph] (ALEF SYMBOL) -> ALEF SYMBOL
+    'α': 'α', # \[Alpha] (GREEK SMALL LETTER ALPHA) -> GREEK SMALL LETTER ALPHA
+    '∧': '∧', # \[And] (LOGICAL AND) -> LOGICAL AND
+    '∠': '∠', # \[Angle] (ANGLE) -> ANGLE
+    'Å': 'Å', # \[Angstrom] (ANGSTROM SIGN) -> ANGSTROM SIGN
+    '♒': '♒', # \[AquariusSign] (AQUARIUS) -> AQUARIUS
+    '♈': '♈', # \[AriesSign] (ARIES) -> ARIES
+    'å': 'å', # \[ARing] (LATIN SMALL LETTER A WITH RING ABOVE) -> LATIN SMALL LETTER A WITH RING ABOVE
+    '⋰': '⋰', # \[AscendingEllipsis] (UP RIGHT DIAGONAL ELLIPSIS) -> UP RIGHT DIAGONAL ELLIPSIS
+    'ã': 'ã', # \[ATilde] (LATIN SMALL LETTER A WITH TILDE) -> LATIN SMALL LETTER A WITH TILDE
+    '∖': '∖', # \[Backslash] (SET MINUS) -> SET MINUS
+    '♫': '♫', # \[BeamedEighthNote] (BEAMED EIGHTH NOTES) -> BEAMED EIGHTH NOTES
+    '♬': '♬', # \[BeamedSixteenthNote] (BEAMED SIXTEENTH NOTES) -> BEAMED SIXTEENTH NOTES
+    '∵': '∵', # \[Because] (BECAUSE) -> BECAUSE
+    'ℶ': 'ℶ', # \[Bet] (BET SYMBOL) -> BET SYMBOL
+    'β': 'β', # \[Beta] (GREEK SMALL LETTER BETA) -> GREEK SMALL LETTER BETA
+    '♝': '♝', # \[BlackBishop] (BLACK CHESS BISHOP) -> BLACK CHESS BISHOP
+    '♚': '♚', # \[BlackKing] (BLACK CHESS KING) -> BLACK CHESS KING
+    '♞': '♞', # \[BlackKnight] (BLACK CHESS KNIGHT) -> BLACK CHESS KNIGHT
+    '♟': '♟', # \[BlackPawn] (BLACK CHESS PAWN) -> BLACK CHESS PAWN
+    '♛': '♛', # \[BlackQueen] (BLACK CHESS QUEEN) -> BLACK CHESS QUEEN
+    '♜': '♜', # \[BlackRook] (BLACK CHESS ROOK) -> BLACK CHESS ROOK
+    '˘': '˘', # \[Breve] (BREVE) -> BREVE
+    '•': '•', # \[Bullet] (BULLET) -> BULLET
+    'ć': 'ć', # \[CAcute] (LATIN SMALL LETTER C WITH ACUTE) -> LATIN SMALL LETTER C WITH ACUTE
+    '♋': '♋', # \[CancerSign] (CANCER) -> CANCER
+    '⌢': '⌢', # \[Cap] (FROWN) -> FROWN
+    'Á': 'Á', # \[CapitalAAcute] (LATIN CAPITAL LETTER A WITH ACUTE) -> LATIN CAPITAL LETTER A WITH ACUTE
+    'Ā': 'Ā', # \[CapitalABar] (LATIN CAPITAL LETTER A WITH MACRON) -> LATIN CAPITAL LETTER A WITH MACRON
+    'Ă': 'Ă', # \[CapitalACup] (LATIN CAPITAL LETTER A WITH BREVE) -> LATIN CAPITAL LETTER A WITH BREVE
+    'Ä': 'Ä', # \[CapitalADoubleDot] (LATIN CAPITAL LETTER A WITH DIAERESIS) -> LATIN CAPITAL LETTER A WITH DIAERESIS
+    'Æ': 'Æ', # \[CapitalAE] (LATIN CAPITAL LETTER AE) -> LATIN CAPITAL LETTER AE
+    'À': 'À', # \[CapitalAGrave] (LATIN CAPITAL LETTER A WITH GRAVE) -> LATIN CAPITAL LETTER A WITH GRAVE
+    'Â': 'Â', # \[CapitalAHat] (LATIN CAPITAL LETTER A WITH CIRCUMFLEX) -> LATIN CAPITAL LETTER A WITH CIRCUMFLEX
+    'Α': 'Α', # \[CapitalAlpha] (GREEK CAPITAL LETTER ALPHA) -> GREEK CAPITAL LETTER ALPHA
+    'Å': 'Å', # \[CapitalARing] (LATIN CAPITAL LETTER A WITH RING ABOVE) -> LATIN CAPITAL LETTER A WITH RING ABOVE
+    'Ã': 'Ã', # \[CapitalATilde] (LATIN CAPITAL LETTER A WITH TILDE) -> LATIN CAPITAL LETTER A WITH TILDE
+    'Β': 'Β', # \[CapitalBeta] (GREEK CAPITAL LETTER BETA) -> GREEK CAPITAL LETTER BETA
+    'Ć': 'Ć', # \[CapitalCAcute] (LATIN CAPITAL LETTER C WITH ACUTE) -> LATIN CAPITAL LETTER C WITH ACUTE
+    'Ç': 'Ç', # \[CapitalCCedilla] (LATIN CAPITAL LETTER C WITH CEDILLA) -> LATIN CAPITAL LETTER C WITH CEDILLA
+    'Č': 'Č', # \[CapitalCHacek] (LATIN CAPITAL LETTER C WITH CARON) -> LATIN CAPITAL LETTER C WITH CARON
+    'Χ': 'Χ', # \[CapitalChi] (GREEK CAPITAL LETTER CHI) -> GREEK CAPITAL LETTER CHI
+    'Δ': 'Δ', # \[CapitalDelta] (GREEK CAPITAL LETTER DELTA) -> GREEK CAPITAL LETTER DELTA
+    'Ď': 'Ď', # \[CapitalDHacek] (LATIN CAPITAL LETTER D WITH CARON) -> LATIN CAPITAL LETTER D WITH CARON
+    '\uf74b': '𝐷', # \[CapitalDifferentialD] (INVALID CHARACTER) -> MATHEMATICAL ITALIC CAPITAL D
+    'Ϝ': 'Ϝ', # \[CapitalDigamma] (GREEK LETTER DIGAMMA) -> GREEK LETTER DIGAMMA
+    'É': 'É', # \[CapitalEAcute] (LATIN CAPITAL LETTER E WITH ACUTE) -> LATIN CAPITAL LETTER E WITH ACUTE
+    'Ē': 'Ē', # \[CapitalEBar] (LATIN CAPITAL LETTER E WITH MACRON) -> LATIN CAPITAL LETTER E WITH MACRON
+    'Ĕ': 'Ĕ', # \[CapitalECup] (LATIN CAPITAL LETTER E WITH BREVE) -> LATIN CAPITAL LETTER E WITH BREVE
+    'Ë': 'Ë', # \[CapitalEDoubleDot] (LATIN CAPITAL LETTER E WITH DIAERESIS) -> LATIN CAPITAL LETTER E WITH DIAERESIS
+    'È': 'È', # \[CapitalEGrave] (LATIN CAPITAL LETTER E WITH GRAVE) -> LATIN CAPITAL LETTER E WITH GRAVE
+    'Ě': 'Ě', # \[CapitalEHacek] (LATIN CAPITAL LETTER E WITH CARON) -> LATIN CAPITAL LETTER E WITH CARON
+    'Ê': 'Ê', # \[CapitalEHat] (LATIN CAPITAL LETTER E WITH CIRCUMFLEX) -> LATIN CAPITAL LETTER E WITH CIRCUMFLEX
+    'Ε': 'Ε', # \[CapitalEpsilon] (GREEK CAPITAL LETTER EPSILON) -> GREEK CAPITAL LETTER EPSILON
+    'Η': 'Η', # \[CapitalEta] (GREEK CAPITAL LETTER ETA) -> GREEK CAPITAL LETTER ETA
+    'Ð': 'Ð', # \[CapitalEth] (LATIN CAPITAL LETTER ETH) -> LATIN CAPITAL LETTER ETH
+    'Γ': 'Γ', # \[CapitalGamma] (GREEK CAPITAL LETTER GAMMA) -> GREEK CAPITAL LETTER GAMMA
+    'Í': 'Í', # \[CapitalIAcute] (LATIN CAPITAL LETTER I WITH ACUTE) -> LATIN CAPITAL LETTER I WITH ACUTE
+    'Ĭ': 'Ĭ', # \[CapitalICup] (LATIN CAPITAL LETTER I WITH BREVE) -> LATIN CAPITAL LETTER I WITH BREVE
+    'Ï': 'Ï', # \[CapitalIDoubleDot] (LATIN CAPITAL LETTER I WITH DIAERESIS) -> LATIN CAPITAL LETTER I WITH DIAERESIS
+    'Ì': 'Ì', # \[CapitalIGrave] (LATIN CAPITAL LETTER I WITH GRAVE) -> LATIN CAPITAL LETTER I WITH GRAVE
+    'Î': 'Î', # \[CapitalIHat] (LATIN CAPITAL LETTER I WITH CIRCUMFLEX) -> LATIN CAPITAL LETTER I WITH CIRCUMFLEX
+    'Ι': 'Ι', # \[CapitalIota] (GREEK CAPITAL LETTER IOTA) -> GREEK CAPITAL LETTER IOTA
+    'Κ': 'Κ', # \[CapitalKappa] (GREEK CAPITAL LETTER KAPPA) -> GREEK CAPITAL LETTER KAPPA
+    'Ϟ': 'Ϟ', # \[CapitalKoppa] (GREEK LETTER KOPPA) -> GREEK LETTER KOPPA
+    'Λ': 'Λ', # \[CapitalLambda] (GREEK CAPITAL LETTER LAMDA) -> GREEK CAPITAL LETTER LAMDA
+    'Ł': 'Ł', # \[CapitalLSlash] (LATIN CAPITAL LETTER L WITH STROKE) -> LATIN CAPITAL LETTER L WITH STROKE
+    'Μ': 'Μ', # \[CapitalMu] (GREEK CAPITAL LETTER MU) -> GREEK CAPITAL LETTER MU
+    'Ň': 'Ň', # \[CapitalNHacek] (LATIN CAPITAL LETTER N WITH CARON) -> LATIN CAPITAL LETTER N WITH CARON
+    'Ñ': 'Ñ', # \[CapitalNTilde] (LATIN CAPITAL LETTER N WITH TILDE) -> LATIN CAPITAL LETTER N WITH TILDE
+    'Ν': 'Ν', # \[CapitalNu] (GREEK CAPITAL LETTER NU) -> GREEK CAPITAL LETTER NU
+    'Ó': 'Ó', # \[CapitalOAcute] (LATIN CAPITAL LETTER O WITH ACUTE) -> LATIN CAPITAL LETTER O WITH ACUTE
+    'Ő': 'Ő', # \[CapitalODoubleAcute] (LATIN CAPITAL LETTER O WITH DOUBLE ACUTE) -> LATIN CAPITAL LETTER O WITH DOUBLE ACUTE
+    'Ö': 'Ö', # \[CapitalODoubleDot] (LATIN CAPITAL LETTER O WITH DIAERESIS) -> LATIN CAPITAL LETTER O WITH DIAERESIS
+    'Œ': 'Œ', # \[CapitalOE] (LATIN CAPITAL LIGATURE OE) -> LATIN CAPITAL LIGATURE OE
+    'Ò': 'Ò', # \[CapitalOGrave] (LATIN CAPITAL LETTER O WITH GRAVE) -> LATIN CAPITAL LETTER O WITH GRAVE
+    'Ô': 'Ô', # \[CapitalOHat] (LATIN CAPITAL LETTER O WITH CIRCUMFLEX) -> LATIN CAPITAL LETTER O WITH CIRCUMFLEX
+    'Ω': 'Ω', # \[CapitalOmega] (GREEK CAPITAL LETTER OMEGA) -> GREEK CAPITAL LETTER OMEGA
+    'Ο': 'Ο', # \[CapitalOmicron] (GREEK CAPITAL LETTER OMICRON) -> GREEK CAPITAL LETTER OMICRON
+    'Ø': 'Ø', # \[CapitalOSlash] (LATIN CAPITAL LETTER O WITH STROKE) -> LATIN CAPITAL LETTER O WITH STROKE
+    'Õ': 'Õ', # \[CapitalOTilde] (LATIN CAPITAL LETTER O WITH TILDE) -> LATIN CAPITAL LETTER O WITH TILDE
+    'Φ': 'Φ', # \[CapitalPhi] (GREEK CAPITAL LETTER PHI) -> GREEK CAPITAL LETTER PHI
+    'Π': 'Π', # \[CapitalPi] (GREEK CAPITAL LETTER PI) -> GREEK CAPITAL LETTER PI
+    'Ψ': 'Ψ', # \[CapitalPsi] (GREEK CAPITAL LETTER PSI) -> GREEK CAPITAL LETTER PSI
+    'Ř': 'Ř', # \[CapitalRHacek] (LATIN CAPITAL LETTER R WITH CARON) -> LATIN CAPITAL LETTER R WITH CARON
+    'Ρ': 'Ρ', # \[CapitalRho] (GREEK CAPITAL LETTER RHO) -> GREEK CAPITAL LETTER RHO
+    'Ϡ': 'Ϡ', # \[CapitalSampi] (GREEK LETTER SAMPI) -> GREEK LETTER SAMPI
+    'Š': 'Š', # \[CapitalSHacek] (LATIN CAPITAL LETTER S WITH CARON) -> LATIN CAPITAL LETTER S WITH CARON
+    'Σ': 'Σ', # \[CapitalSigma] (GREEK CAPITAL LETTER SIGMA) -> GREEK CAPITAL LETTER SIGMA
+    'Ϛ': 'Ϛ', # \[CapitalStigma] (GREEK LETTER STIGMA) -> GREEK LETTER STIGMA
+    'Τ': 'Τ', # \[CapitalTau] (GREEK CAPITAL LETTER TAU) -> GREEK CAPITAL LETTER TAU
+    'Ť': 'Ť', # \[CapitalTHacek] (LATIN CAPITAL LETTER T WITH CARON) -> LATIN CAPITAL LETTER T WITH CARON
+    'Θ': 'Θ', # \[CapitalTheta] (GREEK CAPITAL LETTER THETA) -> GREEK CAPITAL LETTER THETA
+    'Þ': 'Þ', # \[CapitalThorn] (LATIN CAPITAL LETTER THORN) -> LATIN CAPITAL LETTER THORN
+    'Ú': 'Ú', # \[CapitalUAcute] (LATIN CAPITAL LETTER U WITH ACUTE) -> LATIN CAPITAL LETTER U WITH ACUTE
+    'Ű': 'Ű', # \[CapitalUDoubleAcute] (LATIN CAPITAL LETTER U WITH DOUBLE ACUTE) -> LATIN CAPITAL LETTER U WITH DOUBLE ACUTE
+    'Ü': 'Ü', # \[CapitalUDoubleDot] (LATIN CAPITAL LETTER U WITH DIAERESIS) -> LATIN CAPITAL LETTER U WITH DIAERESIS
+    'Ù': 'Ù', # \[CapitalUGrave] (LATIN CAPITAL LETTER U WITH GRAVE) -> LATIN CAPITAL LETTER U WITH GRAVE
+    'Û': 'Û', # \[CapitalUHat] (LATIN CAPITAL LETTER U WITH CIRCUMFLEX) -> LATIN CAPITAL LETTER U WITH CIRCUMFLEX
+    'Υ': 'Υ', # \[CapitalUpsilon] (GREEK CAPITAL LETTER UPSILON) -> GREEK CAPITAL LETTER UPSILON
+    'Ů': 'Ů', # \[CapitalURing] (LATIN CAPITAL LETTER U WITH RING ABOVE) -> LATIN CAPITAL LETTER U WITH RING ABOVE
+    'Ξ': 'Ξ', # \[CapitalXi] (GREEK CAPITAL LETTER XI) -> GREEK CAPITAL LETTER XI
+    'Ý': 'Ý', # \[CapitalYAcute] (LATIN CAPITAL LETTER Y WITH ACUTE) -> LATIN CAPITAL LETTER Y WITH ACUTE
+    'Ζ': 'Ζ', # \[CapitalZeta] (GREEK CAPITAL LETTER ZETA) -> GREEK CAPITAL LETTER ZETA
+    'Ž': 'Ž', # \[CapitalZHacek] (LATIN CAPITAL LETTER Z WITH CARON) -> LATIN CAPITAL LETTER Z WITH CARON
+    '♑': '♑', # \[CapricornSign] (CAPRICORN) -> CAPRICORN
+    'ç': 'ç', # \[CCedilla] (LATIN SMALL LETTER C WITH CEDILLA) -> LATIN SMALL LETTER C WITH CEDILLA
+    '¸': '¸', # \[Cedilla] (CEDILLA) -> CEDILLA
+    '·': '·', # \[CenterDot] (MIDDLE DOT) -> MIDDLE DOT
+    '⋯': '⋯', # \[CenterEllipsis] (MIDLINE HORIZONTAL ELLIPSIS) -> MIDLINE HORIZONTAL ELLIPSIS
+    '¢': '¢', # \[Cent] (CENT SIGN) -> CENT SIGN
+    'č': 'č', # \[CHacek] (LATIN SMALL LETTER C WITH CARON) -> LATIN SMALL LETTER C WITH CARON
+    '☒': '☒', # \[CheckedBox] (BALLOT BOX WITH X) -> BALLOT BOX WITH X
+    '✓': '✓', # \[Checkmark] (CHECK MARK) -> CHECK MARK
+    'χ': 'χ', # \[Chi] (GREEK SMALL LETTER CHI) -> GREEK SMALL LETTER CHI
+    '⊙': '⊙', # \[CircleDot] (CIRCLED DOT OPERATOR) -> CIRCLED DOT OPERATOR
+    '⊖': '⊖', # \[CircleMinus] (CIRCLED MINUS) -> CIRCLED MINUS
+    '⊕': '⊕', # \[CirclePlus] (CIRCLED PLUS) -> CIRCLED PLUS
+    '⊗': '⊗', # \[CircleTimes] (CIRCLED TIMES) -> CIRCLED TIMES
+    '∲': '∲', # \[ClockwiseContourIntegral] (CLOCKWISE CONTOUR INTEGRAL) -> CLOCKWISE CONTOUR INTEGRAL
+    '”': '”', # \[CloseCurlyDoubleQuote] (RIGHT DOUBLE QUOTATION MARK) -> RIGHT DOUBLE QUOTATION MARK
+    '’': '’', # \[CloseCurlyQuote] (RIGHT SINGLE QUOTATION MARK) -> RIGHT SINGLE QUOTATION MARK
+    '⌘': '⌘', # \[CloverLeaf] (PLACE OF INTEREST SIGN) -> PLACE OF INTEREST SIGN
+    '♣': '♣', # \[ClubSuit] (BLACK CLUB SUIT) -> BLACK CLUB SUIT
+    '∶': '∶', # \[Colon] (RATIO) -> RATIO
+    '\uf76a': '⌘', # \[CommandKey] (INVALID CHARACTER) -> PLACE OF INTEREST SIGN
+    '≡': '≡', # \[Congruent] (IDENTICAL TO) -> IDENTICAL TO
+    '\uf7da': '𝕔', # \[ConstantC] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL C
+    '\uf3b1': '⋱', # \[Continuation] (INVALID CHARACTER) -> DOWN RIGHT DIAGONAL ELLIPSIS
+    '∮': '∮', # \[ContourIntegral] (CONTOUR INTEGRAL) -> CONTOUR INTEGRAL
+    '∐': '∐', # \[Coproduct] (N-ARY COPRODUCT) -> N-ARY COPRODUCT
+    '©': '©', # \[Copyright] (COPYRIGHT SIGN) -> COPYRIGHT SIGN
+    '∳': '∳', # \[CounterClockwiseContourIntegral] (ANTICLOCKWISE CONTOUR INTEGRAL) -> ANTICLOCKWISE CONTOUR INTEGRAL
+    '\uf4a0': '⨯', # \[Cross] (INVALID CHARACTER) -> VECTOR OR CROSS PRODUCT
+    '≍': '≍', # \[CupCap] (EQUIVALENT TO) -> EQUIVALENT TO
+    '⌣': '⌣', # \[Cup] (SMILE) -> SMILE
+    'ϒ': 'ϒ', # \[CurlyCapitalUpsilon] (GREEK UPSILON WITH HOOK SYMBOL) -> GREEK UPSILON WITH HOOK SYMBOL
+    'ε': 'ε', # \[CurlyEpsilon] (GREEK SMALL LETTER EPSILON) -> GREEK SMALL LETTER EPSILON
+    'ϰ': 'ϰ', # \[CurlyKappa] (GREEK KAPPA SYMBOL) -> GREEK KAPPA SYMBOL
+    'φ': 'φ', # \[CurlyPhi] (GREEK SMALL LETTER PHI) -> GREEK SMALL LETTER PHI
+    'ϖ': 'ϖ', # \[CurlyPi] (GREEK PI SYMBOL) -> GREEK PI SYMBOL
+    'ϱ': 'ϱ', # \[CurlyRho] (GREEK RHO SYMBOL) -> GREEK RHO SYMBOL
+    'ϑ': 'ϑ', # \[CurlyTheta] (GREEK THETA SYMBOL) -> GREEK THETA SYMBOL
+    '¤': '¤', # \[Currency] (CURRENCY SIGN) -> CURRENCY SIGN
+    '†': '†', # \[Dagger] (DAGGER) -> DAGGER
+    'ℸ': 'ℸ', # \[Dalet] (DALET SYMBOL) -> DALET SYMBOL
+    '–': '–', # \[Dash] (EN DASH) -> EN DASH
+    '°': '°', # \[Degree] (DEGREE SIGN) -> DEGREE SIGN
+    '∇': '∇', # \[Del] (NABLA) -> NABLA
+    'δ': 'δ', # \[Delta] (GREEK SMALL LETTER DELTA) -> GREEK SMALL LETTER DELTA
+    '⋱': '⋱', # \[DescendingEllipsis] (DOWN RIGHT DIAGONAL ELLIPSIS) -> DOWN RIGHT DIAGONAL ELLIPSIS
+    'ď': 'ď', # \[DHacek] (LATIN SMALL LETTER D WITH CARON) -> LATIN SMALL LETTER D WITH CARON
+    '⌀': '⌀', # \[Diameter] (DIAMETER SIGN) -> DIAMETER SIGN
+    '⋄': '⋄', # \[Diamond] (DIAMOND OPERATOR) -> DIAMOND OPERATOR
+    '♢': '♢', # \[DiamondSuit] (WHITE DIAMOND SUIT) -> WHITE DIAMOND SUIT
+    '∆': '∆', # \[DifferenceDelta] (INCREMENT) -> INCREMENT
+    '\uf74c': '𝑑', # \[DifferentialD] (INVALID CHARACTER) -> MATHEMATICAL ITALIC SMALL D
+    'ϝ': 'ϝ', # \[Digamma] (GREEK SMALL LETTER DIGAMMA) -> GREEK SMALL LETTER DIGAMMA
+    '\uf3d5': '→', # \[DirectedEdge] (INVALID CHARACTER) -> RIGHTWARDS ARROW
+    '\uf4a4': 'ϴ', # \[DiscreteRatio] (INVALID CHARACTER) -> GREEK CAPITAL THETA SYMBOL
+    '∣': '∣', # \[Divides] (DIVIDES) -> DIVIDES
+    '÷': '÷', # \[Divide] (DIVISION SIGN) -> DIVISION SIGN
+    '≐': '≐', # \[DotEqual] (APPROACHES THE LIMIT) -> APPROACHES THE LIMIT
+    'ı': 'ı', # \[DotlessI] (LATIN SMALL LETTER DOTLESS I) -> LATIN SMALL LETTER DOTLESS I
+    '\uf700': 'ȷ', # \[DotlessJ] (INVALID CHARACTER) -> LATIN SMALL LETTER DOTLESS J
+    '\uf751': '⛶', # \[DottedSquare] (INVALID CHARACTER) -> SQUARE FOUR CORNERS
+    '∯': '∯', # \[DoubleContourIntegral] (SURFACE INTEGRAL) -> SURFACE INTEGRAL
+    '‡': '‡', # \[DoubleDagger] (DOUBLE DAGGER) -> DOUBLE DAGGER
+    '\uf74a': 'ℽ', # \[DoubledGamma] (INVALID CHARACTER) -> DOUBLE-STRUCK SMALL GAMMA
+    '¨': '¨', # \[DoubleDot] (DIAERESIS) -> DIAERESIS
+    '⇓': '⇓', # \[DoubleDownArrow] (DOWNWARDS DOUBLE ARROW) -> DOWNWARDS DOUBLE ARROW
+    '\uf749': 'ℼ', # \[DoubledPi] (INVALID CHARACTER) -> DOUBLE-STRUCK SMALL PI
+    '⇐': '⇐', # \[DoubleLeftArrow] (LEFTWARDS DOUBLE ARROW) -> LEFTWARDS DOUBLE ARROW
+    '⇔': '⇔', # \[DoubleLeftRightArrow] (LEFT RIGHT DOUBLE ARROW) -> LEFT RIGHT DOUBLE ARROW
+    '⫤': '⫤', # \[DoubleLeftTee] (VERTICAL BAR DOUBLE LEFT TURNSTILE) -> VERTICAL BAR DOUBLE LEFT TURNSTILE
+    '⟸': '⟸', # \[DoubleLongLeftArrow] (LONG LEFTWARDS DOUBLE ARROW) -> LONG LEFTWARDS DOUBLE ARROW
+    '⟺': '⟺', # \[DoubleLongLeftRightArrow] (LONG LEFT RIGHT DOUBLE ARROW) -> LONG LEFT RIGHT DOUBLE ARROW
+    '⟹': '⟹', # \[DoubleLongRightArrow] (LONG RIGHTWARDS DOUBLE ARROW) -> LONG RIGHTWARDS DOUBLE ARROW
+    '″': '″', # \[DoublePrime] (DOUBLE PRIME) -> DOUBLE PRIME
+    '⇒': '⇒', # \[DoubleRightArrow] (RIGHTWARDS DOUBLE ARROW) -> RIGHTWARDS DOUBLE ARROW
+    '⊨': '⊨', # \[DoubleRightTee] (TRUE) -> TRUE
+    '\uf6e6': '𝕒', # \[DoubleStruckA] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL A
+    '\uf6e7': '𝕓', # \[DoubleStruckB] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL B
+    '\uf6e8': '𝕔', # \[DoubleStruckC] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL C
+    '\uf7a4': '𝔸', # \[DoubleStruckCapitalA] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL A
+    '\uf7a5': '𝔹', # \[DoubleStruckCapitalB] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL B
+    '\uf7a6': 'ℂ', # \[DoubleStruckCapitalC] (INVALID CHARACTER) -> DOUBLE-STRUCK CAPITAL C
+    '\uf7a7': '𝔻', # \[DoubleStruckCapitalD] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL D
+    '\uf7a8': '𝔼', # \[DoubleStruckCapitalE] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL E
+    '\uf7a9': '𝔽', # \[DoubleStruckCapitalF] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL F
+    '\uf7aa': '𝔾', # \[DoubleStruckCapitalG] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL G
+    '\uf7ab': 'ℍ', # \[DoubleStruckCapitalH] (INVALID CHARACTER) -> DOUBLE-STRUCK CAPITAL H
+    '\uf7ac': '𝕀', # \[DoubleStruckCapitalI] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL I
+    '\uf7ad': '𝕁', # \[DoubleStruckCapitalJ] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL J
+    '\uf7ae': '𝕂', # \[DoubleStruckCapitalK] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL K
+    '\uf7af': '𝕃', # \[DoubleStruckCapitalL] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL L
+    '\uf7b0': '𝕄', # \[DoubleStruckCapitalM] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL M
+    '\uf7b1': 'ℕ', # \[DoubleStruckCapitalN] (INVALID CHARACTER) -> DOUBLE-STRUCK CAPITAL N
+    '\uf7b2': '𝕆', # \[DoubleStruckCapitalO] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL O
+    '\uf7b3': 'ℙ', # \[DoubleStruckCapitalP] (INVALID CHARACTER) -> DOUBLE-STRUCK CAPITAL P
+    '\uf7b4': 'ℚ', # \[DoubleStruckCapitalQ] (INVALID CHARACTER) -> DOUBLE-STRUCK CAPITAL Q
+    '\uf7b5': 'ℝ', # \[DoubleStruckCapitalR] (INVALID CHARACTER) -> DOUBLE-STRUCK CAPITAL R
+    '\uf7b6': '𝕊', # \[DoubleStruckCapitalS] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL S
+    '\uf7b7': '𝕋', # \[DoubleStruckCapitalT] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL T
+    '\uf7b8': '𝕌', # \[DoubleStruckCapitalU] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL U
+    '\uf7b9': '𝕍', # \[DoubleStruckCapitalV] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL V
+    '\uf7ba': '𝕎', # \[DoubleStruckCapitalW] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL W
+    '\uf7bb': '𝕏', # \[DoubleStruckCapitalX] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL X
+    '\uf7bc': '𝕐', # \[DoubleStruckCapitalY] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK CAPITAL Y
+    '\uf7bd': 'ℤ', # \[DoubleStruckCapitalZ] (INVALID CHARACTER) -> DOUBLE-STRUCK CAPITAL Z
+    '\uf6e9': '𝕕', # \[DoubleStruckD] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL D
+    '\uf6ea': '𝕖', # \[DoubleStruckE] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL E
+    '\uf7e3': '𝟠', # \[DoubleStruckEight] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK DIGIT EIGHT
+    '\uf6eb': '𝕗', # \[DoubleStruckF] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL F
+    '\uf7e0': '𝟝', # \[DoubleStruckFive] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK DIGIT FIVE
+    '\uf7df': '𝟜', # \[DoubleStruckFour] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK DIGIT FOUR
+    '\uf6ec': '𝕘', # \[DoubleStruckG] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL G
+    '\uf6ed': '𝕙', # \[DoubleStruckH] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL H
+    '\uf6ee': '𝕚', # \[DoubleStruckI] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL I
+    '\uf6ef': '𝕛', # \[DoubleStruckJ] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL J
+    '\uf6f0': '𝕜', # \[DoubleStruckK] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL K
+    '\uf6f1': '𝕝', # \[DoubleStruckL] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL L
+    '\uf6f2': '𝕞', # \[DoubleStruckM] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL M
+    '\uf6f3': '𝕟', # \[DoubleStruckN] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL N
+    '\uf7e4': '𝟡', # \[DoubleStruckNine] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK DIGIT NINE
+    '\uf6f4': '𝕠', # \[DoubleStruckO] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL O
+    '\uf7dc': '𝟙', # \[DoubleStruckOne] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK DIGIT ONE
+    '\uf6f5': '𝕡', # \[DoubleStruckP] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL P
+    '\uf6f6': '𝕢', # \[DoubleStruckQ] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL Q
+    '\uf6f7': '𝕣', # \[DoubleStruckR] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL R
+    '\uf6f8': '𝕤', # \[DoubleStruckS] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL S
+    '\uf7e2': '𝟟', # \[DoubleStruckSeven] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK DIGIT SEVEN
+    '\uf7e1': '𝟞', # \[DoubleStruckSix] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK DIGIT SIX
+    '\uf6f9': '𝕥', # \[DoubleStruckT] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL T
+    '\uf7de': '𝟛', # \[DoubleStruckThree] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK DIGIT THREE
+    '\uf7dd': '𝟚', # \[DoubleStruckTwo] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK DIGIT TWO
+    '\uf6fa': '𝕦', # \[DoubleStruckU] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL U
+    '\uf6fb': '𝕧', # \[DoubleStruckV] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL V
+    '\uf6fc': '𝕨', # \[DoubleStruckW] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL W
+    '\uf6fd': '𝕩', # \[DoubleStruckX] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL X
+    '\uf6fe': '𝕪', # \[DoubleStruckY] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL Y
+    '\uf6ff': '𝕫', # \[DoubleStruckZ] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK SMALL Z
+    '\uf7db': '𝟘', # \[DoubleStruckZero] (INVALID CHARACTER) -> MATHEMATICAL DOUBLE-STRUCK DIGIT ZERO
+    '⇑': '⇑', # \[DoubleUpArrow] (UPWARDS DOUBLE ARROW) -> UPWARDS DOUBLE ARROW
+    '⇕': '⇕', # \[DoubleUpDownArrow] (UP DOWN DOUBLE ARROW) -> UP DOWN DOUBLE ARROW
+    '∥': '∥', # \[DoubleVerticalBar] (PARALLEL TO) -> PARALLEL TO
+    '⤓': '⤓', # \[DownArrowBar] (DOWNWARDS ARROW TO BAR) -> DOWNWARDS ARROW TO BAR
+    '↓': '↓', # \[DownArrow] (DOWNWARDS ARROW) -> DOWNWARDS ARROW
+    '⇵': '⇵', # \[DownArrowUpArrow] (DOWNWARDS ARROW LEFTWARDS OF UPWARDS ARROW) -> DOWNWARDS ARROW LEFTWARDS OF UPWARDS ARROW
+    '\uf755': ' ̑', # \[DownBreve] (INVALID CHARACTER) -> COMBINING INVERTED BREVE
+    '¡': '¡', # \[DownExclamation] (INVERTED EXCLAMATION MARK) -> INVERTED EXCLAMATION MARK
+    '⥐': '⥐', # \[DownLeftRightVector] (LEFT BARB DOWN RIGHT BARB DOWN HARPOON) -> LEFT BARB DOWN RIGHT BARB DOWN HARPOON
+    '⥞': '⥞', # \[DownLeftTeeVector] (LEFTWARDS HARPOON WITH BARB DOWN FROM BAR) -> LEFTWARDS HARPOON WITH BARB DOWN FROM BAR
+    '↽': '↽', # \[DownLeftVector] (LEFTWARDS HARPOON WITH BARB DOWNWARDS) -> LEFTWARDS HARPOON WITH BARB DOWNWARDS
+    '⥖': '⥖', # \[DownLeftVectorBar] (LEFTWARDS HARPOON WITH BARB DOWN TO BAR) -> LEFTWARDS HARPOON WITH BARB DOWN TO BAR
+    '▾': '▾', # \[DownPointer] (BLACK DOWN-POINTING SMALL TRIANGLE) -> BLACK DOWN-POINTING SMALL TRIANGLE
+    '¿': '¿', # \[DownQuestion] (INVERTED QUESTION MARK) -> INVERTED QUESTION MARK
+    '⥟': '⥟', # \[DownRightTeeVector] (RIGHTWARDS HARPOON WITH BARB DOWN FROM BAR) -> RIGHTWARDS HARPOON WITH BARB DOWN FROM BAR
+    '⇁': '⇁', # \[DownRightVector] (RIGHTWARDS HARPOON WITH BARB DOWNWARDS) -> RIGHTWARDS HARPOON WITH BARB DOWNWARDS
+    '⥗': '⥗', # \[DownRightVectorBar] (RIGHTWARDS HARPOON WITH BARB DOWN TO BAR) -> RIGHTWARDS HARPOON WITH BARB DOWN TO BAR
+    '↧': '↧', # \[DownTeeArrow] (DOWNWARDS ARROW FROM BAR) -> DOWNWARDS ARROW FROM BAR
+    '⊤': '⊤', # \[DownTee] (DOWN TACK) -> DOWN TACK
+    'é': 'é', # \[EAcute] (LATIN SMALL LETTER E WITH ACUTE) -> LATIN SMALL LETTER E WITH ACUTE
+    '♁': '♁', # \[Earth] (EARTH) -> CIRCLED PLUS
+    'ē': 'ē', # \[EBar] (LATIN SMALL LETTER E WITH MACRON) -> LATIN SMALL LETTER E WITH MACRON
+    'ĕ': 'ĕ', # \[ECup] (LATIN SMALL LETTER E WITH BREVE) -> LATIN SMALL LETTER E WITH BREVE
+    'ë': 'ë', # \[EDoubleDot] (LATIN SMALL LETTER E WITH DIAERESIS) -> LATIN SMALL LETTER E WITH DIAERESIS
+    'è': 'è', # \[EGrave] (LATIN SMALL LETTER E WITH GRAVE) -> LATIN SMALL LETTER E WITH GRAVE
+    'ě': 'ě', # \[EHacek] (LATIN SMALL LETTER E WITH CARON) -> LATIN SMALL LETTER E WITH CARON
+    'ê': 'ê', # \[EHat] (LATIN SMALL LETTER E WITH CIRCUMFLEX) -> LATIN SMALL LETTER E WITH CIRCUMFLEX
+    '♪': '♪', # \[EighthNote] (EIGHTH NOTE) -> EIGHTH NOTE
+    '∈': '∈', # \[Element] (ELEMENT OF) -> ELEMENT OF
+    '…': '…', # \[Ellipsis] (HORIZONTAL ELLIPSIS) -> HORIZONTAL ELLIPSIS
+    '○': '○', # \[EmptyCircle] (WHITE CIRCLE) -> WHITE CIRCLE
+    '◇': '◇', # \[EmptyDiamond] (WHITE DIAMOND) -> WHITE DIAMOND
+    '▽': '▽', # \[EmptyDownTriangle] (WHITE DOWN-POINTING TRIANGLE) -> WHITE DOWN-POINTING TRIANGLE
+    '▯': '▯', # \[EmptyRectangle] (WHITE VERTICAL RECTANGLE) -> WHITE VERTICAL RECTANGLE
+    '∅': '∅', # \[EmptySet] (EMPTY SET) -> EMPTY SET
+    '◦': '◦', # \[EmptySmallCircle] (WHITE BULLET) -> WHITE BULLET
+    '◻': '◻', # \[EmptySmallSquare] (WHITE MEDIUM SQUARE) -> WHITE MEDIUM SQUARE
+    '□': '□', # \[EmptySquare] (WHITE SQUARE) -> WHITE SQUARE
+    '△': '△', # \[EmptyUpTriangle] (WHITE UP-POINTING TRIANGLE) -> WHITE UP-POINTING TRIANGLE
+    '▫': '▫', # \[EmptyVerySmallSquare] (WHITE SMALL SQUARE) -> WHITE SMALL SQUARE
+    'ϵ': 'ϵ', # \[Epsilon] (GREEK LUNATE EPSILON SYMBOL) -> GREEK LUNATE EPSILON SYMBOL
+    '\uf431': '⩵', # \[Equal] (INVALID CHARACTER) -> TWO CONSECUTIVE EQUALS SIGNS
+    '≂': '≂', # \[EqualTilde] (MINUS TILDE) -> MINUS TILDE
+    '⇌': '⇌', # \[Equilibrium] (RIGHTWARDS HARPOON OVER LEFTWARDS HARPOON) -> RIGHTWARDS HARPOON OVER LEFTWARDS HARPOON
+    '⧦': '⇔', # \[Equivalent] (GLEICH STARK) -> LEFT RIGHT DOUBLE ARROW
+    'η': 'η', # \[Eta] (GREEK SMALL LETTER ETA) -> GREEK SMALL LETTER ETA
+    'ð': 'ð', # \[Eth] (LATIN SMALL LETTER ETH) -> LATIN SMALL LETTER ETH
+    '€': '€', # \[Euro] (EURO SIGN) -> EURO SIGN
+    '∃': '∃', # \[Exists] (THERE EXISTS) -> THERE EXISTS
+    '\uf74d': 'ⅇ', # \[ExponentialE] (INVALID CHARACTER) -> DOUBLE-STRUCK ITALIC SMALL E
+    'ﬁ': 'ﬁ', # \[FiLigature] (LATIN SMALL LIGATURE FI) -> LATIN SMALL LIGATURE FI
+    '●': '●', # \[FilledCircle] (BLACK CIRCLE) -> BLACK CIRCLE
+    '◆': '◆', # \[FilledDiamond] (BLACK DIAMOND) -> BLACK DIAMOND
+    '▼': '▼', # \[FilledDownTriangle] (BLACK DOWN-POINTING TRIANGLE) -> BLACK DOWN-POINTING TRIANGLE
+    '◀': '◀', # \[FilledLeftTriangle] (BLACK LEFT-POINTING TRIANGLE) -> BLACK LEFT-POINTING TRIANGLE
+    '▮': '▮', # \[FilledRectangle] (BLACK VERTICAL RECTANGLE) -> BLACK VERTICAL RECTANGLE
+    '▶': '▶', # \[FilledRightTriangle] (BLACK RIGHT-POINTING TRIANGLE) -> BLACK RIGHT-POINTING TRIANGLE
+    '\uf750': '•', # \[FilledSmallCircle] (INVALID CHARACTER) -> BULLET
+    '◼': '◼', # \[FilledSmallSquare] (BLACK MEDIUM SQUARE) -> BLACK MEDIUM SQUARE
+    '■': '■', # \[FilledSquare] (BLACK SQUARE) -> BLACK SQUARE
+    '▲': '▲', # \[FilledUpTriangle] (BLACK UP-POINTING TRIANGLE) -> BLACK UP-POINTING TRIANGLE
+    '▪': '▪', # \[FilledVerySmallSquare] (BLACK SMALL SQUARE) -> BLACK SMALL SQUARE
+    'ς': 'ς', # \[FinalSigma] (GREEK SMALL LETTER FINAL SIGMA) -> GREEK SMALL LETTER FINAL SIGMA
+    '★': '★', # \[FivePointedStar] (BLACK STAR) -> BLACK STAR
+    '♭': '♭', # \[Flat] (MUSIC FLAT SIGN) -> MUSIC FLAT SIGN
+    'ﬂ': 'ﬂ', # \[FlLigature] (LATIN SMALL LIGATURE FL) -> LATIN SMALL LIGATURE FL
+    'ƒ': 'ƒ', # \[Florin] (LATIN SMALL LETTER F WITH HOOK) -> LATIN SMALL LETTER F WITH HOOK
+    '∀': '∀', # \[ForAll] (FOR ALL) -> FOR ALL
+    '\uf800': 'ạ', # \[FormalA] (INVALID CHARACTER) -> LATIN SMALL LETTER A WITH DOT BELOW
+    '\uf854': 'α̣', # \[FormalAlpha] (INVALID CHARACTER) -> GREEK SMALL LETTER ALPHA + COMBINING DOT BELOW
+    '\uf801': 'ḅ', # \[FormalB] (INVALID CHARACTER) -> LATIN SMALL LETTER B WITH DOT BELOW
+    '\uf855': 'β̣', # \[FormalBeta] (INVALID CHARACTER) -> GREEK SMALL LETTER BETA + COMBINING DOT BELOW
+    '\uf802': 'c̣', # \[FormalC] (INVALID CHARACTER) -> LATIN SMALL LETTER C + COMBINING DOT BELOW
+    '\uf81a': 'Ạ', # \[FormalCapitalA] (INVALID CHARACTER) -> LATIN CAPITAL LETTER A WITH DOT BELOW
+    '\uf834': 'Α̣', # \[FormalCapitalAlpha] (INVALID CHARACTER) -> GREEK CAPITAL LETTER ALPHA + COMBINING DOT BELOW
+    '\uf81b': 'Ḅ', # \[FormalCapitalB] (INVALID CHARACTER) -> LATIN CAPITAL LETTER B WITH DOT BELOW
+    '\uf835': 'Β̣', # \[FormalCapitalBeta] (INVALID CHARACTER) -> GREEK CAPITAL LETTER BETA + COMBINING DOT BELOW
+    '\uf81c': 'C̣', # \[FormalCapitalC] (INVALID CHARACTER) -> LATIN CAPITAL LETTER C + COMBINING DOT BELOW
+    '\uf84a': 'Χ̣', # \[FormalCapitalChi] (INVALID CHARACTER) -> GREEK CAPITAL LETTER CHI + COMBINING DOT BELOW
+    '\uf81d': 'Ḍ', # \[FormalCapitalD] (INVALID CHARACTER) -> LATIN CAPITAL LETTER D WITH DOT BELOW
+    '\uf837': 'Δ̣', # \[FormalCapitalDelta] (INVALID CHARACTER) -> GREEK CAPITAL LETTER DELTA + COMBINING DOT BELOW
+    '\uf87f': 'Ϝ̣', # \[FormalCapitalDigamma] (INVALID CHARACTER) -> GREEK LETTER DIGAMMA + COMBINING DOT BELOW
+    '\uf81e': 'Ẹ', # \[FormalCapitalE] (INVALID CHARACTER) -> LATIN CAPITAL LETTER E WITH DOT BELOW
+    '\uf838': 'Ε̣', # \[FormalCapitalEpsilon] (INVALID CHARACTER) -> GREEK CAPITAL LETTER EPSILON + COMBINING DOT BELOW
+    '\uf83a': 'Η̣', # \[FormalCapitalEta] (INVALID CHARACTER) -> GREEK CAPITAL LETTER ETA + COMBINING DOT BELOW
+    '\uf81f': 'F̣', # \[FormalCapitalF] (INVALID CHARACTER) -> LATIN CAPITAL LETTER F + COMBINING DOT BELOW
+    '\uf820': 'G̣', # \[FormalCapitalG] (INVALID CHARACTER) -> LATIN CAPITAL LETTER G + COMBINING DOT BELOW
+    '\uf836': 'Γ', # \[FormalCapitalGamma] (INVALID CHARACTER) -> GREEK CAPITAL LETTER GAMMA
+    '\uf821': 'Ḥ', # \[FormalCapitalH] (INVALID CHARACTER) -> LATIN CAPITAL LETTER H WITH DOT BELOW
+    '\uf822': 'Ị', # \[FormalCapitalI] (INVALID CHARACTER) -> LATIN CAPITAL LETTER I WITH DOT BELOW
+    '\uf83c': 'Ι̣', # \[FormalCapitalIota] (INVALID CHARACTER) -> GREEK CAPITAL LETTER IOTA + COMBINING DOT BELOW + COMBINING DOT BELOW
+    '\uf823': 'J̣', # \[FormalCapitalJ] (INVALID CHARACTER) -> LATIN CAPITAL LETTER J + COMBINING DOT BELOW
+    '\uf824': 'Ḳ', # \[FormalCapitalK] (INVALID CHARACTER) -> LATIN CAPITAL LETTER K WITH DOT BELOW
+    '\uf83d': 'Κ̣', # \[FormalCapitalKappa] (INVALID CHARACTER) -> GREEK CAPITAL LETTER KAPPA + COMBINING DOT BELOW
+    '\uf881': 'Ϟ̣', # \[FormalCapitalKoppa] (INVALID CHARACTER) -> GREEK LETTER KOPPA + COMBINING DOT BELOW
+    '\uf825': 'Ḷ', # \[FormalCapitalL] (INVALID CHARACTER) -> LATIN CAPITAL LETTER L WITH DOT BELOW
+    '\uf83e': 'Λ̣', # \[FormalCapitalLambda] (INVALID CHARACTER) -> GREEK CAPITAL LETTER LAMDA + COMBINING DOT BELOW
+    '\uf826': 'Ṃ', # \[FormalCapitalM] (INVALID CHARACTER) -> LATIN CAPITAL LETTER M WITH DOT BELOW
+    '\uf83f': 'Μ̣', # \[FormalCapitalMu] (INVALID CHARACTER) -> GREEK CAPITAL LETTER MU + COMBINING DOT BELOW
+    '\uf827': 'Ṇ', # \[FormalCapitalN] (INVALID CHARACTER) -> LATIN CAPITAL LETTER N WITH DOT BELOW
+    '\uf840': 'Ν̣', # \[FormalCapitalNu] (INVALID CHARACTER) -> GREEK CAPITAL LETTER NU + COMBINING DOT BELOW
+    '\uf828': 'Ọ', # \[FormalCapitalO] (INVALID CHARACTER) -> LATIN CAPITAL LETTER O WITH DOT BELOW
+    '\uf84c': 'Ω̣', # \[FormalCapitalOmega] (INVALID CHARACTER) -> GREEK CAPITAL LETTER OMEGA + COMBINING DOT BELOW
+    '\uf842': 'Ο̣', # \[FormalCapitalOmicron] (INVALID CHARACTER) -> GREEK CAPITAL LETTER OMICRON + COMBINING DOT BELOW
+    '\uf829': 'P̣', # \[FormalCapitalP] (INVALID CHARACTER) -> LATIN CAPITAL LETTER P + COMBINING DOT BELOW
+    '\uf849': 'Φ̣', # \[FormalCapitalPhi] (INVALID CHARACTER) -> GREEK CAPITAL LETTER PHI + COMBINING DOT BELOW
+    '\uf843': 'Π̣', # \[FormalCapitalPi] (INVALID CHARACTER) -> GREEK CAPITAL LETTER PI + COMBINING DOT BELOW
+    '\uf84b': 'Ψ̣', # \[FormalCapitalPsi] (INVALID CHARACTER) -> GREEK CAPITAL LETTER PSI + COMBINING DOT BELOW
+    '\uf82a': 'Q̣', # \[FormalCapitalQ] (INVALID CHARACTER) -> LATIN CAPITAL LETTER Q + COMBINING DOT BELOW
+    '\uf82b': 'Ṛ', # \[FormalCapitalR] (INVALID CHARACTER) -> LATIN CAPITAL LETTER R WITH DOT BELOW
+    '\uf844': 'Ρ̣', # \[FormalCapitalRho] (INVALID CHARACTER) -> GREEK CAPITAL LETTER RHO + COMBINING DOT BELOW
+    '\uf82c': 'Ṣ', # \[FormalCapitalS] (INVALID CHARACTER) -> LATIN CAPITAL LETTER S WITH DOT BELOW
+    '\uf883': 'Ϡ̣', # \[FormalCapitalSampi] (INVALID CHARACTER) -> GREEK LETTER SAMPI + COMBINING DOT BELOW
+    '\uf846': 'Σ̣', # \[FormalCapitalSigma] (INVALID CHARACTER) -> GREEK CAPITAL LETTER SIGMA + COMBINING DOT BELOW
+    '\uf87d': 'Ϛ̣', # \[FormalCapitalStigma] (INVALID CHARACTER) -> GREEK LETTER STIGMA + COMBINING DOT BELOW
+    '\uf82d': 'Ṭ', # \[FormalCapitalT] (INVALID CHARACTER) -> LATIN CAPITAL LETTER T WITH DOT BELOW
+    '\uf847': 'Τ̣', # \[FormalCapitalTau] (INVALID CHARACTER) -> GREEK CAPITAL LETTER TAU + COMBINING DOT BELOW
+    '\uf83b': 'Θ', # \[FormalCapitalTheta] (INVALID CHARACTER) -> GREEK CAPITAL LETTER THETA
+    '\uf82e': 'Ụ', # \[FormalCapitalU] (INVALID CHARACTER) -> LATIN CAPITAL LETTER U WITH DOT BELOW
+    '\uf848': 'Υ̣', # \[FormalCapitalUpsilon] (INVALID CHARACTER) -> GREEK CAPITAL LETTER UPSILON + COMBINING DOT BELOW
+    '\uf82f': 'Ṿ', # \[FormalCapitalV] (INVALID CHARACTER) -> LATIN CAPITAL LETTER V WITH DOT BELOW
+    '\uf830': 'Ẉ', # \[FormalCapitalW] (INVALID CHARACTER) -> LATIN CAPITAL LETTER W WITH DOT BELOW
+    '\uf831': 'X̣', # \[FormalCapitalX] (INVALID CHARACTER) -> LATIN CAPITAL LETTER X + COMBINING DOT BELOW
+    '\uf841': 'Ξ̣', # \[FormalCapitalXi] (INVALID CHARACTER) -> GREEK CAPITAL LETTER XI + COMBINING DOT BELOW
+    '\uf832': 'Ỵ', # \[FormalCapitalY] (INVALID CHARACTER) -> LATIN CAPITAL LETTER Y WITH DOT BELOW
+    '\uf833': 'Ẓ', # \[FormalCapitalZ] (INVALID CHARACTER) -> LATIN CAPITAL LETTER Z WITH DOT BELOW
+    '\uf839': 'Ζ̣', # \[FormalCapitalZeta] (INVALID CHARACTER) -> GREEK CAPITAL LETTER ZETA + COMBINING DOT BELOW
+    '\uf86a': 'χ̣', # \[FormalChi] (INVALID CHARACTER) -> GREEK SMALL LETTER CHI + COMBINING DOT BELOW
+    '\uf875': 'ϒ̣', # \[FormalCurlyCapitalUpsilon] (INVALID CHARACTER) -> GREEK UPSILON WITH HOOK SYMBOL + COMBINING DOT BELOW
+    '\uf858': 'ε̣', # \[FormalCurlyEpsilon] (INVALID CHARACTER) -> GREEK SMALL LETTER EPSILON + COMBINING DOT BELOW
+    '\uf885': 'ϰ̣', # \[FormalCurlyKappa] (INVALID CHARACTER) -> GREEK KAPPA SYMBOL + COMBINING DOT BELOW
+    '\uf869': 'φ̣', # \[FormalCurlyPhi] (INVALID CHARACTER) -> GREEK SMALL LETTER PHI + COMBINING DOT BELOW
+    '\uf879': 'ϖ̣', # \[FormalCurlyPi] (INVALID CHARACTER) -> GREEK PI SYMBOL + COMBINING DOT BELOW
+    '\uf886': 'ϱ̣', # \[FormalCurlyRho] (INVALID CHARACTER) -> GREEK RHO SYMBOL + COMBINING DOT BELOW
+    '\uf874': 'ϑ̣', # \[FormalCurlyTheta] (INVALID CHARACTER) -> GREEK THETA SYMBOL + COMBINING DOT BELOW
+    '\uf803': 'ḍ', # \[FormalD] (INVALID CHARACTER) -> LATIN SMALL LETTER D WITH DOT BELOW
+    '\uf857': 'δ̣', # \[FormalDelta] (INVALID CHARACTER) -> GREEK SMALL LETTER DELTA + COMBINING DOT BELOW
+    '\uf880': 'ϝ', # \[FormalDigamma] (INVALID CHARACTER) -> GREEK SMALL LETTER DIGAMMA
+    '\uf804': 'ẹ', # \[FormalE] (INVALID CHARACTER) -> LATIN SMALL LETTER E WITH DOT BELOW
+    '\uf88a': 'ϵ̣', # \[FormalEpsilon] (INVALID CHARACTER) -> GREEK LUNATE EPSILON SYMBOL + COMBINING DOT BELOW
+    '\uf85a': 'η̣', # \[FormalEta] (INVALID CHARACTER) -> GREEK SMALL LETTER ETA + COMBINING DOT BELOW
+    '\uf805': 'f̣', # \[FormalF] (INVALID CHARACTER) -> LATIN SMALL LETTER F + COMBINING DOT BELOW
+    '\uf865': 'ς̣', # \[FormalFinalSigma] (INVALID CHARACTER) -> GREEK SMALL LETTER FINAL SIGMA + COMBINING DOT BELOW
+    '\uf806': 'g̣', # \[FormalG] (INVALID CHARACTER) -> LATIN SMALL LETTER G + COMBINING DOT BELOW
+    '\uf856': 'γ̣', # \[FormalGamma] (INVALID CHARACTER) -> GREEK SMALL LETTER GAMMA + COMBINING DOT BELOW
+    '\uf807': 'ḥ', # \[FormalH] (INVALID CHARACTER) -> LATIN SMALL LETTER H WITH DOT BELOW
+    '\uf808': 'ị', # \[FormalI] (INVALID CHARACTER) -> LATIN SMALL LETTER I WITH DOT BELOW
+    '\uf85c': 'Ι̣', # \[FormalIota] (INVALID CHARACTER) -> GREEK CAPITAL LETTER IOTA + COMBINING DOT BELOW + COMBINING DOT BELOW
+    '\uf809': 'j̣', # \[FormalJ] (INVALID CHARACTER) -> LATIN SMALL LETTER J + COMBINING DOT BELOW
+    '\uf80a': 'ḳ', # \[FormalK] (INVALID CHARACTER) -> LATIN SMALL LETTER K WITH DOT BELOW
+    '\uf85d': 'κ̣', # \[FormalKappa] (INVALID CHARACTER) -> GREEK SMALL LETTER KAPPA + COMBINING DOT BELOW
+    '\uf882': 'ϟ̣', # \[FormalKoppa] (INVALID CHARACTER) -> GREEK SMALL LETTER KOPPA + COMBINING DOT BELOW
+    '\uf80b': 'ḷ', # \[FormalL] (INVALID CHARACTER) -> LATIN SMALL LETTER L WITH DOT BELOW
+    '\uf85e': 'λ̣', # \[FormalLambda] (INVALID CHARACTER) -> GREEK SMALL LETTER LAMDA + COMBINING DOT BELOW
+    '\uf80c': 'ṃ', # \[FormalM] (INVALID CHARACTER) -> LATIN SMALL LETTER M + COMBINING DOT BELOW
+    '\uf85f': 'μ̣', # \[FormalMu] (INVALID CHARACTER) -> GREEK SMALL LETTER MU + COMBINING DOT BELOW
+    '\uf80d': 'ṇ', # \[FormalN] (INVALID CHARACTER) -> LATIN SMALL LETTER N + COMBINING DOT BELOW
+    '\uf860': 'ν̣', # \[FormalNu] (INVALID CHARACTER) -> GREEK SMALL LETTER NU + COMBINING DOT BELOW
+    '\uf80e': 'ọ', # \[FormalO] (INVALID CHARACTER) -> LATIN SMALL LETTER O + COMBINING DOT BELOW
+    '\uf86c': 'ω̣', # \[FormalOmega] (INVALID CHARACTER) -> GREEK SMALL LETTER OMEGA + COMBINING DOT BELOW
+    '\uf862': 'ο̣', # \[FormalOmicron] (INVALID CHARACTER) -> GREEK SMALL LETTER OMICRON + COMBINING DOT BELOW
+    '\uf80f': 'p̣', # \[FormalP] (INVALID CHARACTER) -> LATIN SMALL LETTER P + COMBINING DOT BELOW
+    '\uf878': 'ϕ̣', # \[FormalPhi] (INVALID CHARACTER) -> GREEK PHI SYMBOL + COMBINING DOT BELOW
+    '\uf863': 'π̣', # \[FormalPi] (INVALID CHARACTER) -> GREEK SMALL LETTER PI + COMBINING DOT BELOW
+    '\uf86b': 'ψ̣', # \[FormalPsi] (INVALID CHARACTER) -> GREEK SMALL LETTER PSI + COMBINING DOT BELOW
+    '\uf810': 'q̣', # \[FormalQ] (INVALID CHARACTER) -> LATIN SMALL LETTER Q + COMBINING DOT BELOW
+    '\uf811': 'ṛ', # \[FormalR] (INVALID CHARACTER) -> LATIN SMALL LETTER R WITH DOT BELOW
+    '\uf864': 'ρ̣', # \[FormalRho] (INVALID CHARACTER) -> GREEK SMALL LETTER RHO + COMBINING DOT BELOW
+    '\uf812': 'ṣ', # \[FormalS] (INVALID CHARACTER) -> LATIN SMALL LETTER S WITH DOT BELOW
+    '\uf884': 'ϡ̣', # \[FormalSampi] (INVALID CHARACTER) -> GREEK SMALL LETTER SAMPI + COMBINING DOT BELOW
+    '\uf866': 'σ̣', # \[FormalSigma] (INVALID CHARACTER) -> GREEK SMALL LETTER SIGMA + COMBINING DOT BELOW
+    '\uf87e': 'ϛ', # \[FormalStigma] (INVALID CHARACTER) -> GREEK SMALL LETTER STIGMA
+    '\uf813': 'ṭ', # \[FormalT] (INVALID CHARACTER) -> LATIN SMALL LETTER T WITH DOT BELOW
+    '\uf867': 'τ̣', # \[FormalTau] (INVALID CHARACTER) -> GREEK SMALL LETTER TAU + COMBINING DOT BELOW
+    '\uf85b': 'θ̣', # \[FormalTheta] (INVALID CHARACTER) -> GREEK SMALL LETTER THETA + COMBINING DOT BELOW
+    '\uf814': 'ụ', # \[FormalU] (INVALID CHARACTER) -> LATIN SMALL LETTER U WITH DOT BELOW
+    '\uf868': 'υ̣', # \[FormalUpsilon] (INVALID CHARACTER) -> GREEK SMALL LETTER UPSILON + COMBINING DOT BELOW
+    '\uf815': 'ṿ', # \[FormalV] (INVALID CHARACTER) -> LATIN SMALL LETTER V WITH DOT BELOW
+    '\uf816': 'ẉ', # \[FormalW] (INVALID CHARACTER) -> LATIN SMALL LETTER W WITH DOT BELOW
+    '\uf817': 'x̣', # \[FormalX] (INVALID CHARACTER) -> LATIN SMALL LETTER X + COMBINING DOT BELOW
+    '\uf861': 'ξ̣', # \[FormalXi] (INVALID CHARACTER) -> GREEK SMALL LETTER XI + COMBINING DOT BELOW
+    '\uf818': 'ỵ', # \[FormalY] (INVALID CHARACTER) -> LATIN SMALL LETTER Y WITH DOT BELOW
+    '\uf819': 'ẓ', # \[FormalZ] (INVALID CHARACTER) -> LATIN SMALL LETTER Z WITH DOT BELOW
+    '\uf859': 'ζ̣', # \[FormalZeta] (INVALID CHARACTER) -> GREEK SMALL LETTER ZETA + COMBINING DOT BELOW
+    '\uf4a1': '↦', # \[Function] (INVALID CHARACTER) -> RIGHTWARDS ARROW FROM BAR
+    'γ': 'γ', # \[Gamma] (GREEK SMALL LETTER GAMMA) -> GREEK SMALL LETTER GAMMA
+    '♊': '♊', # \[GeminiSign] (GEMINI) -> GEMINI
+    'ℷ': 'ℷ', # \[Gimel] (GIMEL SYMBOL) -> GIMEL SYMBOL
+    '\uf6cc': '𝔞', # \[GothicA] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL A
+    '\uf6cd': '𝔟', # \[GothicB] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL B
+    '\uf6ce': '𝔠', # \[GothicC] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL C
+    '\uf78a': '𝔄', # \[GothicCapitalA] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL A
+    '\uf78b': '𝔅', # \[GothicCapitalB] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL B
+    'ℭ': 'ℭ', # \[GothicCapitalC] (BLACK-LETTER CAPITAL C) -> BLACK-LETTER CAPITAL C
+    '\uf78d': '𝔇', # \[GothicCapitalD] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL D
+    '\uf78e': '𝔈', # \[GothicCapitalE] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL E
+    '\uf78f': '𝔉', # \[GothicCapitalF] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL F
+    '\uf790': '𝔊', # \[GothicCapitalG] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL G
+    'ℌ': 'ℌ', # \[GothicCapitalH] (BLACK-LETTER CAPITAL H) -> BLACK-LETTER CAPITAL H
+    'ℑ': 'ℑ', # \[GothicCapitalI] (BLACK-LETTER CAPITAL I) -> BLACK-LETTER CAPITAL I
+    '\uf793': '𝔍', # \[GothicCapitalJ] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL J
+    '\uf794': '𝔎', # \[GothicCapitalK] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL K
+    '\uf795': '𝔏', # \[GothicCapitalL] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL L
+    '\uf796': '𝔐', # \[GothicCapitalM] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL M
+    '\uf797': '𝔑', # \[GothicCapitalN] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL N
+    '\uf798': '𝔒', # \[GothicCapitalO] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL O
+    '\uf799': '𝔓', # \[GothicCapitalP] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL P
+    '\uf79a': '𝔔', # \[GothicCapitalQ] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL Q
+    'ℜ': 'ℜ', # \[GothicCapitalR] (BLACK-LETTER CAPITAL R) -> BLACK-LETTER CAPITAL R
+    '\uf79c': '𝔖', # \[GothicCapitalS] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL S
+    '\uf79d': '𝔗', # \[GothicCapitalT] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL T
+    '\uf79e': '𝔘', # \[GothicCapitalU] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL U
+    '\uf79f': '𝔙', # \[GothicCapitalV] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL V
+    '\uf7a0': '𝔚', # \[GothicCapitalW] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL W
+    '\uf7a1': '𝔛', # \[GothicCapitalX] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL X
+    '\uf7a2': '𝔜', # \[GothicCapitalY] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR CAPITAL Y
+    'ℨ': 'ℨ', # \[GothicCapitalZ] (BLACK-LETTER CAPITAL Z) -> BLACK-LETTER CAPITAL Z
+    '\uf6cf': '𝔡', # \[GothicD] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL D
+    '\uf6d0': '𝔢', # \[GothicE] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL E
+    '\uf6d1': '𝔣', # \[GothicF] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL F
+    '\uf6d2': '𝔤', # \[GothicG] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL G
+    '\uf6d3': '𝔥', # \[GothicH] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL H
+    '\uf6d4': '𝔦', # \[GothicI] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL I
+    '\uf6d5': '𝔧', # \[GothicJ] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL J
+    '\uf6d6': '𝔨', # \[GothicK] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL K
+    '\uf6d7': '𝔩', # \[GothicL] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL L
+    '\uf6d8': '𝔪', # \[GothicM] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL M
+    '\uf6d9': '𝔫', # \[GothicN] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL N
+    '\uf6da': '𝔬', # \[GothicO] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL O
+    '\uf6db': '𝔭', # \[GothicP] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL P
+    '\uf6dc': '𝔮', # \[GothicQ] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL Q
+    '\uf6dd': '𝔯', # \[GothicR] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL R
+    '\uf6de': '𝔰', # \[GothicS] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL S
+    '\uf6df': '𝔱', # \[GothicT] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL T
+    '\uf6e0': '𝔲', # \[GothicU] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL U
+    '\uf6e1': '𝔳', # \[GothicV] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL V
+    '\uf6e2': '𝔴', # \[GothicW] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL W
+    '\uf6e3': '𝔵', # \[GothicX] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL X
+    '\uf6e4': '𝔶', # \[GothicY] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL Y
+    '\uf6e5': '𝔷', # \[GothicZ] (INVALID CHARACTER) -> MATHEMATICAL FRAKTUR SMALL Z
+    '\uf753': '●', # \[GrayCircle] (INVALID CHARACTER) -> BLACK CIRCLE
+    '\uf752': '■', # \[GraySquare] (INVALID CHARACTER) -> BLACK SQUARE
+    '⋛': '⋛', # \[GreaterEqualLess] (GREATER-THAN EQUAL TO OR LESS-THAN) -> GREATER-THAN EQUAL TO OR LESS-THAN
+    '≥': '≥', # \[GreaterEqual] (GREATER-THAN OR EQUAL TO) -> GREATER-THAN OR EQUAL TO
+    '≧': '≧', # \[GreaterFullEqual] (GREATER-THAN OVER EQUAL TO) -> GREATER-THAN OVER EQUAL TO
+    '≫': '≫', # \[GreaterGreater] (MUCH GREATER-THAN) -> MUCH GREATER-THAN
+    '≷': '≷', # \[GreaterLess] (GREATER-THAN OR LESS-THAN) -> GREATER-THAN OR LESS-THAN
+    '⩾': '⩾', # \[GreaterSlantEqual] (GREATER-THAN OR SLANTED EQUAL TO) -> GREATER-THAN OR SLANTED EQUAL TO
+    '≳': '≳', # \[GreaterTilde] (GREATER-THAN OR EQUIVALENT TO) -> GREATER-THAN OR EQUIVALENT TO
+    'ˇ': 'ˇ', # \[Hacek] (CARON) -> CARON
+    '☺': '☺', # \[HappySmiley] (WHITE SMILING FACE) -> WHITE SMILING FACE
+    'ℏ': 'ℏ', # \[HBar] (PLANCK CONSTANT OVER TWO PI) -> PLANCK CONSTANT OVER TWO PI
+    '♡': '♡', # \[HeartSuit] (WHITE HEART SUIT) -> WHITE HEART SUIT
+    '\uf3ce': '\uf3ce', # \[HermitianConjugate] (INVALID CHARACTER) -> 
+    '─': '─', # \[HorizontalLine] (BOX DRAWINGS LIGHT HORIZONTAL) -> BOX DRAWINGS LIGHT HORIZONTAL
+    '≎': '≎', # \[HumpDownHump] (GEOMETRICALLY EQUIVALENT TO) -> GEOMETRICALLY EQUIVALENT TO
+    '≏': '≏', # \[HumpEqual] (DIFFERENCE BETWEEN) -> DIFFERENCE BETWEEN
+    '‐': '‐', # \[Hyphen] (HYPHEN) -> HYPHEN
+    'í': 'í', # \[IAcute] (LATIN SMALL LETTER I WITH ACUTE) -> LATIN SMALL LETTER I WITH ACUTE
+    'ĭ': 'ĭ', # \[ICup] (LATIN SMALL LETTER I WITH BREVE) -> LATIN SMALL LETTER I WITH BREVE
+    'ï': 'ï', # \[IDoubleDot] (LATIN SMALL LETTER I WITH DIAERESIS) -> LATIN SMALL LETTER I WITH DIAERESIS
+    'ì': 'ì', # \[IGrave] (LATIN SMALL LETTER I WITH GRAVE) -> LATIN SMALL LETTER I WITH GRAVE
+    'î': 'î', # \[IHat] (LATIN SMALL LETTER I WITH CIRCUMFLEX) -> LATIN SMALL LETTER I WITH CIRCUMFLEX
+    '\uf74e': 'ⅈ', # \[ImaginaryI] (INVALID CHARACTER) -> DOUBLE-STRUCK ITALIC SMALL I
+    '\uf74f': 'ⅉ', # \[ImaginaryJ] (INVALID CHARACTER) -> DOUBLE-STRUCK ITALIC SMALL J
+    '\uf523': '⟹', # \[Implies] (INVALID CHARACTER) -> LONG RIGHTWARDS DOUBLE ARROW
+    '∞': '∞', # \[Infinity] (INFINITY) -> INFINITY
+    '∫': '∫', # \[Integral] (INTEGRAL) -> INTEGRAL
+    '⋂': '⋂', # \[Intersection] (N-ARY INTERSECTION) -> N-ARY INTERSECTION
+    'ι': 'ι', # \[Iota] (GREEK SMALL LETTER IOTA) -> GREEK SMALL LETTER IOTA
+    '♃': '♃', # \[Jupiter] (JUPITER) -> JUPITER
+    'κ': 'κ', # \[Kappa] (GREEK SMALL LETTER KAPPA) -> GREEK SMALL LETTER KAPPA
+    'ϟ': 'ϟ', # \[Koppa] (GREEK SMALL LETTER KOPPA) -> GREEK SMALL LETTER KOPPA
+    'λ': 'λ', # \[Lambda] (GREEK SMALL LETTER LAMDA) -> GREEK SMALL LETTER LAMDA
+    '〈': '〈', # \[LeftAngleBracket] (LEFT-POINTING ANGLE BRACKET) -> LEFT-POINTING ANGLE BRACKET
+    '⇤': '⇤', # \[LeftArrowBar] (LEFTWARDS ARROW TO BAR) -> LEFTWARDS ARROW TO BAR
+    '←': '←', # \[LeftArrow] (LEFTWARDS ARROW) -> LEFTWARDS ARROW
+    '⇆': '⇆', # \[LeftArrowRightArrow] (LEFTWARDS ARROW OVER RIGHTWARDS ARROW) -> LEFTWARDS ARROW OVER RIGHTWARDS ARROW
+    '\uf603': '|', # \[LeftBracketingBar] (INVALID CHARACTER) -> VERTICAL LINE
+    '⌈': '⌈', # \[LeftCeiling] (LEFT CEILING) -> LEFT CEILING
+    '〚': '〚', # \[LeftDoubleBracket] (LEFT WHITE SQUARE BRACKET) -> LEFT WHITE SQUARE BRACKET
+    '\uf605': '‖', # \[LeftDoubleBracketingBar] (INVALID CHARACTER) -> DOUBLE VERTICAL LINE
+    '⥡': '⥡', # \[LeftDownTeeVector] (DOWNWARDS HARPOON WITH BARB LEFT FROM BAR) -> DOWNWARDS HARPOON WITH BARB LEFT FROM BAR
+    '⥙': '⥙', # \[LeftDownVectorBar] (DOWNWARDS HARPOON WITH BARB LEFT TO BAR) -> DOWNWARDS HARPOON WITH BARB LEFT TO BAR
+    '⇃': '⇃', # \[LeftDownVector] (DOWNWARDS HARPOON WITH BARB LEFTWARDS) -> DOWNWARDS HARPOON WITH BARB LEFTWARDS
+    '⌊': '⌊', # \[LeftFloor] (LEFT FLOOR) -> LEFT FLOOR
+    '«': '«', # \[LeftGuillemet] (LEFT-POINTING DOUBLE ANGLE QUOTATION MARK) -> LEFT-POINTING DOUBLE ANGLE QUOTATION MARK
+    '◂': '◂', # \[LeftPointer] (BLACK LEFT-POINTING SMALL TRIANGLE) -> BLACK LEFT-POINTING SMALL TRIANGLE
+    '↔': '↔', # \[LeftRightArrow] (LEFT RIGHT ARROW) -> LEFT RIGHT ARROW
+    '⥎': '⥎', # \[LeftRightVector] (LEFT BARB UP RIGHT BARB UP HARPOON) -> LEFT BARB UP RIGHT BARB UP HARPOON
+    '\uf761': '«', # \[LeftSkeleton] (INVALID CHARACTER) -> LEFT-POINTING DOUBLE ANGLE QUOTATION MARK
+    '⊣': '⊣', # \[LeftTee] (LEFT TACK) -> LEFT TACK
+    '↤': '↤', # \[LeftTeeArrow] (LEFTWARDS ARROW FROM BAR) -> LEFTWARDS ARROW FROM BAR
+    '⥚': '⥚', # \[LeftTeeVector] (LEFTWARDS HARPOON WITH BARB UP FROM BAR) -> LEFTWARDS HARPOON WITH BARB UP FROM BAR
+    '⊲': '⊲', # \[LeftTriangle] (NORMAL SUBGROUP OF) -> NORMAL SUBGROUP OF
+    '⧏': '⧏', # \[LeftTriangleBar] (LEFT TRIANGLE BESIDE VERTICAL BAR) -> LEFT TRIANGLE BESIDE VERTICAL BAR
+    '⊴': '⊴', # \[LeftTriangleEqual] (NORMAL SUBGROUP OF OR EQUAL TO) -> NORMAL SUBGROUP OF OR EQUAL TO
+    '⥑': '⥑', # \[LeftUpDownVector] (UP BARB LEFT DOWN BARB LEFT HARPOON) -> UP BARB LEFT DOWN BARB LEFT HARPOON
+    '⥠': '⥠', # \[LeftUpTeeVector] (UPWARDS HARPOON WITH BARB LEFT FROM BAR) -> UPWARDS HARPOON WITH BARB LEFT FROM BAR
+    '↿': '↿', # \[LeftUpVector] (UPWARDS HARPOON WITH BARB LEFTWARDS) -> UPWARDS HARPOON WITH BARB LEFTWARDS
+    '⥘': '⥘', # \[LeftUpVectorBar] (UPWARDS HARPOON WITH BARB LEFT TO BAR) -> UPWARDS HARPOON WITH BARB LEFT TO BAR
+    '↼': '↼', # \[LeftVector] (LEFTWARDS HARPOON WITH BARB UPWARDS) -> LEFTWARDS HARPOON WITH BARB UPWARDS
+    '⥒': '⥒', # \[LeftVectorBar] (LEFTWARDS HARPOON WITH BARB UP TO BAR) -> LEFTWARDS HARPOON WITH BARB UP TO BAR
+    '♌': '♌', # \[LeoSign] (LEO) -> LEO
+    '≤': '≤', # \[LessEqual] (LESS-THAN OR EQUAL TO) -> LESS-THAN OR EQUAL TO
+    '⋚': '⋚', # \[LessEqualGreater] (LESS-THAN EQUAL TO OR GREATER-THAN) -> LESS-THAN EQUAL TO OR GREATER-THAN
+    '≦': '≦', # \[LessFullEqual] (LESS-THAN OVER EQUAL TO) -> LESS-THAN OVER EQUAL TO
+    '≶': '≶', # \[LessGreater] (LESS-THAN OR GREATER-THAN) -> LESS-THAN OR GREATER-THAN
+    '≪': '≪', # \[LessLess] (MUCH LESS-THAN) -> MUCH LESS-THAN
+    '⩽': '⩽', # \[LessSlantEqual] (LESS-THAN OR SLANTED EQUAL TO) -> LESS-THAN OR SLANTED EQUAL TO
+    '≲': '≲', # \[LessTilde] (LESS-THAN OR EQUIVALENT TO) -> LESS-THAN OR EQUIVALENT TO
+    '♎': '♎', # \[LibraSign] (LIBRA) -> LIBRA
+    '—': '—', # \[LongDash] (EM DASH) -> EM DASH
+    '⟵': '⟵', # \[LongLeftArrow] (LONG LEFTWARDS ARROW) -> LONG LEFTWARDS ARROW
+    '⟷': '⟷', # \[LongLeftRightArrow] (LONG LEFT RIGHT ARROW) -> LONG LEFT RIGHT ARROW
+    '⟶': '⟶', # \[LongRightArrow] (LONG RIGHTWARDS ARROW) -> LONG RIGHTWARDS ARROW
+    '↙': '↙', # \[LowerLeftArrow] (SOUTH WEST ARROW) -> SOUTH WEST ARROW
+    '↘': '↘', # \[LowerRightArrow] (SOUTH EAST ARROW) -> SOUTH EAST ARROW
+    'ł': 'ł', # \[LSlash] (LATIN SMALL LETTER L WITH STROKE) -> LATIN SMALL LETTER L WITH STROKE
+    '♂': '♂', # \[Mars] (MALE SIGN) -> MALE SIGN
+    '∡': '∡', # \[MeasuredAngle] (MEASURED ANGLE) -> MEASURED ANGLE
+    '☿': '☿', # \[Mercury] (MERCURY) -> MERCURY
+    '℧': '℧', # \[Mho] (INVERTED OHM SIGN) -> INVERTED OHM SIGN
+    'µ': 'µ', # \[Micro] (MICRO SIGN) -> MICRO SIGN
+    '∓': '∓', # \[MinusPlus] (MINUS-OR-PLUS SIGN) -> MINUS-OR-PLUS SIGN
+    'μ': 'μ', # \[Mu] (GREEK SMALL LETTER MU) -> GREEK SMALL LETTER MU
+    '⊼': '⊼', # \[Nand] (NAND) -> NAND
+    '♮': '♮', # \[Natural] (MUSIC NATURAL SIGN) -> MUSIC NATURAL SIGN
+    '♆': '♆', # \[Neptune] (NEPTUNE) -> NEPTUNE
+    '⪢': '⪢', # \[NestedGreaterGreater] (DOUBLE NESTED GREATER-THAN) -> DOUBLE NESTED GREATER-THAN
+    '⪡': '⪡', # \[NestedLessLess] (DOUBLE NESTED LESS-THAN) -> DOUBLE NESTED LESS-THAN
+    'ň': 'ň', # \[NHacek] (LATIN SMALL LETTER N WITH CARON) -> LATIN SMALL LETTER N WITH CARON
+    '⊽': '⊽', # \[Nor] (NOR) -> NOR
+    '≢': '≢', # \[NotCongruent] (NOT IDENTICAL TO) -> NOT IDENTICAL TO
+    '≭': '≭', # \[NotCupCap] (NOT EQUIVALENT TO) -> NOT EQUIVALENT TO
+    '∦': '∦', # \[NotDoubleVerticalBar] (NOT PARALLEL TO) -> NOT PARALLEL TO
+    '∉': '∉', # \[NotElement] (NOT AN ELEMENT OF) -> NOT AN ELEMENT OF
+    '≠': '≠', # \[NotEqual] (NOT EQUAL TO) -> NOT EQUAL TO
+    '∄': '∄', # \[NotExists] (THERE DOES NOT EXIST) -> THERE DOES NOT EXIST
+    '≯': '≯', # \[NotGreater] (NOT GREATER-THAN) -> NOT GREATER-THAN
+    '≱': '≱', # \[NotGreaterEqual] (NEITHER GREATER-THAN NOR EQUAL TO) -> NEITHER GREATER-THAN NOR EQUAL TO
+    '≩': '≩', # \[NotGreaterFullEqual] (GREATER-THAN BUT NOT EQUAL TO) -> GREATER-THAN BUT NOT EQUAL TO
+    '≹': '≹', # \[NotGreaterLess] (NEITHER GREATER-THAN NOR LESS-THAN) -> NEITHER GREATER-THAN NOR LESS-THAN
+    '≵': '≵', # \[NotGreaterTilde] (NEITHER GREATER-THAN NOR EQUIVALENT TO) -> NEITHER GREATER-THAN NOR EQUIVALENT TO
+    '⋪': '⋪', # \[NotLeftTriangle] (NOT NORMAL SUBGROUP OF) -> NOT NORMAL SUBGROUP OF
+    '⋬': '⋬', # \[NotLeftTriangleEqual] (NOT NORMAL SUBGROUP OF OR EQUAL TO) -> NOT NORMAL SUBGROUP OF OR EQUAL TO
+    '≰': '≰', # \[NotLessEqual] (NEITHER LESS-THAN NOR EQUAL TO) -> NEITHER LESS-THAN NOR EQUAL TO
+    '≨': '≨', # \[NotLessFullEqual] (LESS-THAN BUT NOT EQUAL TO) -> LESS-THAN BUT NOT EQUAL TO
+    '≸': '≸', # \[NotLessGreater] (NEITHER LESS-THAN NOR GREATER-THAN) -> NEITHER LESS-THAN NOR GREATER-THAN
+    '≮': '≮', # \[NotLess] (NOT LESS-THAN) -> NOT LESS-THAN
+    '≴': '≴', # \[NotLessTilde] (NEITHER LESS-THAN NOR EQUIVALENT TO) -> NEITHER LESS-THAN NOR EQUIVALENT TO
+    '⊀': '⊀', # \[NotPrecedes] (DOES NOT PRECEDE) -> DOES NOT PRECEDE
+    '⋠': '⋠', # \[NotPrecedesSlantEqual] (DOES NOT PRECEDE OR EQUAL) -> DOES NOT PRECEDE OR EQUAL
+    '⋨': '⋨', # \[NotPrecedesTilde] (PRECEDES BUT NOT EQUIVALENT TO) -> PRECEDES BUT NOT EQUIVALENT TO
+    '∌': '∌', # \[NotReverseElement] (DOES NOT CONTAIN AS MEMBER) -> DOES NOT CONTAIN AS MEMBER
+    '⋫': '⋫', # \[NotRightTriangle] (DOES NOT CONTAIN AS NORMAL SUBGROUP) -> DOES NOT CONTAIN AS NORMAL SUBGROUP
+    '⋭': '⋭', # \[NotRightTriangleEqual] (DOES NOT CONTAIN AS NORMAL SUBGROUP OR EQUAL) -> DOES NOT CONTAIN AS NORMAL SUBGROUP OR EQUAL
+    '⋢': '⋢', # \[NotSquareSubsetEqual] (NOT SQUARE IMAGE OF OR EQUAL TO) -> NOT SQUARE IMAGE OF OR EQUAL TO
+    '⋣': '⋣', # \[NotSquareSupersetEqual] (NOT SQUARE ORIGINAL OF OR EQUAL TO) -> NOT SQUARE ORIGINAL OF OR EQUAL TO
+    '⊄': '⊄', # \[NotSubset] (NOT A SUBSET OF) -> NOT A SUBSET OF
+    '⊈': '⊈', # \[NotSubsetEqual] (NEITHER A SUBSET OF NOR EQUAL TO) -> NEITHER A SUBSET OF NOR EQUAL TO
+    '⊁': '⊁', # \[NotSucceeds] (DOES NOT SUCCEED) -> DOES NOT SUCCEED
+    '⋡': '⋡', # \[NotSucceedsSlantEqual] (DOES NOT SUCCEED OR EQUAL) -> DOES NOT SUCCEED OR EQUAL
+    '⋩': '⋩', # \[NotSucceedsTilde] (SUCCEEDS BUT NOT EQUIVALENT TO) -> SUCCEEDS BUT NOT EQUIVALENT TO
+    '⊅': '⊅', # \[NotSuperset] (NOT A SUPERSET OF) -> NOT A SUPERSET OF
+    '⊉': '⊉', # \[NotSupersetEqual] (NEITHER A SUPERSET OF NOR EQUAL TO) -> NEITHER A SUPERSET OF NOR EQUAL TO
+    '≁': '≁', # \[NotTilde] (NOT TILDE) -> NOT TILDE
+    '≄': '≄', # \[NotTildeEqual] (NOT ASYMPTOTICALLY EQUAL TO) -> NOT ASYMPTOTICALLY EQUAL TO
+    '≇': '≇', # \[NotTildeFullEqual] (NEITHER APPROXIMATELY NOR ACTUALLY EQUAL TO) -> NEITHER APPROXIMATELY NOR ACTUALLY EQUAL TO
+    '≉': '≉', # \[NotTildeTilde] (NOT ALMOST EQUAL TO) -> NOT ALMOST EQUAL TO
+    'ñ': 'ñ', # \[NTilde] (LATIN SMALL LETTER N WITH TILDE) -> LATIN SMALL LETTER N WITH TILDE
+    'ν': 'ν', # \[Nu] (GREEK SMALL LETTER NU) -> GREEK SMALL LETTER NU
+    '\uf724': '#', # \[NumberSign] (INVALID CHARACTER) -> NUMBER SIGN
+    'ó': 'ó', # \[OAcute] (LATIN SMALL LETTER O WITH ACUTE) -> LATIN SMALL LETTER O WITH ACUTE
+    'ő': 'ő', # \[ODoubleAcute] (LATIN SMALL LETTER O WITH DOUBLE ACUTE) -> LATIN SMALL LETTER O WITH DOUBLE ACUTE
+    'ö': 'ö', # \[ODoubleDot] (LATIN SMALL LETTER O WITH DIAERESIS) -> LATIN SMALL LETTER O WITH DIAERESIS
+    'œ': 'œ', # \[OE] (LATIN SMALL LIGATURE OE) -> LATIN SMALL LIGATURE OE
+    'ò': 'ò', # \[OGrave] (LATIN SMALL LETTER O WITH GRAVE) -> LATIN SMALL LETTER O WITH GRAVE
+    'ô': 'ô', # \[OHat] (LATIN SMALL LETTER O WITH CIRCUMFLEX) -> LATIN SMALL LETTER O WITH CIRCUMFLEX
+    'ω': 'ω', # \[Omega] (GREEK SMALL LETTER OMEGA) -> GREEK SMALL LETTER OMEGA
+    'ο': 'ο', # \[Omicron] (GREEK SMALL LETTER OMICRON) -> GREEK SMALL LETTER OMICRON
+    '“': '“', # \[OpenCurlyDoubleQuote] (LEFT DOUBLE QUOTATION MARK) -> LEFT DOUBLE QUOTATION MARK
+    '‘': '‘', # \[OpenCurlyQuote] (LEFT SINGLE QUOTATION MARK) -> LEFT SINGLE QUOTATION MARK
+    '∨': '∨', # \[Or] (LOGICAL OR) -> LOGICAL OR
+    'ø': 'ø', # \[OSlash] (LATIN SMALL LETTER O WITH STROKE) -> LATIN SMALL LETTER O WITH STROKE
+    'õ': 'õ', # \[OTilde] (LATIN SMALL LETTER O WITH TILDE) -> LATIN SMALL LETTER O WITH TILDE
+    '︷': '︷', # \[OverBrace] (PRESENTATION FORM FOR VERTICAL LEFT CURLY BRACKET) -> PRESENTATION FORM FOR VERTICAL LEFT CURLY BRACKET
+    '⎴': '⎴', # \[OverBracket] (TOP SQUARE BRACKET) -> TOP SQUARE BRACKET
+    '︵': '︵', # \[OverParenthesis] (PRESENTATION FORM FOR VERTICAL LEFT PARENTHESIS) -> PRESENTATION FORM FOR VERTICAL LEFT PARENTHESIS
+    '¶': '¶', # \[Paragraph] (PILCROW SIGN) -> PILCROW SIGN
+    '∂': '∂', # \[PartialD] (PARTIAL DIFFERENTIAL) -> PARTIAL DIFFERENTIAL
+    '\uf3de': '⊙', # \[PermutationProduct] (INVALID CHARACTER) -> CIRCLED DOT OPERATOR
+    '⟂': '⟂', # \[Perpendicular] (PERPENDICULAR) -> PERPENDICULAR
+    'ϕ': 'ϕ', # \[Phi] (GREEK PHI SYMBOL) -> GREEK PHI SYMBOL
+    'π': 'π', # \[Pi] (GREEK SMALL LETTER PI) -> GREEK SMALL LETTER PI
+    '♓': '♓', # \[PiscesSign] (PISCES) -> PISCES
+    '\uf528': '⎕', # \[Placeholder] (INVALID CHARACTER) -> APL FUNCTIONAL SYMBOL QUAD
+    '±': '±', # \[PlusMinus] (PLUS-MINUS SIGN) -> PLUS-MINUS SIGN
+    '♇': '♇', # \[Pluto] (PLUTO) -> PLUTO
+    '≺': '≺', # \[Precedes] (PRECEDES) -> PRECEDES
+    '⪯': '⪯', # \[PrecedesEqual] (PRECEDES ABOVE SINGLE-LINE EQUALS SIGN) -> PRECEDES ABOVE SINGLE-LINE EQUALS SIGN
+    '≼': '≼', # \[PrecedesSlantEqual] (PRECEDES OR EQUAL TO) -> PRECEDES OR EQUAL TO
+    '≾': '≾', # \[PrecedesTilde] (PRECEDES OR EQUIVALENT TO) -> PRECEDES OR EQUIVALENT TO
+    '′': '′', # \[Prime] (PRIME) -> PRIME
+    '∏': '∏', # \[Product] (N-ARY PRODUCT) -> N-ARY PRODUCT
+    '∷': '∷', # \[Proportion] (PROPORTION) -> PROPORTION
+    '∝': '∝', # \[Proportional] (PROPORTIONAL TO) -> PROPORTIONAL TO
+    'ψ': 'ψ', # \[Psi] (GREEK SMALL LETTER PSI) -> GREEK SMALL LETTER PSI
+    '♩': '♩', # \[QuarterNote] (QUARTER NOTE) -> QUARTER NOTE
+    '&': '&', # \[RawAmpersand] (AMPERSAND) -> AMPERSAND
+    '@': '@', # \[RawAt] (COMMERCIAL AT) -> COMMERCIAL AT
+    '`': '`', # \[RawBackquote] (GRAVE ACCENT) -> GRAVE ACCENT
+    '\\': '\\', # \[RawBackslash] (REVERSE SOLIDUS) -> REVERSE SOLIDUS
+    ':': ':', # \[RawColon] (COLON) -> COLON
+    ',': ',', # \[RawComma] (COMMA) -> COMMA
+    '-': '-', # \[RawDash] (HYPHEN-MINUS) -> HYPHEN-MINUS
+    '$': '$', # \[RawDollar] (DOLLAR SIGN) -> DOLLAR SIGN
+    '.': '.', # \[RawDot] (FULL STOP) -> FULL STOP
+    '=': '=', # \[RawEqual] (EQUALS SIGN) -> EQUALS SIGN
+    '!': '!', # \[RawExclamation] (EXCLAMATION MARK) -> EXCLAMATION MARK
+    '>': '>', # \[RawGreater] (GREATER-THAN SIGN) -> GREATER-THAN SIGN
+    '{': '{', # \[RawLeftBrace] (LEFT CURLY BRACKET) -> LEFT CURLY BRACKET
+    '[': '[', # \[RawLeftBracket] (LEFT SQUARE BRACKET) -> LEFT SQUARE BRACKET
+    '(': '(', # \[RawLeftParenthesis] (LEFT PARENTHESIS) -> LEFT PARENTHESIS
+    '<': '<', # \[RawLess] (LESS-THAN SIGN) -> LESS-THAN SIGN
+    '#': '#', # \[RawNumberSign] (NUMBER SIGN) -> NUMBER SIGN
+    '%': '%', # \[RawPercent] (PERCENT SIGN) -> PERCENT SIGN
+    '+': '+', # \[RawPlus] (PLUS SIGN) -> PLUS SIGN
+    '?': '?', # \[RawQuestion] (QUESTION MARK) -> QUESTION MARK
+    "'": "'", # \[RawQuote] (APOSTROPHE) -> APOSTROPHE
+    '}': '}', # \[RawRightBrace] (RIGHT CURLY BRACKET) -> RIGHT CURLY BRACKET
+    ']': ']', # \[RawRightBracket] (RIGHT SQUARE BRACKET) -> RIGHT SQUARE BRACKET
+    ')': ')', # \[RawRightParenthesis] (RIGHT PARENTHESIS) -> RIGHT PARENTHESIS
+    '/': '/', # \[RawSlash] (SOLIDUS) -> SOLIDUS
+    ' ': ' ', # \[RawSpace] (SPACE) -> SPACE
+    '*': '*', # \[RawStar] (ASTERISK) -> ASTERISK
+    '~': '~', # \[RawTilde] (TILDE) -> TILDE
+    '_': '_', # \[RawUnderscore] (LOW LINE) -> LOW LINE
+    '|': '|', # \[RawVerticalBar] (VERTICAL LINE) -> VERTICAL LINE
+    '^': '^', # \[RawWedge] (CIRCUMFLEX ACCENT) -> CIRCUMFLEX ACCENT
+    '®': '®', # \[RegisteredTrademark] (REGISTERED SIGN) -> REGISTERED SIGN
+    '↵': '↵', # \[ReturnIndicator] (DOWNWARDS ARROW WITH CORNER LEFTWARDS) -> DOWNWARDS ARROW WITH CORNER LEFTWARDS
+    '‶': '‶', # \[ReverseDoublePrime] (REVERSED DOUBLE PRIME) -> REVERSED DOUBLE PRIME
+    '∋': '∋', # \[ReverseElement] (CONTAINS AS MEMBER) -> CONTAINS AS MEMBER
+    '⇋': '⇋', # \[ReverseEquilibrium] (LEFTWARDS HARPOON OVER RIGHTWARDS HARPOON) -> LEFTWARDS HARPOON OVER RIGHTWARDS HARPOON
+    '‵': '‵', # \[ReversePrime] (REVERSED PRIME) -> REVERSED PRIME
+    '⥯': '⥯', # \[ReverseUpEquilibrium] (DOWNWARDS HARPOON WITH BARB LEFT BESIDE UPWARDS HARPOON WITH BARB RIGHT) -> DOWNWARDS HARPOON WITH BARB LEFT BESIDE UPWARDS HARPOON WITH BARB RIGHT
+    'ř': 'ř', # \[RHacek] (LATIN SMALL LETTER R WITH CARON) -> LATIN SMALL LETTER R WITH CARON
+    'ρ': 'ρ', # \[Rho] (GREEK SMALL LETTER RHO) -> GREEK SMALL LETTER RHO
+    '∟': '∟', # \[RightAngle] (RIGHT ANGLE) -> RIGHT ANGLE
+    '〉': '〉', # \[RightAngleBracket] (RIGHT-POINTING ANGLE BRACKET) -> RIGHT-POINTING ANGLE BRACKET
+    '→': '→', # \[RightArrow] (RIGHTWARDS ARROW) -> RIGHTWARDS ARROW
+    '⇥': '⇥', # \[RightArrowBar] (RIGHTWARDS ARROW TO BAR) -> RIGHTWARDS ARROW TO BAR
+    '⇄': '⇄', # \[RightArrowLeftArrow] (RIGHTWARDS ARROW OVER LEFTWARDS ARROW) -> RIGHTWARDS ARROW OVER LEFTWARDS ARROW
+    '\uf604': '|', # \[RightBracketingBar] (INVALID CHARACTER) -> VERTICAL LINE
+    '⌉': '⌉', # \[RightCeiling] (RIGHT CEILING) -> RIGHT CEILING
+    '〛': '〛', # \[RightDoubleBracket] (RIGHT WHITE SQUARE BRACKET) -> RIGHT WHITE SQUARE BRACKET
+    '\uf606': '‖', # \[RightDoubleBracketingBar] (INVALID CHARACTER) -> DOUBLE VERTICAL LINE
+    '⥝': '⥝', # \[RightDownTeeVector] (DOWNWARDS HARPOON WITH BARB RIGHT FROM BAR) -> DOWNWARDS HARPOON WITH BARB RIGHT FROM BAR
+    '⇂': '⇂', # \[RightDownVector] (DOWNWARDS HARPOON WITH BARB RIGHTWARDS) -> DOWNWARDS HARPOON WITH BARB RIGHTWARDS
+    '⥕': '⥕', # \[RightDownVectorBar] (DOWNWARDS HARPOON WITH BARB RIGHT TO BAR) -> DOWNWARDS HARPOON WITH BARB RIGHT TO BAR
+    '⌋': '⌋', # \[RightFloor] (RIGHT FLOOR) -> RIGHT FLOOR
+    '»': '»', # \[RightGuillemet] (RIGHT-POINTING DOUBLE ANGLE QUOTATION MARK) -> RIGHT-POINTING DOUBLE ANGLE QUOTATION MARK
+    '▸': '▸', # \[RightPointer] (BLACK RIGHT-POINTING SMALL TRIANGLE) -> BLACK RIGHT-POINTING SMALL TRIANGLE
+    '\uf762': '»', # \[RightSkeleton] (INVALID CHARACTER) -> RIGHT-POINTING DOUBLE ANGLE QUOTATION MARK
+    '⊢': '⊢', # \[RightTee] (RIGHT TACK) -> RIGHT TACK
+    '↦': '↦', # \[RightTeeArrow] (RIGHTWARDS ARROW FROM BAR) -> RIGHTWARDS ARROW FROM BAR
+    '⥛': '⥛', # \[RightTeeVector] (RIGHTWARDS HARPOON WITH BARB UP FROM BAR) -> RIGHTWARDS HARPOON WITH BARB UP FROM BAR
+    '⊳': '⊳', # \[RightTriangle] (CONTAINS AS NORMAL SUBGROUP) -> CONTAINS AS NORMAL SUBGROUP
+    '⧐': '⧐', # \[RightTriangleBar] (VERTICAL BAR BESIDE RIGHT TRIANGLE) -> VERTICAL BAR BESIDE RIGHT TRIANGLE
+    '⊵': '⊵', # \[RightTriangleEqual] (CONTAINS AS NORMAL SUBGROUP OR EQUAL TO) -> CONTAINS AS NORMAL SUBGROUP OR EQUAL TO
+    '⥏': '⥏', # \[RightUpDownVector] (UP BARB RIGHT DOWN BARB RIGHT HARPOON) -> UP BARB RIGHT DOWN BARB RIGHT HARPOON
+    '⥜': '⥜', # \[RightUpTeeVector] (UPWARDS HARPOON WITH BARB RIGHT FROM BAR) -> UPWARDS HARPOON WITH BARB RIGHT FROM BAR
+    '↾': '↾', # \[RightUpVector] (UPWARDS HARPOON WITH BARB RIGHTWARDS) -> UPWARDS HARPOON WITH BARB RIGHTWARDS
+    '⥔': '⥔', # \[RightUpVectorBar] (UPWARDS HARPOON WITH BARB RIGHT TO BAR) -> UPWARDS HARPOON WITH BARB RIGHT TO BAR
+    '⇀': '⇀', # \[RightVector] (RIGHTWARDS HARPOON WITH BARB UPWARDS) -> RIGHTWARDS HARPOON WITH BARB UPWARDS
+    '⥓': '⥓', # \[RightVectorBar] (RIGHTWARDS HARPOON WITH BARB UP TO BAR) -> RIGHTWARDS HARPOON WITH BARB UP TO BAR
+    '⥰': '⥰', # \[RoundImplies] (RIGHT DOUBLE ARROW WITH ROUNDED HEAD) -> RIGHT DOUBLE ARROW WITH ROUNDED HEAD
+    '\uf522': '→', # \[Rule] (INVALID CHARACTER) -> RIGHTWARDS ARROW
+    '\uf51f': '⧴', # \[RuleDelayed] (INVALID CHARACTER) -> RULE-DELAYED
+    '☹': '☹', # \[SadSmiley] (WHITE FROWNING FACE) -> WHITE FROWNING FACE
+    '♐': '♐', # \[SagittariusSign] (SAGITTARIUS) -> SAGITTARIUS
+    # This one is changed because GREEK LETTER SAMPI is already used for \[CapitalSampi]
+    'Ϡ': 'ϡ', # \[Sampi] (GREEK LETTER SAMPI) -> GREEK SMALL LETTER SAMPI
+    '♄': '♄', # \[Saturn] (SATURN) -> SATURN
+    '♏': '♏', # \[ScorpioSign] (SCORPIUS) -> SCORPIUS
+    '\uf6b2': '𝒶', # \[ScriptA] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL A
+    '\uf6b3': '𝒷', # \[ScriptB] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL B
+    '\uf6b4': '𝒸', # \[ScriptC] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL C
+    '\uf770': '𝒜', # \[ScriptCapitalA] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL A
+    'ℬ': 'ℬ', # \[ScriptCapitalB] (SCRIPT CAPITAL B) -> SCRIPT CAPITAL B
+    '\uf772': '𝒞', # \[ScriptCapitalC] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL C
+    '\uf773': '𝒟', # \[ScriptCapitalD] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL D
+    'ℰ': 'ℰ', # \[ScriptCapitalE] (SCRIPT CAPITAL E) -> SCRIPT CAPITAL E
+    'ℱ': 'ℱ', # \[ScriptCapitalF] (SCRIPT CAPITAL F) -> SCRIPT CAPITAL F
+    '\uf776': '𝒢', # \[ScriptCapitalG] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL G
+    'ℋ': 'ℋ', # \[ScriptCapitalH] (SCRIPT CAPITAL H) -> SCRIPT CAPITAL H
+    'ℐ': 'ℐ', # \[ScriptCapitalI] (SCRIPT CAPITAL I) -> SCRIPT CAPITAL I
+    '\uf779': '𝒥', # \[ScriptCapitalJ] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL J
+    '\uf77a': '𝒦', # \[ScriptCapitalK] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL K
+    'ℒ': 'ℒ', # \[ScriptCapitalL] (SCRIPT CAPITAL L) -> SCRIPT CAPITAL L
+    'ℳ': 'ℳ', # \[ScriptCapitalM] (SCRIPT CAPITAL M) -> SCRIPT CAPITAL M
+    '\uf77d': '𝒩', # \[ScriptCapitalN] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL N
+    '\uf77e': '𝒪', # \[ScriptCapitalO] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL O
+    '℘': '𝒫', # \[ScriptCapitalP] (SCRIPT CAPITAL P) -> MATHEMATICAL SCRIPT CAPITAL P
+    '\uf780': '𝒬', # \[ScriptCapitalQ] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL Q
+    'ℛ': 'ℛ', # \[ScriptCapitalR] (SCRIPT CAPITAL R) -> SCRIPT CAPITAL R
+    '\uf782': '𝒮', # \[ScriptCapitalS] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL S
+    '\uf783': '𝒯', # \[ScriptCapitalT] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL T
+    '\uf784': '𝒰', # \[ScriptCapitalU] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL U
+    '\uf785': '𝒱', # \[ScriptCapitalV] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL V
+    '\uf786': '𝒲', # \[ScriptCapitalW] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL W
+    '\uf787': '𝒳', # \[ScriptCapitalX] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL X
+    '\uf788': '𝒴', # \[ScriptCapitalY] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL Y
+    '\uf789': '𝒵', # \[ScriptCapitalZ] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT CAPITAL Z
+    '\uf6b5': '𝒹', # \[ScriptD] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL D
+    '\uf730': '𝒾', # \[ScriptDotlessI] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL I
+    '\uf731': '𝒿', # \[ScriptDotlessJ] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL J
+    'ℯ': 'ℯ', # \[ScriptE] (SCRIPT SMALL E) -> SCRIPT SMALL E
+    '\uf6b7': '𝒻', # \[ScriptF] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL F
+    'ℊ': 'ℊ', # \[ScriptG] (SCRIPT SMALL G) -> SCRIPT SMALL G
+    '\uf6b9': '𝒽', # \[ScriptH] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL H
+    '\uf6ba': '𝒾', # \[ScriptI] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL I
+    '\uf6bb': '𝒿', # \[ScriptJ] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL J
+    '\uf6bc': '𝓀', # \[ScriptK] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL K
+    'ℓ': 'ℓ', # \[ScriptL] (SCRIPT SMALL L) -> SCRIPT SMALL L
+    '\uf6be': '𝓂', # \[ScriptM] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL M
+    '\uf6bf': '𝓃', # \[ScriptN] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL N
+    'ℴ': 'ℴ', # \[ScriptO] (SCRIPT SMALL O) -> SCRIPT SMALL O
+    '\uf6c1': '𝓅', # \[ScriptP] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL P
+    '\uf6c2': '𝓆', # \[ScriptQ] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL Q
+    '\uf6c3': '𝓇', # \[ScriptR] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL R
+    '\uf6c4': '𝓈', # \[ScriptS] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL S
+    '\uf6c5': '𝓉', # \[ScriptT] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL T
+    '\uf6c6': '𝓊', # \[ScriptU] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL U
+    '\uf6c7': '𝓋', # \[ScriptV] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL V
+    '\uf6c8': '𝓌', # \[ScriptW] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL W
+    '\uf6c9': '𝓍', # \[ScriptX] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL X
+    '\uf6ca': '𝓎', # \[ScriptY] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL Y
+    '\uf6cb': '𝓏', # \[ScriptZ] (INVALID CHARACTER) -> MATHEMATICAL SCRIPT SMALL Z
+    '§': '§', # \[Section] (SECTION SIGN) -> SECTION SIGN
+    'š': 'š', # \[SHacek] (LATIN SMALL LETTER S WITH CARON) -> LATIN SMALL LETTER S WITH CARON
+    '♯': '♯', # \[Sharp] (MUSIC SHARP SIGN) -> MUSIC SHARP SIGN
+    '\uf52b': '↓', # \[ShortDownArrow] (INVALID CHARACTER) -> DOWNWARDS ARROW
+    '\uf526': '←', # \[ShortLeftArrow] (INVALID CHARACTER) -> LEFTWARDS ARROW
+    '\uf525': '→', # \[ShortRightArrow] (INVALID CHARACTER) -> RIGHTWARDS ARROW
+    '\uf52a': '↑', # \[ShortUpArrow] (INVALID CHARACTER) -> UPWARDS ARROW
+    'σ': 'σ', # \[Sigma] (GREEK SMALL LETTER SIGMA) -> GREEK SMALL LETTER SIGMA
+    '✶': '✶', # \[SixPointedStar] (SIX POINTED BLACK STAR) -> SIX POINTED BLACK STAR
+    '⁃': '⁃', # \[SkeletonIndicator] (HYPHEN BULLET) -> HYPHEN BULLET
+    '∘': '∘', # \[SmallCircle] (RING OPERATOR) -> RING OPERATOR
+    '␣': '␣', # \[SpaceIndicator] (OPEN BOX) -> OPEN BOX
+    '♠': '♠', # \[SpadeSuit] (BLACK SPADE SUIT) -> BLACK SPADE SUIT
+    '\uf3bb': '⋮', # \[SpanFromAbove] (INVALID CHARACTER) -> VERTICAL ELLIPSIS
+    '\uf3bc': '⋱', # \[SpanFromBoth] (INVALID CHARACTER) -> DOWN RIGHT DIAGONAL ELLIPSIS
+    '\uf3ba': '⋯', # \[SpanFromLeft] (INVALID CHARACTER) -> MIDLINE HORIZONTAL ELLIPSIS
+    '∢': '∢', # \[SphericalAngle] (SPHERICAL ANGLE) -> SPHERICAL ANGLE
+    '√': '√', # \[Sqrt] (SQUARE ROOT) -> SQUARE ROOT
+    '\uf520': '▫', # \[Square] (INVALID CHARACTER) -> WHITE SMALL SQUARE
+    '⊓': '⊓', # \[SquareIntersection] (SQUARE CAP) -> SQUARE CAP
+    '⊏': '⊏', # \[SquareSubset] (SQUARE IMAGE OF) -> SQUARE IMAGE OF
+    '⊑': '⊑', # \[SquareSubsetEqual] (SQUARE IMAGE OF OR EQUAL TO) -> SQUARE IMAGE OF OR EQUAL TO
+    '⊐': '⊐', # \[SquareSuperset] (SQUARE ORIGINAL OF) -> SQUARE ORIGINAL OF
+    '⊒': '⊒', # \[SquareSupersetEqual] (SQUARE ORIGINAL OF OR EQUAL TO) -> SQUARE ORIGINAL OF OR EQUAL TO
+    '⊔': '⊔', # \[SquareUnion] (SQUARE CUP) -> SQUARE CUP
+    '⋆': '⋆', # \[Star] (STAR OPERATOR) -> STAR OPERATOR
+    '£': '£', # \[Sterling] (POUND SIGN) -> POUND SIGN
+    'ϛ': 'ϛ', # \[Stigma] (GREEK SMALL LETTER STIGMA) -> GREEK SMALL LETTER STIGMA
+    '⊂': '⊂', # \[Subset] (SUBSET OF) -> SUBSET OF
+    '⊆': '⊆', # \[SubsetEqual] (SUBSET OF OR EQUAL TO) -> SUBSET OF OR EQUAL TO
+    '≻': '≻', # \[Succeeds] (SUCCEEDS) -> SUCCEEDS
+    '⪰': '⪰', # \[SucceedsEqual] (SUCCEEDS ABOVE SINGLE-LINE EQUALS SIGN) -> SUCCEEDS ABOVE SINGLE-LINE EQUALS SIGN
+    '≽': '≽', # \[SucceedsSlantEqual] (SUCCEEDS OR EQUAL TO) -> SUCCEEDS OR EQUAL TO
+    '≿': '≿', # \[SucceedsTilde] (SUCCEEDS OR EQUIVALENT TO) -> SUCCEEDS OR EQUIVALENT TO
+    '∍': '∍', # \[SuchThat] (SMALL CONTAINS AS MEMBER) -> SMALL CONTAINS AS MEMBER
+    '∑': '∑', # \[Sum] (N-ARY SUMMATION) -> N-ARY SUMMATION
+    '⊃': '⊃', # \[Superset] (SUPERSET OF) -> SUPERSET OF
+    '⊇': '⊇', # \[SupersetEqual] (SUPERSET OF OR EQUAL TO) -> SUPERSET OF OR EQUAL TO
+    'ß': 'ß', # \[SZ] (LATIN SMALL LETTER SHARP S) -> LATIN SMALL LETTER SHARP S
+    'τ': 'τ', # \[Tau] (GREEK SMALL LETTER TAU) -> GREEK SMALL LETTER TAU
+    '♉': '♉', # \[TaurusSign] (TAURUS) -> TAURUS
+    '\uf3da': '⊗', # \[TensorProduct] (INVALID CHARACTER) -> CIRCLED TIMES
+    'ť': 'ť', # \[THacek] (LATIN SMALL LETTER T WITH CARON) -> LATIN SMALL LETTER T WITH CARON
+    '∴': '∴', # \[Therefore] (THEREFORE) -> THEREFORE
+    'θ': 'θ', # \[Theta] (GREEK SMALL LETTER THETA) -> GREEK SMALL LETTER THETA
+    '\u2005': '\u2005', # \[ThickSpace] (FOUR-PER-EM SPACE) -> FOUR-PER-EM SPACE
+    'þ': 'þ', # \[Thorn] (LATIN SMALL LETTER THORN) -> LATIN SMALL LETTER THORN
+    '∼': '∼', # \[Tilde] (TILDE OPERATOR) -> TILDE OPERATOR
+    '≃': '≃', # \[TildeEqual] (ASYMPTOTICALLY EQUAL TO) -> ASYMPTOTICALLY EQUAL TO
+    '≅': '≅', # \[TildeFullEqual] (APPROXIMATELY EQUAL TO) -> APPROXIMATELY EQUAL TO
+    '≈': '≈', # \[TildeTilde] (ALMOST EQUAL TO) -> ALMOST EQUAL TO
+    '×': '×', # \[Times] (MULTIPLICATION SIGN) -> MULTIPLICATION SIGN
+    '™': '™', # \[Trademark] (TRADE MARK SIGN) -> TRADE MARK SIGN
+    '\uf3c7': 'ᵀ', # \[Transpose] (INVALID CHARACTER) -> MODIFIER LETTER CAPITAL T
+    '\uf758': '⋯', # \[TripleDot] (INVALID CHARACTER) -> MIDLINE HORIZONTAL ELLIPSIS
+    'ú': 'ú', # \[UAcute] (LATIN SMALL LETTER U WITH ACUTE) -> LATIN SMALL LETTER U WITH ACUTE
+    'ű': 'ű', # \[UDoubleAcute] (LATIN SMALL LETTER U WITH DOUBLE ACUTE) -> LATIN SMALL LETTER U WITH DOUBLE ACUTE
+    'ü': 'ü', # \[UDoubleDot] (LATIN SMALL LETTER U WITH DIAERESIS) -> LATIN SMALL LETTER U WITH DIAERESIS
+    'ù': 'ù', # \[UGrave] (LATIN SMALL LETTER U WITH GRAVE) -> LATIN SMALL LETTER U WITH GRAVE
+    'û': 'û', # \[UHat] (LATIN SMALL LETTER U WITH CIRCUMFLEX) -> LATIN SMALL LETTER U WITH CIRCUMFLEX
+    '︸': '︸', # \[UnderBrace] (PRESENTATION FORM FOR VERTICAL RIGHT CURLY BRACKET) -> PRESENTATION FORM FOR VERTICAL RIGHT CURLY BRACKET
+    '⎵': '⎵', # \[UnderBracket] (BOTTOM SQUARE BRACKET) -> BOTTOM SQUARE BRACKET
+    '︶': '︶', # \[UnderParenthesis] (PRESENTATION FORM FOR VERTICAL RIGHT PARENTHESIS) -> PRESENTATION FORM FOR VERTICAL RIGHT PARENTHESIS
+    '\uf3d4': '↔', # \[UndirectedEdge] (INVALID CHARACTER) -> LEFT RIGHT ARROW
+    '⋃': '⋃', # \[Union] (N-ARY UNION) -> N-ARY UNION
+    '⊎': '⊎', # \[UnionPlus] (MULTISET UNION) -> MULTISET UNION
+    '↑': '↑', # \[UpArrow] (UPWARDS ARROW) -> UPWARDS ARROW
+    '⤒': '⤒', # \[UpArrowBar] (UPWARDS ARROW TO BAR) -> UPWARDS ARROW TO BAR
+    '⇅': '⇅', # \[UpArrowDownArrow] (UPWARDS ARROW LEFTWARDS OF DOWNWARDS ARROW) -> UPWARDS ARROW LEFTWARDS OF DOWNWARDS ARROW
+    '↕': '↕', # \[UpDownArrow] (UP DOWN ARROW) -> UP DOWN ARROW
+    '⥮': '⥮', # \[UpEquilibrium] (UPWARDS HARPOON WITH BARB LEFT BESIDE DOWNWARDS HARPOON WITH BARB RIGHT) -> UPWARDS HARPOON WITH BARB LEFT BESIDE DOWNWARDS HARPOON WITH BARB RIGHT
+    '↖': '↖', # \[UpperLeftArrow] (NORTH WEST ARROW) -> NORTH WEST ARROW
+    '↗': '↗', # \[UpperRightArrow] (NORTH EAST ARROW) -> NORTH EAST ARROW
+    '▴': '▴', # \[UpPointer] (BLACK UP-POINTING SMALL TRIANGLE) -> BLACK UP-POINTING SMALL TRIANGLE
+    'υ': 'υ', # \[Upsilon] (GREEK SMALL LETTER UPSILON) -> GREEK SMALL LETTER UPSILON
+    '⊥': '⊥', # \[UpTee] (UP TACK) -> UP TACK
+    '↥': '↥', # \[UpTeeArrow] (UPWARDS ARROW FROM BAR) -> UPWARDS ARROW FROM BAR
+    '♅': '⛢', # \[Uranus] (URANUS) -> ASTRONOMICAL SYMBOL FOR URANUS
+    'ů': 'ů', # \[URing] (LATIN SMALL LETTER U WITH RING ABOVE) -> LATIN SMALL LETTER U WITH RING ABOVE
+    '⋁': '⋁', # \[Vee] (N-ARY LOGICAL OR) -> N-ARY LOGICAL OR
+    '♀': '♀', # \[Venus] (FEMALE SIGN) -> FEMALE SIGN
+    '\uf3d0': '|', # \[VerticalBar] (INVALID CHARACTER) -> VERTICAL LINE
+    '⋮': '⋮', # \[VerticalEllipsis] (VERTICAL ELLIPSIS) -> VERTICAL ELLIPSIS
+    '│': '│', # \[VerticalLine] (BOX DRAWINGS LIGHT VERTICAL) -> BOX DRAWINGS LIGHT VERTICAL
+    '≀': '≀', # \[VerticalTilde] (WREATH PRODUCT) -> WREATH PRODUCT
+    '\u200a': '\u200a', # \[VeryThinSpace] (HAIR SPACE) -> HAIR SPACE
+    '♍': '♍', # \[VirgoSign] (VIRGO) -> VIRGO
+    '\uf725': '\uf725', # \[WarningSign] (INVALID CHARACTER) -> 
+    '⌚': '⌚', # \[WatchIcon] (WATCH) -> WATCH
+    '⋀': '⋀', # \[Wedge] (N-ARY LOGICAL AND) -> N-ARY LOGICAL AND
+    '℘': '℘', # \[WeierstrassP] (SCRIPT CAPITAL P) -> SCRIPT CAPITAL P
+    '♗': '♗', # \[WhiteBishop] (WHITE CHESS BISHOP) -> WHITE CHESS BISHOP
+    '♔': '♔', # \[WhiteKing] (WHITE CHESS KING) -> WHITE CHESS KING
+    '♘': '♘', # \[WhiteKnight] (WHITE CHESS KNIGHT) -> WHITE CHESS KNIGHT
+    '♙': '♙', # \[WhitePawn] (WHITE CHESS PAWN) -> WHITE CHESS PAWN
+    '♕': '♕', # \[WhiteQueen] (WHITE CHESS QUEEN) -> WHITE CHESS QUEEN
+    '♖': '♖', # \[WhiteRook] (WHITE CHESS ROOK) -> WHITE CHESS ROOK
+    'ξ': 'ξ', # \[Xi] (GREEK SMALL LETTER XI) -> GREEK SMALL LETTER XI
+    '⊻': '⊻', # \[Xor] (XOR) -> XOR
+    'ý': 'ý', # \[YAcute] (LATIN SMALL LETTER Y WITH ACUTE) -> LATIN SMALL LETTER Y WITH ACUTE
+    'ÿ': 'ÿ', # \[YDoubleDot] (LATIN SMALL LETTER Y WITH DIAERESIS) -> LATIN SMALL LETTER Y WITH DIAERESIS
+    '¥': '¥', # \[Yen] (YEN SIGN) -> YEN SIGN
+    'ζ': 'ζ', # \[Zeta] (GREEK SMALL LETTER ZETA) -> GREEK SMALL LETTER ZETA
+    'ž': 'ž', # \[ZHacek] (LATIN SMALL LETTER Z WITH CARON) -> LATIN SMALL LETTER Z WITH CARON
 }
 
-aliased_characters = {
-    "a'": "\u00E1",
-    "a-": "\u0101",
-    "au": "\u0103",
-    'a"': "\u00E4",
-    "ae": "\u00E6",
-    "a`": "\u00E0",
-    "a^": "\u00E2",
-    "al": "\u2135",
-    "esc": "\uF768",
-    "am": "\uF760",
-    "a": "\u03B1",
-    "alpha": "\u03B1",
-    "alt": "\uF7D1",
-    "&&": "\u2227",
-    "and": "\u2227",
-    "Ang": "\u212B",
-    "ao": "\u00E5",
-    "a~": "\u00E3",
-    "\\": "\u2216",
-    "be": "\u2136",
-    "b": "\u03B2",
-    "beta": "\u03B2",
-    "bv": "\u02D8",
-    "bu": "\u2022",
-    "c'": "\u0107",
-    "A'": "\u00C1",
-    "A-": "\u0100",
-    "Au": "\u0102",
-    'A"': "\u00C4",
-    "AE": "\u00C6",
-    "A`": "\u00C0",
-    "A^": "\u00C2",
-    "A": "\u0391",
-    "Alpha": "\u0391",
-    "Ao": "\u00C5",
-    "A~": "\u00C3",
-    "B": "\u0392",
-    "Beta": "\u0392",
-    "C'": "\u0106",
-    "C,": "\u00C7",
-    "Cv": "\u010C",
-    "Ch": "\u03A7",
-    "Chi": "\u03A7",
-    "C": "\u03A7",
-    "D": "\u0394",
-    "Delta": "\u0394",
-    "Dv": "\u010E",
-    "DD": "\uF74B",
-    "Di": "\u03DC",
-    "Digamma": "\u03DC",
-    "E'": "\u00C9",
-    "E-": "\u0112",
-    "Eu": "\u0114",
-    'E"': "\u00CB",
-    "E`": "\u00C8",
-    "Ev": "\u011A",
-    "E^": "\u00CA",
-    "E": "\u0395",
-    "Epsilon": "\u0395",
-    "Et": "\u0397",
-    "Eta": "\u0397",
-    "H": "\u0397",
-    "D-": "\u00D0",
-    "G": "\u0393",
-    "Gamma": "\u0393",
-    "I'": "\u00CD",
-    "Iu": "\u012C",
-    'I"': "\u00CF",
-    "I`": "\u00CC",
-    "I^": "\u00CE",
-    "I": "\u0399",
-    "Iota": "\u0399",
-    "K": "\u039A",
-    "Kappa": "\u039A",
-    "Ko": "\u03DE",
-    "Koppa": "\u03DE",
-    "L": "\u039B",
-    "Lambda": "\u039B",
-    "L/": "\u0141",
-    "M": "\u039C",
-    "Mu": "\u039C",
-    "Nv": "\u0147",
-    "N~": "\u00D1",
-    "N": "\u039D",
-    "Nu": "\u039D",
-    "O'": "\u00D3",
-    "O''": "\u0150",
-    'O"': "\u00D6",
-    "OE": "\u0152",
-    "O`": "\u00D2",
-    "O^": "\u00D4",
-    "O": "\u03A9",
-    "Omega": "\u03A9",
-    "W": "\u03A9",
-    "Om": "\u039F",
-    "Omicron": "\u039F",
-    "O/": "\u00D8",
-    "O~": "\u00D5",
-    "Ph": "\u03A6",
-    "Phi": "\u03A6",
-    "F": "\u03A6",
-    "P": "\u03A0",
-    "Pi": "\u03A0",
-    "Ps": "\u03A8",
-    "Psi": "\u03A8",
-    "Y": "\u03A8",
-    "Rv": "\u0158",
-    "R": "\u03A1",
-    "Rho": "\u03A1",
-    "Sa": "\u03E0",
-    "Sampi": "\u03E0",
-    "Sv": "\u0160",
-    "S": "\u03A3",
-    "Sigma": "\u03A3",
-    "T": "\u03A4",
-    "Tau": "\u03A4",
-    "Tv": "\u0164",
-    "Th": "\u0398",
-    "Theta": "\u0398",
-    "Q": "\u0398",
-    "Thn": "\u00DE",
-    "U'": "\u00DA",
-    "U''": "\u0170",
-    'U"': "\u00DC",
-    "U`": "\u00D9",
-    "U^": "\u00DB",
-    "U": "\u03A5",
-    "Upsilon": "\u03A5",
-    "Uo": "\u016E",
-    "X": "\u039E",
-    "Xi": "\u039E",
-    "Y'": "\u00DD",
-    "Z": "\u0396",
-    "Zeta": "\u0396",
-    "Zv": "\u017D",
-    "c,": "\u00E7",
-    "cd": "\u00B8",
-    ".": "\u00B7",
-    "cent": "\u00A2",
-    "cv": "\u010D",
-    "ch": "\u03C7",
-    "chi": "\u03C7",
-    "c": "\u03C7",
-    "c.": "\u2299",
-    "c-": "\u2296",
-    "c+": "\u2295",
-    "c*": "\u2297",
-    "ccint": "\u2232",
-    "cl": "\u2318",
-    ":": "\u2236",
-    "cmd": "\uF76A",
-    "===": "\u2261",
-    "co": "\uF3C8",
-    "conj": "\uF3C8",
-    "ct": "\uF3C9",
-    "cont": "\uF3B1",
-    "cint": "\u222E",
-    "ctrl": "\uF763",
-    "coprod": "\u2210",
-    "cccint": "\u2233",
-    "cross": "\uF4A0",
-    "cU": "\u03D2",
-    "cUpsilon": "\u03D2",
-    "ce": "\u03B5",
-    "cepsilon": "\u03B5",
-    "ck": "\u03F0",
-    "ckappa": "\u03F0",
-    "j": "\u03C6",
-    "cph": "\u03C6",
-    "cphi": "\u03C6",
-    "cp": "\u03D6",
-    "cpi": "\u03D6",
-    "cr": "\u03F1",
-    "crho": "\u03F1",
-    "cq": "\u03D1",
-    "cth": "\u03D1",
-    "ctheta": "\u03D1",
-    "dg": "\u2020",
-    "da": "\u2138",
-    "-": "\u2013",
-    "deg": "\u00B0",
-    " del": "\uF7D0",
-    "del": "\u2207",
-    "d": "\u03B4",
-    "delta": "\u03B4",
-    "dv": "\u010F",
-    "dia": "\u22C4",
-    "diffd": "\u2206",
-    "dd": "\uF74C",
-    "di": "\u03DD",
-    "digamma": "\u03DD",
-    "dratio": "\uF4A4",
-    "shift": "\uF4A3",
-    "dhy": "\u00AD",
-    "dlsep": "\uF76E",
-    "dpsep": "\uF76F",
-    "div": "\u00F7",
-    ".=": "\u2250",
-    "ddg": "\u2021",
-    "gg": "\uF74A",
-    "pp": "\uF749",
-    " <=": "\u21D0",
-    "<=>": "\u21D4",
-    "<==": "\u27F8",
-    "<==>": "\u27FA",
-    "==>": "\u27F9",
-    "''": "\u2033",
-    " =>": "\u21D2",
-    "dsa": "\uF6E6",
-    "dsb": "\uF6E7",
-    "dsc": "\uF6E8",
-    "dsA": "\uF7A4",
-    "dsB": "\uF7A5",
-    "dsC": "\uF7A6",
-    "dsD": "\uF7A7",
-    "dsE": "\uF7A8",
-    "dsF": "\uF7A9",
-    "dsG": "\uF7AA",
-    "dsH": "\uF7AB",
-    "dsI": "\uF7AC",
-    "dsJ": "\uF7AD",
-    "dsK": "\uF7AE",
-    "dsL": "\uF7AF",
-    "dsM": "\uF7B0",
-    "dsN": "\uF7B1",
-    "dsO": "\uF7B2",
-    "dsP": "\uF7B3",
-    "dsQ": "\uF7B4",
-    "dsR": "\uF7B5",
-    "dsS": "\uF7B6",
-    "dsT": "\uF7B7",
-    "dsU": "\uF7B8",
-    "dsV": "\uF7B9",
-    "dsW": "\uF7BA",
-    "dsX": "\uF7BB",
-    "dsY": "\uF7BC",
-    "dsZ": "\uF7BD",
-    "dsd": "\uF6E9",
-    "dse": "\uF6EA",
-    "ds8": "\uF7E3",
-    "dsf": "\uF6EB",
-    "ds5": "\uF7E0",
-    "ds4": "\uF7DF",
-    "dsg": "\uF6EC",
-    "dsh": "\uF6ED",
-    "dsi": "\uF6EE",
-    "dsj": "\uF6EF",
-    "dsk": "\uF6F0",
-    "dsl": "\uF6F1",
-    "dsm": "\uF6F2",
-    "dsn": "\uF6F3",
-    "ds9": "\uF7E4",
-    "dso": "\uF6F4",
-    "ds1": "\uF7DC",
-    "dsp": "\uF6F5",
-    "dsq": "\uF6F6",
-    "dsr": "\uF6F7",
-    "dss": "\uF6F8",
-    "ds7": "\uF7E2",
-    "ds6": "\uF7E1",
-    "dst": "\uF6F9",
-    "ds3": "\uF7DE",
-    "ds2": "\uF7DD",
-    "dsu": "\uF6FA",
-    "dsv": "\uF6FB",
-    "dsw": "\uF6FC",
-    "dsx": "\uF6FD",
-    "dsy": "\uF6FE",
-    "dsz": "\uF6FF",
-    "ds0": "\uF7DB",
-    " ||": "\u2225",
-    "dbv": "\uF755",
-    "d!": "\u00A1",
-    "d?": "\u00BF",
-    "dT": "\u22A4",
-    "e'": "\u00E9",
-    "e-": "\u0113",
-    "eu": "\u0115",
-    'e"': "\u00EB",
-    "e`": "\u00E8",
-    "ev": "\u011B",
-    "e^": "\u00EA",
-    "el": "\u2208",
-    "elem": "\u2208",
-    "...": "\u2026",
-    "eci": "\u25CB",
-    "es": "\u2205",
-    "esci": "\u25E6",
-    "essq": "\u25FB",
-    "esq": "\u25A1",
-    "ent": "\uF7D4",
-    "e": "\u03F5",
-    "epsilon": "\u03F5",
-    "==": "\uF431",
-    "=~": "\u2242",
-    "equi": "\u21CC",
-    "equiv": "\u29E6",
-    " esc": "\uF769",
-    "et": "\u03B7",
-    "eta": "\u03B7",
-    "h": "\u03B7",
-    "d-": "\u00F0",
-    "ex": "\u2203",
-    "ee": "\uF74D",
-    "fci": "\u25CF",
-    "fsci": "\uF750",
-    "fssq": "\u25FC",
-    "fsq": "\u25A0",
-    "fvssq": "\u25AA",
-    "fs": "\u03C2",
-    "*5": "\u2605",
-    "fa": "\u2200",
-    "$a": "\uF800",
-    "$b": "\uF801",
-    "$c": "\uF802",
-    "$A": "\uF81A",
-    "$B": "\uF81B",
-    "$C": "\uF81C",
-    "$D": "\uF81D",
-    "$E": "\uF81E",
-    "$F": "\uF81F",
-    "$G": "\uF820",
-    "$H": "\uF821",
-    "$I": "\uF822",
-    "$J": "\uF823",
-    "$K": "\uF824",
-    "$L": "\uF825",
-    "$M": "\uF826",
-    "$N": "\uF827",
-    "$O": "\uF828",
-    "$P": "\uF829",
-    "$Q": "\uF82A",
-    "$R": "\uF82B",
-    "$S": "\uF82C",
-    "$T": "\uF82D",
-    "$U": "\uF82E",
-    "$V": "\uF82F",
-    "$W": "\uF830",
-    "$X": "\uF831",
-    "$Y": "\uF832",
-    "$Z": "\uF833",
-    "$d": "\uF803",
-    "$e": "\uF804",
-    "$f": "\uF805",
-    "$g": "\uF806",
-    "$h": "\uF807",
-    "$i": "\uF808",
-    "$j": "\uF809",
-    "$k": "\uF80A",
-    "$l": "\uF80B",
-    "$m": "\uF80C",
-    "$n": "\uF80D",
-    "$o": "\uF80E",
-    "$p": "\uF80F",
-    "$q": "\uF810",
-    "$r": "\uF811",
-    "$s": "\uF812",
-    "$t": "\uF813",
-    "$u": "\uF814",
-    "$v": "\uF815",
-    "$w": "\uF816",
-    "$x": "\uF817",
-    "$y": "\uF818",
-    "$z": "\uF819",
-    ":-@": "\uF721",
-    "fn": "\uF4A1",
-    "g": "\u03B3",
-    "gamma": "\u03B3",
-    "gi": "\u2137",
-    "goa": "\uF6CC",
-    "gob": "\uF6CD",
-    "goc": "\uF6CE",
-    "goA": "\uF78A",
-    "goB": "\uF78B",
-    "goC": "\u212D",
-    "goD": "\uF78D",
-    "goE": "\uF78E",
-    "goF": "\uF78F",
-    "goG": "\uF790",
-    "goH": "\u210C",
-    "goI": "\u2111",
-    "goJ": "\uF793",
-    "goK": "\uF794",
-    "goL": "\uF795",
-    "goM": "\uF796",
-    "goN": "\uF797",
-    "goO": "\uF798",
-    "goP": "\uF799",
-    "goQ": "\uF79A",
-    "goR": "\u211C",
-    "goS": "\uF79C",
-    "goT": "\uF79D",
-    "goU": "\uF79E",
-    "goV": "\uF79F",
-    "goW": "\uF7A0",
-    "goX": "\uF7A1",
-    "goY": "\uF7A2",
-    "goZ": "\u2128",
-    "god": "\uF6CF",
-    "goe": "\uF6D0",
-    "go8": "\uF7ED",
-    "gof": "\uF6D1",
-    "go5": "\uF7EA",
-    "go4": "\uF7E9",
-    "gog": "\uF6D2",
-    "goh": "\uF6D3",
-    "goi": "\uF6D4",
-    "goj": "\uF6D5",
-    "gok": "\uF6D6",
-    "gol": "\uF6D7",
-    "gom": "\uF6D8",
-    "gon": "\uF6D9",
-    "go9": "\uF7EF",
-    "goo": "\uF6DA",
-    "go1": "\uF7E6",
-    "gop": "\uF6DB",
-    "goq": "\uF6DC",
-    "gor": "\uF6DD",
-    "gos": "\uF6DE",
-    "go7": "\uF7EC",
-    "go6": "\uF7EB",
-    "got": "\uF6DF",
-    "go3": "\uF7E8",
-    "go2": "\uF7E7",
-    "gou": "\uF6E0",
-    "gov": "\uF6E1",
-    "gow": "\uF6E2",
-    "gox": "\uF6E3",
-    "goy": "\uF6E4",
-    "goz": "\uF6E5",
-    "go0": "\uF7E5",
-    "gci": "\uF753",
-    "gsq": "\uF752",
-    ">=": "\u2265",
-    ">/": "\u2A7E",
-    ">~": "\u2273",
-    "hck": "\u02C7",
-    ":)": "\u263A",
-    ":-)": "\u263A",
-    "hb": "\u210F",
-    "hc": "\uF3CE",
-    "hline": "\u2500",
-    "h=": "\u224F",
-    "i'": "\u00ED",
-    "iu": "\u012D",
-    'i"': "\u00EF",
-    "i`": "\u00EC",
-    "i^": "\u00EE",
-    "ii": "\uF74E",
-    "jj": "\uF74F",
-    "+": "\uF39E",
-    "=>": "\uF523",
-    "inf": "\u221E",
-    "int": "\u222B",
-    "inter": "\u22C2",
-    "@": "\uF76D",
-    ",": "\uF765",
-    "is": "\uF360",
-    "i": "\u03B9",
-    "iota": "\u03B9",
-    "k": "\u03BA",
-    "kappa": "\u03BA",
-    "ko": "\u03DF",
-    "koppa": "\u03DF",
-    "l": "\u03BB",
-    "lambda": "\u03BB",
-    "<": "\u2329",
-    "<-": "\u2190",
-    "l|": "\uF603",
-    "lc": "\u2308",
-    "[[": "\u301A",
-    "l||": "\uF605",
-    "lf": "\u230A",
-    "g<<": "\u00AB",
-    "[": "\uF76B",
-    "<->": "\u2194",
-    "lT": "\u22A3",
-    "<=": "\u2264",
-    "</": "\u2A7D",
-    "<~": "\u2272",
-    "_": "\uF754",
-    "ls": "\uF754",
-    "--": "\u2014",
-    "<--": "\u27F5",
-    "<-->": "\u27F7",
-    "-->": "\u27F6",
-    "l/": "\u0142",
-    "math": "\uF757",
-    "   ": "\u205F",
-    "mho": "\u2127",
-    "mi": "\u00B5",
-    "-+": "\u2213",
-    "m": "\u03BC",
-    "mu": "\u03BC",
-    "nand": "\u22BC",
-    "-   ": "\uF383",
-    "-    ": "\uF384",
-    "-  ": "\uF382",
-    "- ": "\uF380",
-    ":-|": "\uF722",
-    "nv": "\u0148",
-    "nb": "\u2060",
-    "nbs": "\u00A0",
-    "nor": "\u22BD",
-    "!===": "\u2262",
-    "!||": "\u2226",
-    "!el": "\u2209",
-    "!elem": "\u2209",
-    "!=": "\u2260",
-    "!=~": "\uF400",
-    "!ex": "\u2204",
-    "!>": "\u226F",
-    "!>=": "\u2271",
-    "!>/": "\uF429",
-    "!>~": "\u2275",
-    "!h=": "\uF401",
-    "!<=": "\u2270",
-    "!<": "\u226E",
-    "!</": "\uF424",
-    "!<~": "\u2274",
-    "!": "\u00AC",
-    "not": "\u00AC",
-    "!mem": "\u220C",
-    "!sub": "\u2284",
-    "!sub=": "\u2288",
-    "!sup": "\u2285",
-    "!sup=": "\u2289",
-    "!~": "\u2241",
-    "!~=": "\u2244",
-    "!~==": "\u2247",
-    "!~~": "\u2249",
-    "!|": "\u2224",
-    "n~": "\u00F1",
-    "n": "\u03BD",
-    "nu": "\u03BD",
-    "null": "\uF3A0",
-    "o'": "\u00F3",
-    "o''": "\u0151",
-    'o"': "\u00F6",
-    "oe": "\u0153",
-    "o`": "\u00F2",
-    "o^": "\u00F4",
-    "o": "\u03C9",
-    "omega": "\u03C9",
-    "w": "\u03C9",
-    "om": "\u03BF",
-    "omicron": "\u03BF",
-    "opt": "\uF7D2",
-    "||": "\u2228",
-    "or": "\u2228",
-    "o/": "\u00F8",
-    "o~": "\u00F5",
-    "o{": "\uFE37",
-    "o[": "\u23B4",
-    "o(": "\uFE35",
-    "pd": "\u2202",
-    "ph": "\u03D5",
-    "phi": "\u03D5",
-    "f": "\u03D5",
-    "p": "\u03C0",
-    "pi": "\u03C0",
-    "pw": "\uF361",
-    "pl": "\uF528",
-    "+-": "\u00B1",
-    "'": "\u2032",
-    "prod": "\u220F",
-    "prop": "\u221D",
-    "ps": "\u03C8",
-    "psi": "\u03C8",
-    "y": "\u03C8",
-    "rtm": "\u00AE",
-    "ret": "\u21B5",
-    " ret": "\uF766",
-    "``": "\u2036",
-    "mem": "\u220B",
-    "`": "\u2035",
-    "rv": "\u0159",
-    "r": "\u03C1",
-    "rho": "\u03C1",
-    ">": "\u232A",
-    " ->": "\u2192",
-    "r|": "\uF604",
-    "rc": "\u2309",
-    "]]": "\u301B",
-    "r||": "\uF606",
-    "rf": "\u230B",
-    "g>>": "\u00BB",
-    "]": "\uF76C",
-    "rT": "\u22A2",
-    "vec": "\u21C0",
-    "->": "\uF522",
-    ":>": "\uF51F",
-    ":-(": "\u2639",
-    "sa": "\u03E0",
-    "sampi": "\u03E0",
-    "sca": "\uF6B2",
-    "scb": "\uF6B3",
-    "scc": "\uF6B4",
-    "scA": "\uF770",
-    "scB": "\u212C",
-    "scC": "\uF772",
-    "scD": "\uF773",
-    "scE": "\u2130",
-    "scF": "\u2131",
-    "scG": "\uF776",
-    "scH": "\u210B",
-    "scI": "\u2110",
-    "scJ": "\uF779",
-    "scK": "\uF77A",
-    "scL": "\u2112",
-    "scM": "\u2133",
-    "scN": "\uF77D",
-    "scO": "\uF77E",
-    "scP": "\u2118",
-    "scQ": "\uF780",
-    "scR": "\u211B",
-    "scS": "\uF782",
-    "scT": "\uF783",
-    "scU": "\uF784",
-    "scV": "\uF785",
-    "scW": "\uF786",
-    "scX": "\uF787",
-    "scY": "\uF788",
-    "scZ": "\uF789",
-    "scd": "\uF6B5",
-    "sce": "\u212F",
-    "sc8": "\uF7F8",
-    "scf": "\uF6B7",
-    "sc5": "\uF7F5",
-    "sc4": "\uF7F4",
-    "scg": "\u210A",
-    "sch": "\uF6B9",
-    "sci": "\uF6BA",
-    "scj": "\uF6BB",
-    "sck": "\uF6BC",
-    "scl": "\u2113",
-    "scm": "\uF6BE",
-    "scn": "\uF6BF",
-    "sc9": "\uF7F9",
-    "sco": "\u2134",
-    "sc1": "\uF7F1",
-    "scp": "\uF6C1",
-    "scq": "\uF6C2",
-    "scr": "\uF6C3",
-    "scs": "\uF6C4",
-    "sc7": "\uF7F7",
-    "sc6": "\uF7F6",
-    "sct": "\uF6C5",
-    "sc3": "\uF7F3",
-    "sc2": "\uF7F2",
-    "scu": "\uF6C6",
-    "scv": "\uF6C7",
-    "scw": "\uF6C8",
-    "scx": "\uF6C9",
-    "scy": "\uF6CA",
-    "scz": "\uF6CB",
-    "sc0": "\uF7F0",
-    "spl": "\uF527",
-    "sv": "\u0161",
-    "s": "\u03C3",
-    "sigma": "\u03C3",
-    "*6": "\u2736",
-    "sc": "\u2218",
-    "space": "\u2423",
-    "spc": "\uF7BF",
-    "sqrt": "\u221A",
-    "sq": "\uF520",
-    "star": "\u22C6",
-    "sti": "\u03DB",
-    "stigma": "\u03DB",
-    "sub": "\u2282",
-    "sub=": "\u2286",
-    "st": "\u220D",
-    "sum": "\u2211",
-    "sup": "\u2283",
-    "sup=": "\u2287",
-    "sz": "\u00DF",
-    "ss": "\u00DF",
-    "tab": "\uF7BE",
-    "t": "\u03C4",
-    "tau": "\u03C4",
-    "tv": "\u0165",
-    "tf": "\u2234",
-    "th": "\u03B8",
-    "theta": "\u03B8",
-    "q": "\u03B8",
-    "    ": "\u2005",
-    "  ": "\u2009",
-    "thn": "\u00FE",
-    "~": "\u223C",
-    "~=": "\u2243",
-    "~==": "\u2245",
-    "~~": "\u2248",
-    "*": "\u00D7",
-    "tm": "\u2122",
-    "tr": "\uF3C7",
-    "u'": "\u00FA",
-    "u''": "\u0171",
-    'u"': "\u00FC",
-    "u`": "\u00F9",
-    "u^": "\u00FB",
-    "u{": "\uFE38",
-    "u[": "\u23B5",
-    "u(": "\uFE36",
-    "un": "\u22C3",
-    "u": "\u03C5",
-    "upsilon": "\u03C5",
-    "uT": "\u22A5",
-    "uo": "\u016F",
-    "v": "\u22C1",
-    " |": "\u2223",
-    "vline": "\u2502",
-    "|": "\uF432",
-    " ": "\u200A",
-    "^": "\u22C0",
-    "wp": "\u2118",
-    "wf": "\uF720",
-    "wolf": "\uF720",
-    "x": "\u03BE",
-    "xi": "\u03BE",
-    "xnor": "\uF4A2",
-    "xor": "\u22BB",
-    "y'": "\u00FD",
-    "z": "\u03B6",
-    "zeta": "\u03B6",
-    "zv": "\u017E",
+# Conversion from unicode to WL
+UNICODE_TO_WL = {
+    '𝐷': '\uf74b', # MATHEMATICAL ITALIC CAPITAL D -> \[CapitalDifferentialD]
+    '⌘': '\uf76a', # PLACE OF INTEREST SIGN -> \[CommandKey]
+    '𝕔': '\uf6e8', # MATHEMATICAL DOUBLE-STRUCK SMALL C -> \[DoubleStruckC]
+    '⨯': '\uf4a0', # VECTOR OR CROSS PRODUCT -> \[Cross]
+    '𝑑': '\uf74c', # MATHEMATICAL ITALIC SMALL D -> \[DifferentialD]
+    '→': '\uf525', # RIGHTWARDS ARROW -> \[ShortRightArrow]
+    'ϴ': '\uf4a4', # GREEK CAPITAL THETA SYMBOL -> \[DiscreteRatio]
+    'ȷ': '\uf700', # LATIN SMALL LETTER DOTLESS J -> \[DotlessJ]
+    '⛶': '\uf751', # SQUARE FOUR CORNERS -> \[DottedSquare]
+    'ℽ': '\uf74a', # DOUBLE-STRUCK SMALL GAMMA -> \[DoubledGamma]
+    'ℼ': '\uf749', # DOUBLE-STRUCK SMALL PI -> \[DoubledPi]
+    '𝕒': '\uf6e6', # MATHEMATICAL DOUBLE-STRUCK SMALL A -> \[DoubleStruckA]
+    '𝕓': '\uf6e7', # MATHEMATICAL DOUBLE-STRUCK SMALL B -> \[DoubleStruckB]
+    '𝔸': '\uf7a4', # MATHEMATICAL DOUBLE-STRUCK CAPITAL A -> \[DoubleStruckCapitalA]
+    '𝔹': '\uf7a5', # MATHEMATICAL DOUBLE-STRUCK CAPITAL B -> \[DoubleStruckCapitalB]
+    'ℂ': '\uf7a6', # DOUBLE-STRUCK CAPITAL C -> \[DoubleStruckCapitalC]
+    '𝔻': '\uf7a7', # MATHEMATICAL DOUBLE-STRUCK CAPITAL D -> \[DoubleStruckCapitalD]
+    '𝔼': '\uf7a8', # MATHEMATICAL DOUBLE-STRUCK CAPITAL E -> \[DoubleStruckCapitalE]
+    '𝔽': '\uf7a9', # MATHEMATICAL DOUBLE-STRUCK CAPITAL F -> \[DoubleStruckCapitalF]
+    '𝔾': '\uf7aa', # MATHEMATICAL DOUBLE-STRUCK CAPITAL G -> \[DoubleStruckCapitalG]
+    'ℍ': '\uf7ab', # DOUBLE-STRUCK CAPITAL H -> \[DoubleStruckCapitalH]
+    '𝕀': '\uf7ac', # MATHEMATICAL DOUBLE-STRUCK CAPITAL I -> \[DoubleStruckCapitalI]
+    '𝕁': '\uf7ad', # MATHEMATICAL DOUBLE-STRUCK CAPITAL J -> \[DoubleStruckCapitalJ]
+    '𝕂': '\uf7ae', # MATHEMATICAL DOUBLE-STRUCK CAPITAL K -> \[DoubleStruckCapitalK]
+    '𝕃': '\uf7af', # MATHEMATICAL DOUBLE-STRUCK CAPITAL L -> \[DoubleStruckCapitalL]
+    '𝕄': '\uf7b0', # MATHEMATICAL DOUBLE-STRUCK CAPITAL M -> \[DoubleStruckCapitalM]
+    'ℕ': '\uf7b1', # DOUBLE-STRUCK CAPITAL N -> \[DoubleStruckCapitalN]
+    '𝕆': '\uf7b2', # MATHEMATICAL DOUBLE-STRUCK CAPITAL O -> \[DoubleStruckCapitalO]
+    'ℙ': '\uf7b3', # DOUBLE-STRUCK CAPITAL P -> \[DoubleStruckCapitalP]
+    'ℚ': '\uf7b4', # DOUBLE-STRUCK CAPITAL Q -> \[DoubleStruckCapitalQ]
+    'ℝ': '\uf7b5', # DOUBLE-STRUCK CAPITAL R -> \[DoubleStruckCapitalR]
+    '𝕊': '\uf7b6', # MATHEMATICAL DOUBLE-STRUCK CAPITAL S -> \[DoubleStruckCapitalS]
+    '𝕋': '\uf7b7', # MATHEMATICAL DOUBLE-STRUCK CAPITAL T -> \[DoubleStruckCapitalT]
+    '𝕌': '\uf7b8', # MATHEMATICAL DOUBLE-STRUCK CAPITAL U -> \[DoubleStruckCapitalU]
+    '𝕍': '\uf7b9', # MATHEMATICAL DOUBLE-STRUCK CAPITAL V -> \[DoubleStruckCapitalV]
+    '𝕎': '\uf7ba', # MATHEMATICAL DOUBLE-STRUCK CAPITAL W -> \[DoubleStruckCapitalW]
+    '𝕏': '\uf7bb', # MATHEMATICAL DOUBLE-STRUCK CAPITAL X -> \[DoubleStruckCapitalX]
+    '𝕐': '\uf7bc', # MATHEMATICAL DOUBLE-STRUCK CAPITAL Y -> \[DoubleStruckCapitalY]
+    'ℤ': '\uf7bd', # DOUBLE-STRUCK CAPITAL Z -> \[DoubleStruckCapitalZ]
+    '𝕕': '\uf6e9', # MATHEMATICAL DOUBLE-STRUCK SMALL D -> \[DoubleStruckD]
+    '𝕖': '\uf6ea', # MATHEMATICAL DOUBLE-STRUCK SMALL E -> \[DoubleStruckE]
+    '𝟠': '\uf7e3', # MATHEMATICAL DOUBLE-STRUCK DIGIT EIGHT -> \[DoubleStruckEight]
+    '𝕗': '\uf6eb', # MATHEMATICAL DOUBLE-STRUCK SMALL F -> \[DoubleStruckF]
+    '𝟝': '\uf7e0', # MATHEMATICAL DOUBLE-STRUCK DIGIT FIVE -> \[DoubleStruckFive]
+    '𝟜': '\uf7df', # MATHEMATICAL DOUBLE-STRUCK DIGIT FOUR -> \[DoubleStruckFour]
+    '𝕘': '\uf6ec', # MATHEMATICAL DOUBLE-STRUCK SMALL G -> \[DoubleStruckG]
+    '𝕙': '\uf6ed', # MATHEMATICAL DOUBLE-STRUCK SMALL H -> \[DoubleStruckH]
+    '𝕚': '\uf6ee', # MATHEMATICAL DOUBLE-STRUCK SMALL I -> \[DoubleStruckI]
+    '𝕛': '\uf6ef', # MATHEMATICAL DOUBLE-STRUCK SMALL J -> \[DoubleStruckJ]
+    '𝕜': '\uf6f0', # MATHEMATICAL DOUBLE-STRUCK SMALL K -> \[DoubleStruckK]
+    '𝕝': '\uf6f1', # MATHEMATICAL DOUBLE-STRUCK SMALL L -> \[DoubleStruckL]
+    '𝕞': '\uf6f2', # MATHEMATICAL DOUBLE-STRUCK SMALL M -> \[DoubleStruckM]
+    '𝕟': '\uf6f3', # MATHEMATICAL DOUBLE-STRUCK SMALL N -> \[DoubleStruckN]
+    '𝟡': '\uf7e4', # MATHEMATICAL DOUBLE-STRUCK DIGIT NINE -> \[DoubleStruckNine]
+    '𝕠': '\uf6f4', # MATHEMATICAL DOUBLE-STRUCK SMALL O -> \[DoubleStruckO]
+    '𝟙': '\uf7dc', # MATHEMATICAL DOUBLE-STRUCK DIGIT ONE -> \[DoubleStruckOne]
+    '𝕡': '\uf6f5', # MATHEMATICAL DOUBLE-STRUCK SMALL P -> \[DoubleStruckP]
+    '𝕢': '\uf6f6', # MATHEMATICAL DOUBLE-STRUCK SMALL Q -> \[DoubleStruckQ]
+    '𝕣': '\uf6f7', # MATHEMATICAL DOUBLE-STRUCK SMALL R -> \[DoubleStruckR]
+    '𝕤': '\uf6f8', # MATHEMATICAL DOUBLE-STRUCK SMALL S -> \[DoubleStruckS]
+    '𝟟': '\uf7e2', # MATHEMATICAL DOUBLE-STRUCK DIGIT SEVEN -> \[DoubleStruckSeven]
+    '𝟞': '\uf7e1', # MATHEMATICAL DOUBLE-STRUCK DIGIT SIX -> \[DoubleStruckSix]
+    '𝕥': '\uf6f9', # MATHEMATICAL DOUBLE-STRUCK SMALL T -> \[DoubleStruckT]
+    '𝟛': '\uf7de', # MATHEMATICAL DOUBLE-STRUCK DIGIT THREE -> \[DoubleStruckThree]
+    '𝟚': '\uf7dd', # MATHEMATICAL DOUBLE-STRUCK DIGIT TWO -> \[DoubleStruckTwo]
+    '𝕦': '\uf6fa', # MATHEMATICAL DOUBLE-STRUCK SMALL U -> \[DoubleStruckU]
+    '𝕧': '\uf6fb', # MATHEMATICAL DOUBLE-STRUCK SMALL V -> \[DoubleStruckV]
+    '𝕨': '\uf6fc', # MATHEMATICAL DOUBLE-STRUCK SMALL W -> \[DoubleStruckW]
+    '𝕩': '\uf6fd', # MATHEMATICAL DOUBLE-STRUCK SMALL X -> \[DoubleStruckX]
+    '𝕪': '\uf6fe', # MATHEMATICAL DOUBLE-STRUCK SMALL Y -> \[DoubleStruckY]
+    '𝕫': '\uf6ff', # MATHEMATICAL DOUBLE-STRUCK SMALL Z -> \[DoubleStruckZ]
+    '𝟘': '\uf7db', # MATHEMATICAL DOUBLE-STRUCK DIGIT ZERO -> \[DoubleStruckZero]
+    '⩵': '\uf431', # TWO CONSECUTIVE EQUALS SIGNS -> \[Equal]
+    '⇔': '⧦', # LEFT RIGHT DOUBLE ARROW -> \[Equivalent]
+    'ⅇ': '\uf74d', # DOUBLE-STRUCK ITALIC SMALL E -> \[ExponentialE]
+    '•': '\uf750', # BULLET -> \[FilledSmallCircle]
+    'ạ': '\uf800', # LATIN SMALL LETTER A WITH DOT BELOW -> \[FormalA]
+    'α̣': '\uf854', # GREEK SMALL LETTER ALPHA + COMBINING DOT BELOW -> \[FormalAlpha]
+    'ḅ': '\uf801', # LATIN SMALL LETTER B WITH DOT BELOW -> \[FormalB]
+    'β̣': '\uf855', # GREEK SMALL LETTER BETA + COMBINING DOT BELOW -> \[FormalBeta]
+    'c̣': '\uf802', # LATIN SMALL LETTER C + COMBINING DOT BELOW -> \[FormalC]
+    'Ạ': '\uf81a', # LATIN CAPITAL LETTER A WITH DOT BELOW -> \[FormalCapitalA]
+    'Α̣': '\uf834', # GREEK CAPITAL LETTER ALPHA + COMBINING DOT BELOW -> \[FormalCapitalAlpha]
+    'Ḅ': '\uf81b', # LATIN CAPITAL LETTER B WITH DOT BELOW -> \[FormalCapitalB]
+    'Β̣': '\uf835', # GREEK CAPITAL LETTER BETA + COMBINING DOT BELOW -> \[FormalCapitalBeta]
+    'C̣': '\uf81c', # LATIN CAPITAL LETTER C + COMBINING DOT BELOW -> \[FormalCapitalC]
+    'Χ̣': '\uf84a', # GREEK CAPITAL LETTER CHI + COMBINING DOT BELOW -> \[FormalCapitalChi]
+    'Ḍ': '\uf81d', # LATIN CAPITAL LETTER D WITH DOT BELOW -> \[FormalCapitalD]
+    'Δ̣': '\uf837', # GREEK CAPITAL LETTER DELTA + COMBINING DOT BELOW -> \[FormalCapitalDelta]
+    'Ϝ̣': '\uf87f', # GREEK LETTER DIGAMMA + COMBINING DOT BELOW -> \[FormalCapitalDigamma]
+    'Ẹ': '\uf81e', # LATIN CAPITAL LETTER E WITH DOT BELOW -> \[FormalCapitalE]
+    'Ε̣': '\uf838', # GREEK CAPITAL LETTER EPSILON + COMBINING DOT BELOW -> \[FormalCapitalEpsilon]
+    'Η̣': '\uf83a', # GREEK CAPITAL LETTER ETA + COMBINING DOT BELOW -> \[FormalCapitalEta]
+    'F̣': '\uf81f', # LATIN CAPITAL LETTER F + COMBINING DOT BELOW -> \[FormalCapitalF]
+    'G̣': '\uf820', # LATIN CAPITAL LETTER G + COMBINING DOT BELOW -> \[FormalCapitalG]
+    'Γ': '\uf836', # GREEK CAPITAL LETTER GAMMA -> \[FormalCapitalGamma]
+    'Ḥ': '\uf821', # LATIN CAPITAL LETTER H WITH DOT BELOW -> \[FormalCapitalH]
+    'Ị': '\uf822', # LATIN CAPITAL LETTER I WITH DOT BELOW -> \[FormalCapitalI]
+    'Ι̣': '\uf85c', # GREEK CAPITAL LETTER IOTA + COMBINING DOT BELOW -> \[FormalIota]
+    'J̣': '\uf823', # LATIN CAPITAL LETTER J + COMBINING DOT BELOW -> \[FormalCapitalJ]
+    'Ḳ': '\uf824', # LATIN CAPITAL LETTER K WITH DOT BELOW -> \[FormalCapitalK]
+    'Κ̣': '\uf83d', # GREEK CAPITAL LETTER KAPPA + COMBINING DOT BELOW -> \[FormalCapitalKappa]
+    'Ϟ̣': '\uf881', # GREEK LETTER KOPPA + COMBINING DOT BELOW -> \[FormalCapitalKoppa]
+    'Ḷ': '\uf825', # LATIN CAPITAL LETTER L WITH DOT BELOW -> \[FormalCapitalL]
+    'Λ̣': '\uf83e', # GREEK CAPITAL LETTER LAMDA + COMBINING DOT BELOW -> \[FormalCapitalLambda]
+    'Ṃ': '\uf826', # LATIN CAPITAL LETTER M WITH DOT BELOW -> \[FormalCapitalM]
+    'Μ̣': '\uf83f', # GREEK CAPITAL LETTER MU + COMBINING DOT BELOW -> \[FormalCapitalMu]
+    'Ṇ': '\uf827', # LATIN CAPITAL LETTER N WITH DOT BELOW -> \[FormalCapitalN]
+    'Ν̣': '\uf840', # GREEK CAPITAL LETTER NU + COMBINING DOT BELOW -> \[FormalCapitalNu]
+    'Ọ': '\uf828', # LATIN CAPITAL LETTER O WITH DOT BELOW -> \[FormalCapitalO]
+    'Ω̣': '\uf84c', # GREEK CAPITAL LETTER OMEGA + COMBINING DOT BELOW -> \[FormalCapitalOmega]
+    'Ο̣': '\uf842', # GREEK CAPITAL LETTER OMICRON + COMBINING DOT BELOW -> \[FormalCapitalOmicron]
+    'P̣': '\uf829', # LATIN CAPITAL LETTER P + COMBINING DOT BELOW -> \[FormalCapitalP]
+    'Φ̣': '\uf849', # GREEK CAPITAL LETTER PHI + COMBINING DOT BELOW -> \[FormalCapitalPhi]
+    'Π̣': '\uf843', # GREEK CAPITAL LETTER PI + COMBINING DOT BELOW -> \[FormalCapitalPi]
+    'Ψ̣': '\uf84b', # GREEK CAPITAL LETTER PSI + COMBINING DOT BELOW -> \[FormalCapitalPsi]
+    'Q̣': '\uf82a', # LATIN CAPITAL LETTER Q + COMBINING DOT BELOW -> \[FormalCapitalQ]
+    'Ṛ': '\uf82b', # LATIN CAPITAL LETTER R WITH DOT BELOW -> \[FormalCapitalR]
+    'Ρ̣': '\uf844', # GREEK CAPITAL LETTER RHO + COMBINING DOT BELOW -> \[FormalCapitalRho]
+    'Ṣ': '\uf82c', # LATIN CAPITAL LETTER S WITH DOT BELOW -> \[FormalCapitalS]
+    'Ϡ̣': '\uf883', # GREEK LETTER SAMPI + COMBINING DOT BELOW -> \[FormalCapitalSampi]
+    'Σ̣': '\uf846', # GREEK CAPITAL LETTER SIGMA + COMBINING DOT BELOW -> \[FormalCapitalSigma]
+    'Ϛ̣': '\uf87d', # GREEK LETTER STIGMA + COMBINING DOT BELOW -> \[FormalCapitalStigma]
+    'Ṭ': '\uf82d', # LATIN CAPITAL LETTER T WITH DOT BELOW -> \[FormalCapitalT]
+    'Τ̣': '\uf847', # GREEK CAPITAL LETTER TAU + COMBINING DOT BELOW -> \[FormalCapitalTau]
+    'Θ': '\uf83b', # GREEK CAPITAL LETTER THETA -> \[FormalCapitalTheta]
+    'Ụ': '\uf82e', # LATIN CAPITAL LETTER U WITH DOT BELOW -> \[FormalCapitalU]
+    'Υ̣': '\uf848', # GREEK CAPITAL LETTER UPSILON + COMBINING DOT BELOW -> \[FormalCapitalUpsilon]
+    'Ṿ': '\uf82f', # LATIN CAPITAL LETTER V WITH DOT BELOW -> \[FormalCapitalV]
+    'Ẉ': '\uf830', # LATIN CAPITAL LETTER W WITH DOT BELOW -> \[FormalCapitalW]
+    'X̣': '\uf831', # LATIN CAPITAL LETTER X + COMBINING DOT BELOW -> \[FormalCapitalX]
+    'Ξ̣': '\uf841', # GREEK CAPITAL LETTER XI + COMBINING DOT BELOW -> \[FormalCapitalXi]
+    'Ỵ': '\uf832', # LATIN CAPITAL LETTER Y WITH DOT BELOW -> \[FormalCapitalY]
+    'Ẓ': '\uf833', # LATIN CAPITAL LETTER Z WITH DOT BELOW -> \[FormalCapitalZ]
+    'Ζ̣': '\uf839', # GREEK CAPITAL LETTER ZETA + COMBINING DOT BELOW -> \[FormalCapitalZeta]
+    'χ̣': '\uf86a', # GREEK SMALL LETTER CHI + COMBINING DOT BELOW -> \[FormalChi]
+    'ϒ̣': '\uf875', # GREEK UPSILON WITH HOOK SYMBOL + COMBINING DOT BELOW -> \[FormalCurlyCapitalUpsilon]
+    'ε̣': '\uf858', # GREEK SMALL LETTER EPSILON + COMBINING DOT BELOW -> \[FormalCurlyEpsilon]
+    'ϰ̣': '\uf885', # GREEK KAPPA SYMBOL + COMBINING DOT BELOW -> \[FormalCurlyKappa]
+    'φ̣': '\uf869', # GREEK SMALL LETTER PHI + COMBINING DOT BELOW -> \[FormalCurlyPhi]
+    'ϖ̣': '\uf879', # GREEK PI SYMBOL + COMBINING DOT BELOW -> \[FormalCurlyPi]
+    'ϱ̣': '\uf886', # GREEK RHO SYMBOL + COMBINING DOT BELOW -> \[FormalCurlyRho]
+    'ϑ̣': '\uf874', # GREEK THETA SYMBOL + COMBINING DOT BELOW -> \[FormalCurlyTheta]
+    'ḍ': '\uf803', # LATIN SMALL LETTER D WITH DOT BELOW -> \[FormalD]
+    'δ̣': '\uf857', # GREEK SMALL LETTER DELTA + COMBINING DOT BELOW -> \[FormalDelta]
+    'ϝ': '\uf880', # GREEK SMALL LETTER DIGAMMA -> \[FormalDigamma]
+    'ẹ': '\uf804', # LATIN SMALL LETTER E WITH DOT BELOW -> \[FormalE]
+    'ϵ̣': '\uf88a', # GREEK LUNATE EPSILON SYMBOL + COMBINING DOT BELOW -> \[FormalEpsilon]
+    'η̣': '\uf85a', # GREEK SMALL LETTER ETA + COMBINING DOT BELOW -> \[FormalEta]
+    'f̣': '\uf805', # LATIN SMALL LETTER F + COMBINING DOT BELOW -> \[FormalF]
+    'ς̣': '\uf865', # GREEK SMALL LETTER FINAL SIGMA + COMBINING DOT BELOW -> \[FormalFinalSigma]
+    'g̣': '\uf806', # LATIN SMALL LETTER G + COMBINING DOT BELOW -> \[FormalG]
+    'γ̣': '\uf856', # GREEK SMALL LETTER GAMMA + COMBINING DOT BELOW -> \[FormalGamma]
+    'ḥ': '\uf807', # LATIN SMALL LETTER H WITH DOT BELOW -> \[FormalH]
+    'ị': '\uf808', # LATIN SMALL LETTER I WITH DOT BELOW -> \[FormalI]
+    'j̣': '\uf809', # LATIN SMALL LETTER J + COMBINING DOT BELOW -> \[FormalJ]
+    'ḳ': '\uf80a', # LATIN SMALL LETTER K WITH DOT BELOW -> \[FormalK]
+    'κ̣': '\uf85d', # GREEK SMALL LETTER KAPPA + COMBINING DOT BELOW -> \[FormalKappa]
+    'ϟ̣': '\uf882', # GREEK SMALL LETTER KOPPA + COMBINING DOT BELOW -> \[FormalKoppa]
+    'ḷ': '\uf80b', # LATIN SMALL LETTER L WITH DOT BELOW -> \[FormalL]
+    'λ̣': '\uf85e', # GREEK SMALL LETTER LAMDA + COMBINING DOT BELOW -> \[FormalLambda]
+    'ṃ': '\uf80c', # LATIN SMALL LETTER M + COMBINING DOT BELOW -> \[FormalM]
+    'μ̣': '\uf85f', # GREEK SMALL LETTER MU + COMBINING DOT BELOW -> \[FormalMu]
+    'ṇ': '\uf80d', # LATIN SMALL LETTER N + COMBINING DOT BELOW -> \[FormalN]
+    'ν̣': '\uf860', # GREEK SMALL LETTER NU + COMBINING DOT BELOW -> \[FormalNu]
+    'ọ': '\uf80e', # LATIN SMALL LETTER O + COMBINING DOT BELOW -> \[FormalO]
+    'ω̣': '\uf86c', # GREEK SMALL LETTER OMEGA + COMBINING DOT BELOW -> \[FormalOmega]
+    'ο̣': '\uf862', # GREEK SMALL LETTER OMICRON + COMBINING DOT BELOW -> \[FormalOmicron]
+    'p̣': '\uf80f', # LATIN SMALL LETTER P + COMBINING DOT BELOW -> \[FormalP]
+    'ϕ̣': '\uf878', # GREEK PHI SYMBOL + COMBINING DOT BELOW -> \[FormalPhi]
+    'π̣': '\uf863', # GREEK SMALL LETTER PI + COMBINING DOT BELOW -> \[FormalPi]
+    'ψ̣': '\uf86b', # GREEK SMALL LETTER PSI + COMBINING DOT BELOW -> \[FormalPsi]
+    'q̣': '\uf810', # LATIN SMALL LETTER Q + COMBINING DOT BELOW -> \[FormalQ]
+    'ṛ': '\uf811', # LATIN SMALL LETTER R WITH DOT BELOW -> \[FormalR]
+    'ρ̣': '\uf864', # GREEK SMALL LETTER RHO + COMBINING DOT BELOW -> \[FormalRho]
+    'ṣ': '\uf812', # LATIN SMALL LETTER S WITH DOT BELOW -> \[FormalS]
+    # This one is changed because GREEK LETTER SAMPI is mapped to \[CapitalSampi] already
+    'ϡ̣': '\uf884', # GREEK SMALL LETTER SAMPI + COMBINING DOT BELOW -> \[FormalSampi]
+    'σ̣': '\uf866', # GREEK SMALL LETTER SIGMA + COMBINING DOT BELOW -> \[FormalSigma]
+    'ϛ': '\uf87e', # GREEK SMALL LETTER STIGMA -> \[FormalStigma]
+    'ṭ': '\uf813', # LATIN SMALL LETTER T WITH DOT BELOW -> \[FormalT]
+    'τ̣': '\uf867', # GREEK SMALL LETTER TAU + COMBINING DOT BELOW -> \[FormalTau]
+    'θ̣': '\uf85b', # GREEK SMALL LETTER THETA + COMBINING DOT BELOW -> \[FormalTheta]
+    'ụ': '\uf814', # LATIN SMALL LETTER U WITH DOT BELOW -> \[FormalU]
+    'υ̣': '\uf868', # GREEK SMALL LETTER UPSILON + COMBINING DOT BELOW -> \[FormalUpsilon]
+    'ṿ': '\uf815', # LATIN SMALL LETTER V WITH DOT BELOW -> \[FormalV]
+    'ẉ': '\uf816', # LATIN SMALL LETTER W WITH DOT BELOW -> \[FormalW]
+    'x̣': '\uf817', # LATIN SMALL LETTER X + COMBINING DOT BELOW -> \[FormalX]
+    'ξ̣': '\uf861', # GREEK SMALL LETTER XI + COMBINING DOT BELOW -> \[FormalXi]
+    'ỵ': '\uf818', # LATIN SMALL LETTER Y WITH DOT BELOW -> \[FormalY]
+    'ẓ': '\uf819', # LATIN SMALL LETTER Z WITH DOT BELOW -> \[FormalZ]
+    'ζ̣': '\uf859', # GREEK SMALL LETTER ZETA + COMBINING DOT BELOW -> \[FormalZeta]
+    '↦': '\uf4a1', # RIGHTWARDS ARROW FROM BAR -> \[Function]
+    '𝔞': '\uf6cc', # MATHEMATICAL FRAKTUR SMALL A -> \[GothicA]
+    '𝔟': '\uf6cd', # MATHEMATICAL FRAKTUR SMALL B -> \[GothicB]
+    '𝔠': '\uf6ce', # MATHEMATICAL FRAKTUR SMALL C -> \[GothicC]
+    '𝔄': '\uf78a', # MATHEMATICAL FRAKTUR CAPITAL A -> \[GothicCapitalA]
+    '𝔅': '\uf78b', # MATHEMATICAL FRAKTUR CAPITAL B -> \[GothicCapitalB]
+    '𝔇': '\uf78d', # MATHEMATICAL FRAKTUR CAPITAL D -> \[GothicCapitalD]
+    '𝔈': '\uf78e', # MATHEMATICAL FRAKTUR CAPITAL E -> \[GothicCapitalE]
+    '𝔉': '\uf78f', # MATHEMATICAL FRAKTUR CAPITAL F -> \[GothicCapitalF]
+    '𝔊': '\uf790', # MATHEMATICAL FRAKTUR CAPITAL G -> \[GothicCapitalG]
+    '𝔍': '\uf793', # MATHEMATICAL FRAKTUR CAPITAL J -> \[GothicCapitalJ]
+    '𝔎': '\uf794', # MATHEMATICAL FRAKTUR CAPITAL K -> \[GothicCapitalK]
+    '𝔏': '\uf795', # MATHEMATICAL FRAKTUR CAPITAL L -> \[GothicCapitalL]
+    '𝔐': '\uf796', # MATHEMATICAL FRAKTUR CAPITAL M -> \[GothicCapitalM]
+    '𝔑': '\uf797', # MATHEMATICAL FRAKTUR CAPITAL N -> \[GothicCapitalN]
+    '𝔒': '\uf798', # MATHEMATICAL FRAKTUR CAPITAL O -> \[GothicCapitalO]
+    '𝔓': '\uf799', # MATHEMATICAL FRAKTUR CAPITAL P -> \[GothicCapitalP]
+    '𝔔': '\uf79a', # MATHEMATICAL FRAKTUR CAPITAL Q -> \[GothicCapitalQ]
+    '𝔖': '\uf79c', # MATHEMATICAL FRAKTUR CAPITAL S -> \[GothicCapitalS]
+    '𝔗': '\uf79d', # MATHEMATICAL FRAKTUR CAPITAL T -> \[GothicCapitalT]
+    '𝔘': '\uf79e', # MATHEMATICAL FRAKTUR CAPITAL U -> \[GothicCapitalU]
+    '𝔙': '\uf79f', # MATHEMATICAL FRAKTUR CAPITAL V -> \[GothicCapitalV]
+    '𝔚': '\uf7a0', # MATHEMATICAL FRAKTUR CAPITAL W -> \[GothicCapitalW]
+    '𝔛': '\uf7a1', # MATHEMATICAL FRAKTUR CAPITAL X -> \[GothicCapitalX]
+    '𝔜': '\uf7a2', # MATHEMATICAL FRAKTUR CAPITAL Y -> \[GothicCapitalY]
+    '𝔡': '\uf6cf', # MATHEMATICAL FRAKTUR SMALL D -> \[GothicD]
+    '𝔢': '\uf6d0', # MATHEMATICAL FRAKTUR SMALL E -> \[GothicE]
+    '𝔣': '\uf6d1', # MATHEMATICAL FRAKTUR SMALL F -> \[GothicF]
+    '𝔤': '\uf6d2', # MATHEMATICAL FRAKTUR SMALL G -> \[GothicG]
+    '𝔥': '\uf6d3', # MATHEMATICAL FRAKTUR SMALL H -> \[GothicH]
+    '𝔦': '\uf6d4', # MATHEMATICAL FRAKTUR SMALL I -> \[GothicI]
+    '𝔧': '\uf6d5', # MATHEMATICAL FRAKTUR SMALL J -> \[GothicJ]
+    '𝔨': '\uf6d6', # MATHEMATICAL FRAKTUR SMALL K -> \[GothicK]
+    '𝔩': '\uf6d7', # MATHEMATICAL FRAKTUR SMALL L -> \[GothicL]
+    '𝔪': '\uf6d8', # MATHEMATICAL FRAKTUR SMALL M -> \[GothicM]
+    '𝔫': '\uf6d9', # MATHEMATICAL FRAKTUR SMALL N -> \[GothicN]
+    '𝔬': '\uf6da', # MATHEMATICAL FRAKTUR SMALL O -> \[GothicO]
+    '𝔭': '\uf6db', # MATHEMATICAL FRAKTUR SMALL P -> \[GothicP]
+    '𝔮': '\uf6dc', # MATHEMATICAL FRAKTUR SMALL Q -> \[GothicQ]
+    '𝔯': '\uf6dd', # MATHEMATICAL FRAKTUR SMALL R -> \[GothicR]
+    '𝔰': '\uf6de', # MATHEMATICAL FRAKTUR SMALL S -> \[GothicS]
+    '𝔱': '\uf6df', # MATHEMATICAL FRAKTUR SMALL T -> \[GothicT]
+    '𝔲': '\uf6e0', # MATHEMATICAL FRAKTUR SMALL U -> \[GothicU]
+    '𝔳': '\uf6e1', # MATHEMATICAL FRAKTUR SMALL V -> \[GothicV]
+    '𝔴': '\uf6e2', # MATHEMATICAL FRAKTUR SMALL W -> \[GothicW]
+    '𝔵': '\uf6e3', # MATHEMATICAL FRAKTUR SMALL X -> \[GothicX]
+    '𝔶': '\uf6e4', # MATHEMATICAL FRAKTUR SMALL Y -> \[GothicY]
+    '𝔷': '\uf6e5', # MATHEMATICAL FRAKTUR SMALL Z -> \[GothicZ]
+    '●': '\uf753', # BLACK CIRCLE -> \[GrayCircle]
+    'ⅈ': '\uf74e', # DOUBLE-STRUCK ITALIC SMALL I -> \[ImaginaryI]
+    'ⅉ': '\uf74f', # DOUBLE-STRUCK ITALIC SMALL J -> \[ImaginaryJ]
+    '⟹': '\uf523', # LONG RIGHTWARDS DOUBLE ARROW -> \[Implies]
+    '|': '\uf3d0', # VERTICAL LINE -> 
+    '«': '\uf761', # LEFT-POINTING DOUBLE ANGLE QUOTATION MARK -> \[LeftSkeleton]
+    # The following is ommited so that `a := b` or `a = b` don't get converted to `a :\uf7d9 b` or `a \uf7d9 b`
+    # '=': '\uf7d9', # EQUALS SIGN -> \[LongEqual]
+    '#': '\uf724', # NUMBER SIGN -> \[NumberSign]
+    '⊙': '\uf3de', # CIRCLED DOT OPERATOR -> \[PermutationProduct]
+    '⎕': '\uf528', # APL FUNCTIONAL SYMBOL QUAD -> \[Placeholder]
+    '»': '\uf762', # RIGHT-POINTING DOUBLE ANGLE QUOTATION MARK -> \[RightSkeleton]
+    '⧴': '\uf51f', # RULE-DELAYED -> \[RuleDelayed]
+    'ϡ': 'Ϡ', # GREEK SMALL LETTER SAMPI -> \[Sampi]
+    '𝒶': '\uf6b2', # MATHEMATICAL SCRIPT SMALL A -> \[ScriptA]
+    '𝒷': '\uf6b3', # MATHEMATICAL SCRIPT SMALL B -> \[ScriptB]
+    '𝒸': '\uf6b4', # MATHEMATICAL SCRIPT SMALL C -> \[ScriptC]
+    '𝒜': '\uf770', # MATHEMATICAL SCRIPT CAPITAL A -> \[ScriptCapitalA]
+    '𝒞': '\uf772', # MATHEMATICAL SCRIPT CAPITAL C -> \[ScriptCapitalC]
+    '𝒟': '\uf773', # MATHEMATICAL SCRIPT CAPITAL D -> \[ScriptCapitalD]
+    '𝒢': '\uf776', # MATHEMATICAL SCRIPT CAPITAL G -> \[ScriptCapitalG]
+    '𝒥': '\uf779', # MATHEMATICAL SCRIPT CAPITAL J -> \[ScriptCapitalJ]
+    '𝒦': '\uf77a', # MATHEMATICAL SCRIPT CAPITAL K -> \[ScriptCapitalK]
+    '𝒩': '\uf77d', # MATHEMATICAL SCRIPT CAPITAL N -> \[ScriptCapitalN]
+    '𝒪': '\uf77e', # MATHEMATICAL SCRIPT CAPITAL O -> \[ScriptCapitalO]
+    '𝒬': '\uf780', # MATHEMATICAL SCRIPT CAPITAL Q -> \[ScriptCapitalQ]
+    '𝒮': '\uf782', # MATHEMATICAL SCRIPT CAPITAL S -> \[ScriptCapitalS]
+    '𝒯': '\uf783', # MATHEMATICAL SCRIPT CAPITAL T -> \[ScriptCapitalT]
+    '𝒰': '\uf784', # MATHEMATICAL SCRIPT CAPITAL U -> \[ScriptCapitalU]
+    '𝒱': '\uf785', # MATHEMATICAL SCRIPT CAPITAL V -> \[ScriptCapitalV]
+    '𝒲': '\uf786', # MATHEMATICAL SCRIPT CAPITAL W -> \[ScriptCapitalW]
+    '𝒳': '\uf787', # MATHEMATICAL SCRIPT CAPITAL X -> \[ScriptCapitalX]
+    '𝒴': '\uf788', # MATHEMATICAL SCRIPT CAPITAL Y -> \[ScriptCapitalY]
+    '𝒵': '\uf789', # MATHEMATICAL SCRIPT CAPITAL Z -> \[ScriptCapitalZ]
+    '𝒹': '\uf6b5', # MATHEMATICAL SCRIPT SMALL D -> \[ScriptD]
+    '𝒾': '\uf6ba', # MATHEMATICAL SCRIPT SMALL I -> \[ScriptI]
+    '𝒿': '\uf6bb', # MATHEMATICAL SCRIPT SMALL J -> \[ScriptJ]
+    '𝒻': '\uf6b7', # MATHEMATICAL SCRIPT SMALL F -> \[ScriptF]
+    '𝒽': '\uf6b9', # MATHEMATICAL SCRIPT SMALL H -> \[ScriptH]
+    '𝓀': '\uf6bc', # MATHEMATICAL SCRIPT SMALL K -> \[ScriptK]
+    '𝓂': '\uf6be', # MATHEMATICAL SCRIPT SMALL M -> \[ScriptM]
+    '𝓃': '\uf6bf', # MATHEMATICAL SCRIPT SMALL N -> \[ScriptN]
+    '𝓅': '\uf6c1', # MATHEMATICAL SCRIPT SMALL P -> \[ScriptP]
+    '𝓆': '\uf6c2', # MATHEMATICAL SCRIPT SMALL Q -> \[ScriptQ]
+    '𝓇': '\uf6c3', # MATHEMATICAL SCRIPT SMALL R -> \[ScriptR]
+    '𝓈': '\uf6c4', # MATHEMATICAL SCRIPT SMALL S -> \[ScriptS]
+    '𝓉': '\uf6c5', # MATHEMATICAL SCRIPT SMALL T -> \[ScriptT]
+    '𝓊': '\uf6c6', # MATHEMATICAL SCRIPT SMALL U -> \[ScriptU]
+    '𝓋': '\uf6c7', # MATHEMATICAL SCRIPT SMALL V -> \[ScriptV]
+    '𝓌': '\uf6c8', # MATHEMATICAL SCRIPT SMALL W -> \[ScriptW]
+    '𝓍': '\uf6c9', # MATHEMATICAL SCRIPT SMALL X -> \[ScriptX]
+    '𝓎': '\uf6ca', # MATHEMATICAL SCRIPT SMALL Y -> \[ScriptY]
+    '𝓏': '\uf6cb', # MATHEMATICAL SCRIPT SMALL Z -> \[ScriptZ]
+    '↓': '\uf52b', # DOWNWARDS ARROW -> \[ShortDownArrow]
+    '←': '\uf526', # LEFTWARDS ARROW -> \[ShortLeftArrow]
+    '↑': '\uf52a', # UPWARDS ARROW -> \[ShortUpArrow]
+    '▫': '\uf520', # WHITE SMALL SQUARE -> \[Square]
+    '⊗': '\uf3da', # CIRCLED TIMES -> \[TensorProduct]
+    'ᵀ': '\uf3c7', # MODIFIER LETTER CAPITAL T -> \[Transpose]
+    '↔': '\uf3d4', # LEFT RIGHT ARROW -> \[UndirectedEdge]
+    '⛢': '♅', # ASTRONOMICAL SYMBOL FOR URANUS -> \[Uranus]
 }
+
+# Conversion from WL to the fully qualified names
+WL_TO_NAMED = {
+    'á': 'AAcute',
+    'ā': 'ABar',
+    'ă': 'ACup',
+    'ä': 'ADoubleDot',
+    'æ': 'AE',
+    'à': 'AGrave',
+    'â': 'AHat',
+    'ℵ': 'Aleph',
+    '\uf764': 'AliasDelimiter',
+    '\uf768': 'AliasIndicator',
+    '\uf760': 'AlignmentMarker',
+    'α': 'Alpha',
+    '\uf7d1': 'AltKey',
+    '∧': 'And',
+    '∠': 'Angle',
+    'Å': 'Angstrom',
+    '♒': 'AquariusSign',
+    '♈': 'AriesSign',
+    'å': 'ARing',
+    '⋰': 'AscendingEllipsis',
+    'ã': 'ATilde',
+    '\uf3a8': 'AutoLeftMatch',
+    '\uf3ae': 'AutoOperand',
+    '\uf3a4': 'AutoPlaceholder',
+    '\uf3a9': 'AutoRightMatch',
+    '\uf3ad': 'AutoSpace',
+    '∖': 'Backslash',
+    '♫': 'BeamedEighthNote',
+    '♬': 'BeamedSixteenthNote',
+    '∵': 'Because',
+    'ℶ': 'Bet',
+    'β': 'Beta',
+    '♝': 'BlackBishop',
+    '♚': 'BlackKing',
+    '♞': 'BlackKnight',
+    '♟': 'BlackPawn',
+    '♛': 'BlackQueen',
+    '♜': 'BlackRook',
+    '˘': 'Breve',
+    '•': 'Bullet',
+    'ć': 'CAcute',
+    '♋': 'CancerSign',
+    '⌢': 'Cap',
+    'Á': 'CapitalAAcute',
+    'Ā': 'CapitalABar',
+    'Ă': 'CapitalACup',
+    'Ä': 'CapitalADoubleDot',
+    'Æ': 'CapitalAE',
+    'À': 'CapitalAGrave',
+    'Â': 'CapitalAHat',
+    'Α': 'CapitalAlpha',
+    'Å': 'CapitalARing',
+    'Ã': 'CapitalATilde',
+    'Β': 'CapitalBeta',
+    'Ć': 'CapitalCAcute',
+    'Ç': 'CapitalCCedilla',
+    'Č': 'CapitalCHacek',
+    'Χ': 'CapitalChi',
+    'Δ': 'CapitalDelta',
+    'Ď': 'CapitalDHacek',
+    '\uf74b': 'CapitalDifferentialD',
+    'Ϝ': 'CapitalDigamma',
+    'É': 'CapitalEAcute',
+    'Ē': 'CapitalEBar',
+    'Ĕ': 'CapitalECup',
+    'Ë': 'CapitalEDoubleDot',
+    'È': 'CapitalEGrave',
+    'Ě': 'CapitalEHacek',
+    'Ê': 'CapitalEHat',
+    'Ε': 'CapitalEpsilon',
+    'Η': 'CapitalEta',
+    'Ð': 'CapitalEth',
+    'Γ': 'CapitalGamma',
+    'Í': 'CapitalIAcute',
+    'Ĭ': 'CapitalICup',
+    'Ï': 'CapitalIDoubleDot',
+    'Ì': 'CapitalIGrave',
+    'Î': 'CapitalIHat',
+    'Ι': 'CapitalIota',
+    'Κ': 'CapitalKappa',
+    'Ϟ': 'CapitalKoppa',
+    'Λ': 'CapitalLambda',
+    'Ł': 'CapitalLSlash',
+    'Μ': 'CapitalMu',
+    'Ň': 'CapitalNHacek',
+    'Ñ': 'CapitalNTilde',
+    'Ν': 'CapitalNu',
+    'Ó': 'CapitalOAcute',
+    'Ő': 'CapitalODoubleAcute',
+    'Ö': 'CapitalODoubleDot',
+    'Œ': 'CapitalOE',
+    'Ò': 'CapitalOGrave',
+    'Ô': 'CapitalOHat',
+    'Ω': 'CapitalOmega',
+    'Ο': 'CapitalOmicron',
+    'Ø': 'CapitalOSlash',
+    'Õ': 'CapitalOTilde',
+    'Φ': 'CapitalPhi',
+    'Π': 'CapitalPi',
+    'Ψ': 'CapitalPsi',
+    'Ř': 'CapitalRHacek',
+    'Ρ': 'CapitalRho',
+    'Ϡ': 'CapitalSampi',
+    'Š': 'CapitalSHacek',
+    'Σ': 'CapitalSigma',
+    'Ϛ': 'CapitalStigma',
+    'Τ': 'CapitalTau',
+    'Ť': 'CapitalTHacek',
+    'Θ': 'CapitalTheta',
+    'Þ': 'CapitalThorn',
+    'Ú': 'CapitalUAcute',
+    'Ű': 'CapitalUDoubleAcute',
+    'Ü': 'CapitalUDoubleDot',
+    'Ù': 'CapitalUGrave',
+    'Û': 'CapitalUHat',
+    'Υ': 'CapitalUpsilon',
+    'Ů': 'CapitalURing',
+    'Ξ': 'CapitalXi',
+    'Ý': 'CapitalYAcute',
+    'Ζ': 'CapitalZeta',
+    'Ž': 'CapitalZHacek',
+    '♑': 'CapricornSign',
+    'ç': 'CCedilla',
+    '¸': 'Cedilla',
+    '·': 'CenterDot',
+    '⋯': 'CenterEllipsis',
+    '¢': 'Cent',
+    'č': 'CHacek',
+    '☒': 'CheckedBox',
+    '✓': 'Checkmark',
+    'χ': 'Chi',
+    '⊙': 'CircleDot',
+    '⊖': 'CircleMinus',
+    '⊕': 'CirclePlus',
+    '⊗': 'CircleTimes',
+    '∲': 'ClockwiseContourIntegral',
+    '”': 'CloseCurlyDoubleQuote',
+    '’': 'CloseCurlyQuote',
+    '⌘': 'CloverLeaf',
+    '♣': 'ClubSuit',
+    '∶': 'Colon',
+    '\uf76a': 'CommandKey',
+    '\uf3d3': 'Conditioned',
+    '≡': 'Congruent',
+    '\uf3c8': 'Conjugate',
+    '\uf3c9': 'ConjugateTranspose',
+    '\uf7da': 'ConstantC',
+    '\uf3b1': 'Continuation',
+    '∮': 'ContourIntegral',
+    '\uf763': 'ControlKey',
+    '∐': 'Coproduct',
+    '©': 'Copyright',
+    '∳': 'CounterClockwiseContourIntegral',
+    '\uf4a0': 'Cross',
+    '≍': 'CupCap',
+    '⌣': 'Cup',
+    'ϒ': 'CurlyCapitalUpsilon',
+    'ε': 'CurlyEpsilon',
+    'ϰ': 'CurlyKappa',
+    'φ': 'CurlyPhi',
+    'ϖ': 'CurlyPi',
+    'ϱ': 'CurlyRho',
+    'ϑ': 'CurlyTheta',
+    '¤': 'Currency',
+    '†': 'Dagger',
+    'ℸ': 'Dalet',
+    '–': 'Dash',
+    '°': 'Degree',
+    '\uf7d0': 'DeleteKey',
+    '∇': 'Del',
+    'δ': 'Delta',
+    '⋱': 'DescendingEllipsis',
+    'ď': 'DHacek',
+    '⌀': 'Diameter',
+    '⋄': 'Diamond',
+    '♢': 'DiamondSuit',
+    '∆': 'DifferenceDelta',
+    '\uf74c': 'DifferentialD',
+    'ϝ': 'Digamma',
+    '\uf3d5': 'DirectedEdge',
+    '\uf4a4': 'DiscreteRatio',
+    '\uf4a3': 'DiscreteShift',
+    '\xad': 'DiscretionaryHyphen',
+    '\uf76e': 'DiscretionaryLineSeparator',
+    '\uf3bf': 'DiscretionaryPageBreakAbove',
+    '\uf3c6': 'DiscretionaryPageBreakBelow',
+    '\uf76f': 'DiscretionaryParagraphSeparator',
+    '\uf3d2': 'Distributed',
+    '∣': 'Divides',
+    '÷': 'Divide',
+    '≐': 'DotEqual',
+    'ı': 'DotlessI',
+    '\uf700': 'DotlessJ',
+    '\uf751': 'DottedSquare',
+    '∯': 'DoubleContourIntegral',
+    '‡': 'DoubleDagger',
+    '\uf74a': 'DoubledGamma',
+    '¨': 'DoubleDot',
+    '⇓': 'DoubleDownArrow',
+    '\uf749': 'DoubledPi',
+    '⇐': 'DoubleLeftArrow',
+    '⇔': 'DoubleLeftRightArrow',
+    '⫤': 'DoubleLeftTee',
+    '⟸': 'DoubleLongLeftArrow',
+    '⟺': 'DoubleLongLeftRightArrow',
+    '⟹': 'DoubleLongRightArrow',
+    '″': 'DoublePrime',
+    '⇒': 'DoubleRightArrow',
+    '⊨': 'DoubleRightTee',
+    '\uf6e6': 'DoubleStruckA',
+    '\uf6e7': 'DoubleStruckB',
+    '\uf6e8': 'DoubleStruckC',
+    '\uf7a4': 'DoubleStruckCapitalA',
+    '\uf7a5': 'DoubleStruckCapitalB',
+    '\uf7a6': 'DoubleStruckCapitalC',
+    '\uf7a7': 'DoubleStruckCapitalD',
+    '\uf7a8': 'DoubleStruckCapitalE',
+    '\uf7a9': 'DoubleStruckCapitalF',
+    '\uf7aa': 'DoubleStruckCapitalG',
+    '\uf7ab': 'DoubleStruckCapitalH',
+    '\uf7ac': 'DoubleStruckCapitalI',
+    '\uf7ad': 'DoubleStruckCapitalJ',
+    '\uf7ae': 'DoubleStruckCapitalK',
+    '\uf7af': 'DoubleStruckCapitalL',
+    '\uf7b0': 'DoubleStruckCapitalM',
+    '\uf7b1': 'DoubleStruckCapitalN',
+    '\uf7b2': 'DoubleStruckCapitalO',
+    '\uf7b3': 'DoubleStruckCapitalP',
+    '\uf7b4': 'DoubleStruckCapitalQ',
+    '\uf7b5': 'DoubleStruckCapitalR',
+    '\uf7b6': 'DoubleStruckCapitalS',
+    '\uf7b7': 'DoubleStruckCapitalT',
+    '\uf7b8': 'DoubleStruckCapitalU',
+    '\uf7b9': 'DoubleStruckCapitalV',
+    '\uf7ba': 'DoubleStruckCapitalW',
+    '\uf7bb': 'DoubleStruckCapitalX',
+    '\uf7bc': 'DoubleStruckCapitalY',
+    '\uf7bd': 'DoubleStruckCapitalZ',
+    '\uf6e9': 'DoubleStruckD',
+    '\uf6ea': 'DoubleStruckE',
+    '\uf7e3': 'DoubleStruckEight',
+    '\uf6eb': 'DoubleStruckF',
+    '\uf7e0': 'DoubleStruckFive',
+    '\uf7df': 'DoubleStruckFour',
+    '\uf6ec': 'DoubleStruckG',
+    '\uf6ed': 'DoubleStruckH',
+    '\uf6ee': 'DoubleStruckI',
+    '\uf6ef': 'DoubleStruckJ',
+    '\uf6f0': 'DoubleStruckK',
+    '\uf6f1': 'DoubleStruckL',
+    '\uf6f2': 'DoubleStruckM',
+    '\uf6f3': 'DoubleStruckN',
+    '\uf7e4': 'DoubleStruckNine',
+    '\uf6f4': 'DoubleStruckO',
+    '\uf7dc': 'DoubleStruckOne',
+    '\uf6f5': 'DoubleStruckP',
+    '\uf6f6': 'DoubleStruckQ',
+    '\uf6f7': 'DoubleStruckR',
+    '\uf6f8': 'DoubleStruckS',
+    '\uf7e2': 'DoubleStruckSeven',
+    '\uf7e1': 'DoubleStruckSix',
+    '\uf6f9': 'DoubleStruckT',
+    '\uf7de': 'DoubleStruckThree',
+    '\uf7dd': 'DoubleStruckTwo',
+    '\uf6fa': 'DoubleStruckU',
+    '\uf6fb': 'DoubleStruckV',
+    '\uf6fc': 'DoubleStruckW',
+    '\uf6fd': 'DoubleStruckX',
+    '\uf6fe': 'DoubleStruckY',
+    '\uf6ff': 'DoubleStruckZ',
+    '\uf7db': 'DoubleStruckZero',
+    '⇑': 'DoubleUpArrow',
+    '⇕': 'DoubleUpDownArrow',
+    '∥': 'DoubleVerticalBar',
+    '⤓': 'DownArrowBar',
+    '↓': 'DownArrow',
+    '⇵': 'DownArrowUpArrow',
+    '\uf755': 'DownBreve',
+    '¡': 'DownExclamation',
+    '⥐': 'DownLeftRightVector',
+    '⥞': 'DownLeftTeeVector',
+    '↽': 'DownLeftVector',
+    '⥖': 'DownLeftVectorBar',
+    '▾': 'DownPointer',
+    '¿': 'DownQuestion',
+    '⥟': 'DownRightTeeVector',
+    '⇁': 'DownRightVector',
+    '⥗': 'DownRightVectorBar',
+    '↧': 'DownTeeArrow',
+    '⊤': 'DownTee',
+    'é': 'EAcute',
+    '♁': 'Earth',
+    'ē': 'EBar',
+    'ĕ': 'ECup',
+    'ë': 'EDoubleDot',
+    'è': 'EGrave',
+    'ě': 'EHacek',
+    'ê': 'EHat',
+    '♪': 'EighthNote',
+    '∈': 'Element',
+    '…': 'Ellipsis',
+    '○': 'EmptyCircle',
+    '◇': 'EmptyDiamond',
+    '▽': 'EmptyDownTriangle',
+    '▯': 'EmptyRectangle',
+    '∅': 'EmptySet',
+    '◦': 'EmptySmallCircle',
+    '◻': 'EmptySmallSquare',
+    '□': 'EmptySquare',
+    '△': 'EmptyUpTriangle',
+    '▫': 'EmptyVerySmallSquare',
+    '\uf7d4': 'EnterKey',
+    '\uf3b9': 'EntityEnd',
+    '\uf3b8': 'EntityStart',
+    'ϵ': 'Epsilon',
+    '\uf431': 'Equal',
+    '≂': 'EqualTilde',
+    '⇌': 'Equilibrium',
+    '⧦': 'Equivalent',
+    '\uf767': 'ErrorIndicator',
+    '\uf769': 'EscapeKey',
+    'η': 'Eta',
+    'ð': 'Eth',
+    '€': 'Euro',
+    '∃': 'Exists',
+    '\uf74d': 'ExponentialE',
+    'ﬁ': 'FiLigature',
+    '●': 'FilledCircle',
+    '◆': 'FilledDiamond',
+    '▼': 'FilledDownTriangle',
+    '◀': 'FilledLeftTriangle',
+    '▮': 'FilledRectangle',
+    '▶': 'FilledRightTriangle',
+    '\uf750': 'FilledSmallCircle',
+    '◼': 'FilledSmallSquare',
+    '■': 'FilledSquare',
+    '▲': 'FilledUpTriangle',
+    '▪': 'FilledVerySmallSquare',
+    'ς': 'FinalSigma',
+    '\uf7fa': 'FirstPage',
+    '★': 'FivePointedStar',
+    '♭': 'Flat',
+    'ﬂ': 'FlLigature',
+    'ƒ': 'Florin',
+    '∀': 'ForAll',
+    '\uf800': 'FormalA',
+    '\uf854': 'FormalAlpha',
+    '\uf801': 'FormalB',
+    '\uf855': 'FormalBeta',
+    '\uf802': 'FormalC',
+    '\uf81a': 'FormalCapitalA',
+    '\uf834': 'FormalCapitalAlpha',
+    '\uf81b': 'FormalCapitalB',
+    '\uf835': 'FormalCapitalBeta',
+    '\uf81c': 'FormalCapitalC',
+    '\uf84a': 'FormalCapitalChi',
+    '\uf81d': 'FormalCapitalD',
+    '\uf837': 'FormalCapitalDelta',
+    '\uf87f': 'FormalCapitalDigamma',
+    '\uf81e': 'FormalCapitalE',
+    '\uf838': 'FormalCapitalEpsilon',
+    '\uf83a': 'FormalCapitalEta',
+    '\uf81f': 'FormalCapitalF',
+    '\uf820': 'FormalCapitalG',
+    '\uf836': 'FormalCapitalGamma',
+    '\uf821': 'FormalCapitalH',
+    '\uf822': 'FormalCapitalI',
+    '\uf83c': 'FormalCapitalIota',
+    '\uf823': 'FormalCapitalJ',
+    '\uf824': 'FormalCapitalK',
+    '\uf83d': 'FormalCapitalKappa',
+    '\uf881': 'FormalCapitalKoppa',
+    '\uf825': 'FormalCapitalL',
+    '\uf83e': 'FormalCapitalLambda',
+    '\uf826': 'FormalCapitalM',
+    '\uf83f': 'FormalCapitalMu',
+    '\uf827': 'FormalCapitalN',
+    '\uf840': 'FormalCapitalNu',
+    '\uf828': 'FormalCapitalO',
+    '\uf84c': 'FormalCapitalOmega',
+    '\uf842': 'FormalCapitalOmicron',
+    '\uf829': 'FormalCapitalP',
+    '\uf849': 'FormalCapitalPhi',
+    '\uf843': 'FormalCapitalPi',
+    '\uf84b': 'FormalCapitalPsi',
+    '\uf82a': 'FormalCapitalQ',
+    '\uf82b': 'FormalCapitalR',
+    '\uf844': 'FormalCapitalRho',
+    '\uf82c': 'FormalCapitalS',
+    '\uf883': 'FormalCapitalSampi',
+    '\uf846': 'FormalCapitalSigma',
+    '\uf87d': 'FormalCapitalStigma',
+    '\uf82d': 'FormalCapitalT',
+    '\uf847': 'FormalCapitalTau',
+    '\uf83b': 'FormalCapitalTheta',
+    '\uf82e': 'FormalCapitalU',
+    '\uf848': 'FormalCapitalUpsilon',
+    '\uf82f': 'FormalCapitalV',
+    '\uf830': 'FormalCapitalW',
+    '\uf831': 'FormalCapitalX',
+    '\uf841': 'FormalCapitalXi',
+    '\uf832': 'FormalCapitalY',
+    '\uf833': 'FormalCapitalZ',
+    '\uf839': 'FormalCapitalZeta',
+    '\uf86a': 'FormalChi',
+    '\uf875': 'FormalCurlyCapitalUpsilon',
+    '\uf858': 'FormalCurlyEpsilon',
+    '\uf885': 'FormalCurlyKappa',
+    '\uf869': 'FormalCurlyPhi',
+    '\uf879': 'FormalCurlyPi',
+    '\uf886': 'FormalCurlyRho',
+    '\uf874': 'FormalCurlyTheta',
+    '\uf803': 'FormalD',
+    '\uf857': 'FormalDelta',
+    '\uf880': 'FormalDigamma',
+    '\uf804': 'FormalE',
+    '\uf88a': 'FormalEpsilon',
+    '\uf85a': 'FormalEta',
+    '\uf805': 'FormalF',
+    '\uf865': 'FormalFinalSigma',
+    '\uf806': 'FormalG',
+    '\uf856': 'FormalGamma',
+    '\uf807': 'FormalH',
+    '\uf808': 'FormalI',
+    '\uf85c': 'FormalIota',
+    '\uf809': 'FormalJ',
+    '\uf80a': 'FormalK',
+    '\uf85d': 'FormalKappa',
+    '\uf882': 'FormalKoppa',
+    '\uf80b': 'FormalL',
+    '\uf85e': 'FormalLambda',
+    '\uf80c': 'FormalM',
+    '\uf85f': 'FormalMu',
+    '\uf80d': 'FormalN',
+    '\uf860': 'FormalNu',
+    '\uf80e': 'FormalO',
+    '\uf86c': 'FormalOmega',
+    '\uf862': 'FormalOmicron',
+    '\uf80f': 'FormalP',
+    '\uf878': 'FormalPhi',
+    '\uf863': 'FormalPi',
+    '\uf86b': 'FormalPsi',
+    '\uf810': 'FormalQ',
+    '\uf811': 'FormalR',
+    '\uf864': 'FormalRho',
+    '\uf812': 'FormalS',
+    '\uf884': 'FormalSampi',
+    '\uf866': 'FormalSigma',
+    '\uf87e': 'FormalStigma',
+    '\uf813': 'FormalT',
+    '\uf867': 'FormalTau',
+    '\uf85b': 'FormalTheta',
+    '\uf814': 'FormalU',
+    '\uf868': 'FormalUpsilon',
+    '\uf815': 'FormalV',
+    '\uf816': 'FormalW',
+    '\uf817': 'FormalX',
+    '\uf861': 'FormalXi',
+    '\uf818': 'FormalY',
+    '\uf819': 'FormalZ',
+    '\uf859': 'FormalZeta',
+    '\uf721': 'FreakedSmiley',
+    '\uf4a1': 'Function',
+    'γ': 'Gamma',
+    '♊': 'GeminiSign',
+    'ℷ': 'Gimel',
+    '\uf6cc': 'GothicA',
+    '\uf6cd': 'GothicB',
+    '\uf6ce': 'GothicC',
+    '\uf78a': 'GothicCapitalA',
+    '\uf78b': 'GothicCapitalB',
+    'ℭ': 'GothicCapitalC',
+    '\uf78d': 'GothicCapitalD',
+    '\uf78e': 'GothicCapitalE',
+    '\uf78f': 'GothicCapitalF',
+    '\uf790': 'GothicCapitalG',
+    'ℌ': 'GothicCapitalH',
+    'ℑ': 'GothicCapitalI',
+    '\uf793': 'GothicCapitalJ',
+    '\uf794': 'GothicCapitalK',
+    '\uf795': 'GothicCapitalL',
+    '\uf796': 'GothicCapitalM',
+    '\uf797': 'GothicCapitalN',
+    '\uf798': 'GothicCapitalO',
+    '\uf799': 'GothicCapitalP',
+    '\uf79a': 'GothicCapitalQ',
+    'ℜ': 'GothicCapitalR',
+    '\uf79c': 'GothicCapitalS',
+    '\uf79d': 'GothicCapitalT',
+    '\uf79e': 'GothicCapitalU',
+    '\uf79f': 'GothicCapitalV',
+    '\uf7a0': 'GothicCapitalW',
+    '\uf7a1': 'GothicCapitalX',
+    '\uf7a2': 'GothicCapitalY',
+    'ℨ': 'GothicCapitalZ',
+    '\uf6cf': 'GothicD',
+    '\uf6d0': 'GothicE',
+    '\uf7ed': 'GothicEight',
+    '\uf6d1': 'GothicF',
+    '\uf7ea': 'GothicFive',
+    '\uf7e9': 'GothicFour',
+    '\uf6d2': 'GothicG',
+    '\uf6d3': 'GothicH',
+    '\uf6d4': 'GothicI',
+    '\uf6d5': 'GothicJ',
+    '\uf6d6': 'GothicK',
+    '\uf6d7': 'GothicL',
+    '\uf6d8': 'GothicM',
+    '\uf6d9': 'GothicN',
+    '\uf7ef': 'GothicNine',
+    '\uf6da': 'GothicO',
+    '\uf7e6': 'GothicOne',
+    '\uf6db': 'GothicP',
+    '\uf6dc': 'GothicQ',
+    '\uf6dd': 'GothicR',
+    '\uf6de': 'GothicS',
+    '\uf7ec': 'GothicSeven',
+    '\uf7eb': 'GothicSix',
+    '\uf6df': 'GothicT',
+    '\uf7e8': 'GothicThree',
+    '\uf7e7': 'GothicTwo',
+    '\uf6e0': 'GothicU',
+    '\uf6e1': 'GothicV',
+    '\uf6e2': 'GothicW',
+    '\uf6e3': 'GothicX',
+    '\uf6e4': 'GothicY',
+    '\uf6e5': 'GothicZ',
+    '\uf7e5': 'GothicZero',
+    '\uf753': 'GrayCircle',
+    '\uf752': 'GraySquare',
+    '⋛': 'GreaterEqualLess',
+    '≥': 'GreaterEqual',
+    '≧': 'GreaterFullEqual',
+    '≫': 'GreaterGreater',
+    '≷': 'GreaterLess',
+    '⩾': 'GreaterSlantEqual',
+    '≳': 'GreaterTilde',
+    'ˇ': 'Hacek',
+    '☺': 'HappySmiley',
+    'ℏ': 'HBar',
+    '♡': 'HeartSuit',
+    '\uf3ce': 'HermitianConjugate',
+    '─': 'HorizontalLine',
+    '≎': 'HumpDownHump',
+    '≏': 'HumpEqual',
+    '‐': 'Hyphen',
+    'í': 'IAcute',
+    'ĭ': 'ICup',
+    'ï': 'IDoubleDot',
+    'ì': 'IGrave',
+    'î': 'IHat',
+    '\uf74e': 'ImaginaryI',
+    '\uf74f': 'ImaginaryJ',
+    '\uf39e': 'ImplicitPlus',
+    '\uf523': 'Implies',
+    '∞': 'Infinity',
+    '∫': 'Integral',
+    '⋂': 'Intersection',
+    '\uf76d': 'InvisibleApplication',
+    '\uf765': 'InvisibleComma',
+    '\uf3b4': 'InvisiblePostfixScriptBase',
+    '\uf3b3': 'InvisiblePrefixScriptBase',
+    '\uf360': 'InvisibleSpace',
+    '\u2062': 'InvisibleTimes',
+    'ι': 'Iota',
+    '♃': 'Jupiter',
+    'κ': 'Kappa',
+    '\uf756': 'KernelIcon',
+    'ϟ': 'Koppa',
+    'λ': 'Lambda',
+    '\uf7fb': 'LastPage',
+    '〈': 'LeftAngleBracket',
+    '⇤': 'LeftArrowBar',
+    '←': 'LeftArrow',
+    '⇆': 'LeftArrowRightArrow',
+    '\uf113': 'LeftAssociation',
+    '\uf603': 'LeftBracketingBar',
+    '⌈': 'LeftCeiling',
+    '〚': 'LeftDoubleBracket',
+    '\uf605': 'LeftDoubleBracketingBar',
+    '⥡': 'LeftDownTeeVector',
+    '⥙': 'LeftDownVectorBar',
+    '⇃': 'LeftDownVector',
+    '⌊': 'LeftFloor',
+    '«': 'LeftGuillemet',
+    '\uf76b': 'LeftModified',
+    '◂': 'LeftPointer',
+    '↔': 'LeftRightArrow',
+    '⥎': 'LeftRightVector',
+    '\uf761': 'LeftSkeleton',
+    '⊣': 'LeftTee',
+    '↤': 'LeftTeeArrow',
+    '⥚': 'LeftTeeVector',
+    '⊲': 'LeftTriangle',
+    '⧏': 'LeftTriangleBar',
+    '⊴': 'LeftTriangleEqual',
+    '⥑': 'LeftUpDownVector',
+    '⥠': 'LeftUpTeeVector',
+    '↿': 'LeftUpVector',
+    '⥘': 'LeftUpVectorBar',
+    '↼': 'LeftVector',
+    '⥒': 'LeftVectorBar',
+    '♌': 'LeoSign',
+    '≤': 'LessEqual',
+    '⋚': 'LessEqualGreater',
+    '≦': 'LessFullEqual',
+    '≶': 'LessGreater',
+    '≪': 'LessLess',
+    '⩽': 'LessSlantEqual',
+    '≲': 'LessTilde',
+    '\uf754': 'LetterSpace',
+    '♎': 'LibraSign',
+    '\uf723': 'LightBulb',
+    '—': 'LongDash',
+    '\uf7d9': 'LongEqual',
+    '⟵': 'LongLeftArrow',
+    '⟷': 'LongLeftRightArrow',
+    '⟶': 'LongRightArrow',
+    '↙': 'LowerLeftArrow',
+    '↘': 'LowerRightArrow',
+    'ł': 'LSlash',
+    '♂': 'Mars',
+    '\uf757': 'MathematicaIcon',
+    '∡': 'MeasuredAngle',
+    '\u205f': 'MediumSpace',
+    '☿': 'Mercury',
+    '℧': 'Mho',
+    'µ': 'Micro',
+    '∓': 'MinusPlus',
+    'μ': 'Mu',
+    '⊼': 'Nand',
+    '♮': 'Natural',
+    '\uf383': 'NegativeMediumSpace',
+    '\uf384': 'NegativeThickSpace',
+    '\uf382': 'NegativeThinSpace',
+    '\uf380': 'NegativeVeryThinSpace',
+    '♆': 'Neptune',
+    '⪢': 'NestedGreaterGreater',
+    '⪡': 'NestedLessLess',
+    '\uf722': 'NeutralSmiley',
+    'ň': 'NHacek',
+    '\u2060': 'NoBreak',
+    '\xa0': 'NonBreakingSpace',
+    '⊽': 'Nor',
+    '≢': 'NotCongruent',
+    '≭': 'NotCupCap',
+    '∦': 'NotDoubleVerticalBar',
+    '∉': 'NotElement',
+    '≠': 'NotEqual',
+    '\uf400': 'NotEqualTilde',
+    '∄': 'NotExists',
+    '≯': 'NotGreater',
+    '≱': 'NotGreaterEqual',
+    '≩': 'NotGreaterFullEqual',
+    '\uf427': 'NotGreaterGreater',
+    '≹': 'NotGreaterLess',
+    '\uf429': 'NotGreaterSlantEqual',
+    '≵': 'NotGreaterTilde',
+    '\uf402': 'NotHumpDownHump',
+    '\uf401': 'NotHumpEqual',
+    '⋪': 'NotLeftTriangle',
+    '\uf412': 'NotLeftTriangleBar',
+    '⋬': 'NotLeftTriangleEqual',
+    '≰': 'NotLessEqual',
+    '≨': 'NotLessFullEqual',
+    '≸': 'NotLessGreater',
+    '≮': 'NotLess',
+    '\uf422': 'NotLessLess',
+    '\uf424': 'NotLessSlantEqual',
+    '≴': 'NotLessTilde',
+    '¬': 'Not',
+    '\uf428': 'NotNestedGreaterGreater',
+    '\uf423': 'NotNestedLessLess',
+    '⊀': 'NotPrecedes',
+    '\uf42b': 'NotPrecedesEqual',
+    '⋠': 'NotPrecedesSlantEqual',
+    '⋨': 'NotPrecedesTilde',
+    '∌': 'NotReverseElement',
+    '⋫': 'NotRightTriangle',
+    '\uf413': 'NotRightTriangleBar',
+    '⋭': 'NotRightTriangleEqual',
+    '\uf42e': 'NotSquareSubset',
+    '⋢': 'NotSquareSubsetEqual',
+    '\uf42f': 'NotSquareSuperset',
+    '⋣': 'NotSquareSupersetEqual',
+    '⊄': 'NotSubset',
+    '⊈': 'NotSubsetEqual',
+    '⊁': 'NotSucceeds',
+    '\uf42d': 'NotSucceedsEqual',
+    '⋡': 'NotSucceedsSlantEqual',
+    '⋩': 'NotSucceedsTilde',
+    '⊅': 'NotSuperset',
+    '⊉': 'NotSupersetEqual',
+    '≁': 'NotTilde',
+    '≄': 'NotTildeEqual',
+    '≇': 'NotTildeFullEqual',
+    '≉': 'NotTildeTilde',
+    '\uf3d1': 'NotVerticalBar',
+    'ñ': 'NTilde',
+    'ν': 'Nu',
+    '\uf3a0': 'Null',
+    '\uf724': 'NumberSign',
+    'ó': 'OAcute',
+    'ő': 'ODoubleAcute',
+    'ö': 'ODoubleDot',
+    'œ': 'OE',
+    'ò': 'OGrave',
+    'ô': 'OHat',
+    'ω': 'Omega',
+    'ο': 'Omicron',
+    '“': 'OpenCurlyDoubleQuote',
+    '‘': 'OpenCurlyQuote',
+    '\uf7d2': 'OptionKey',
+    '∨': 'Or',
+    'ø': 'OSlash',
+    'õ': 'OTilde',
+    '︷': 'OverBrace',
+    '⎴': 'OverBracket',
+    '︵': 'OverParenthesis',
+    '¶': 'Paragraph',
+    '∂': 'PartialD',
+    '\uf3de': 'PermutationProduct',
+    '⟂': 'Perpendicular',
+    'ϕ': 'Phi',
+    'π': 'Pi',
+    '\uf361': 'Piecewise',
+    '♓': 'PiscesSign',
+    '\uf528': 'Placeholder',
+    '±': 'PlusMinus',
+    '♇': 'Pluto',
+    '≺': 'Precedes',
+    '⪯': 'PrecedesEqual',
+    '≼': 'PrecedesSlantEqual',
+    '≾': 'PrecedesTilde',
+    '′': 'Prime',
+    '∏': 'Product',
+    '∷': 'Proportion',
+    '∝': 'Proportional',
+    'ψ': 'Psi',
+    '♩': 'QuarterNote',
+    '®': 'RegisteredTrademark',
+    '↵': 'ReturnIndicator',
+    '\uf766': 'ReturnKey',
+    '‶': 'ReverseDoublePrime',
+    '∋': 'ReverseElement',
+    '⇋': 'ReverseEquilibrium',
+    '‵': 'ReversePrime',
+    '⥯': 'ReverseUpEquilibrium',
+    'ř': 'RHacek',
+    'ρ': 'Rho',
+    '∟': 'RightAngle',
+    '〉': 'RightAngleBracket',
+    '→': 'RightArrow',
+    '⇥': 'RightArrowBar',
+    '⇄': 'RightArrowLeftArrow',
+    '\uf114': 'RightAssociation',
+    '\uf604': 'RightBracketingBar',
+    '⌉': 'RightCeiling',
+    '〛': 'RightDoubleBracket',
+    '\uf606': 'RightDoubleBracketingBar',
+    '⥝': 'RightDownTeeVector',
+    '⇂': 'RightDownVector',
+    '⥕': 'RightDownVectorBar',
+    '⌋': 'RightFloor',
+    '»': 'RightGuillemet',
+    '\uf76c': 'RightModified',
+    '▸': 'RightPointer',
+    '\uf762': 'RightSkeleton',
+    '⊢': 'RightTee',
+    '↦': 'RightTeeArrow',
+    '⥛': 'RightTeeVector',
+    '⊳': 'RightTriangle',
+    '⧐': 'RightTriangleBar',
+    '⊵': 'RightTriangleEqual',
+    '⥏': 'RightUpDownVector',
+    '⥜': 'RightUpTeeVector',
+    '↾': 'RightUpVector',
+    '⥔': 'RightUpVectorBar',
+    '⇀': 'RightVector',
+    '⥓': 'RightVectorBar',
+    '⥰': 'RoundImplies',
+    '\uf3b2': 'RoundSpaceIndicator',
+    '\uf522': 'Rule',
+    '\uf51f': 'RuleDelayed',
+    '☹': 'SadSmiley',
+    '♐': 'SagittariusSign',
+    'Ϡ': 'Sampi',
+    '♄': 'Saturn',
+    '♏': 'ScorpioSign',
+    '\uf6b2': 'ScriptA',
+    '\uf6b3': 'ScriptB',
+    '\uf6b4': 'ScriptC',
+    '\uf770': 'ScriptCapitalA',
+    'ℬ': 'ScriptCapitalB',
+    '\uf772': 'ScriptCapitalC',
+    '\uf773': 'ScriptCapitalD',
+    'ℰ': 'ScriptCapitalE',
+    'ℱ': 'ScriptCapitalF',
+    '\uf776': 'ScriptCapitalG',
+    'ℋ': 'ScriptCapitalH',
+    'ℐ': 'ScriptCapitalI',
+    '\uf779': 'ScriptCapitalJ',
+    '\uf77a': 'ScriptCapitalK',
+    'ℒ': 'ScriptCapitalL',
+    'ℳ': 'ScriptCapitalM',
+    '\uf77d': 'ScriptCapitalN',
+    '\uf77e': 'ScriptCapitalO',
+    '℘': 'ScriptCapitalP',
+    '\uf780': 'ScriptCapitalQ',
+    'ℛ': 'ScriptCapitalR',
+    '\uf782': 'ScriptCapitalS',
+    '\uf783': 'ScriptCapitalT',
+    '\uf784': 'ScriptCapitalU',
+    '\uf785': 'ScriptCapitalV',
+    '\uf786': 'ScriptCapitalW',
+    '\uf787': 'ScriptCapitalX',
+    '\uf788': 'ScriptCapitalY',
+    '\uf789': 'ScriptCapitalZ',
+    '\uf6b5': 'ScriptD',
+    '\uf730': 'ScriptDotlessI',
+    '\uf731': 'ScriptDotlessJ',
+    'ℯ': 'ScriptE',
+    '\uf7f8': 'ScriptEight',
+    '\uf6b7': 'ScriptF',
+    '\uf7f5': 'ScriptFive',
+    '\uf7f4': 'ScriptFour',
+    'ℊ': 'ScriptG',
+    '\uf6b9': 'ScriptH',
+    '\uf6ba': 'ScriptI',
+    '\uf6bb': 'ScriptJ',
+    '\uf6bc': 'ScriptK',
+    'ℓ': 'ScriptL',
+    '\uf6be': 'ScriptM',
+    '\uf6bf': 'ScriptN',
+    '\uf7f9': 'ScriptNine',
+    'ℴ': 'ScriptO',
+    '\uf7f1': 'ScriptOne',
+    '\uf6c1': 'ScriptP',
+    '\uf6c2': 'ScriptQ',
+    '\uf6c3': 'ScriptR',
+    '\uf6c4': 'ScriptS',
+    '\uf7f7': 'ScriptSeven',
+    '\uf7f6': 'ScriptSix',
+    '\uf6c5': 'ScriptT',
+    '\uf7f3': 'ScriptThree',
+    '\uf7f2': 'ScriptTwo',
+    '\uf6c6': 'ScriptU',
+    '\uf6c7': 'ScriptV',
+    '\uf6c8': 'ScriptW',
+    '\uf6c9': 'ScriptX',
+    '\uf6ca': 'ScriptY',
+    '\uf6cb': 'ScriptZ',
+    '\uf7f0': 'ScriptZero',
+    '§': 'Section',
+    '\uf527': 'SelectionPlaceholder',
+    'š': 'SHacek',
+    '♯': 'Sharp',
+    '\uf52b': 'ShortDownArrow',
+    '\uf526': 'ShortLeftArrow',
+    '\uf525': 'ShortRightArrow',
+    '\uf52a': 'ShortUpArrow',
+    'σ': 'Sigma',
+    '✶': 'SixPointedStar',
+    '⁃': 'SkeletonIndicator',
+    '∘': 'SmallCircle',
+    '␣': 'SpaceIndicator',
+    '\uf7bf': 'SpaceKey',
+    '♠': 'SpadeSuit',
+    '\uf3bb': 'SpanFromAbove',
+    '\uf3bc': 'SpanFromBoth',
+    '\uf3ba': 'SpanFromLeft',
+    '∢': 'SphericalAngle',
+    '√': 'Sqrt',
+    '\uf520': 'Square',
+    '⊓': 'SquareIntersection',
+    '⊏': 'SquareSubset',
+    '⊑': 'SquareSubsetEqual',
+    '⊐': 'SquareSuperset',
+    '⊒': 'SquareSupersetEqual',
+    '⊔': 'SquareUnion',
+    '⋆': 'Star',
+    '£': 'Sterling',
+    'ϛ': 'Stigma',
+    '⊂': 'Subset',
+    '⊆': 'SubsetEqual',
+    '≻': 'Succeeds',
+    '⪰': 'SucceedsEqual',
+    '≽': 'SucceedsSlantEqual',
+    '≿': 'SucceedsTilde',
+    '∍': 'SuchThat',
+    '∑': 'Sum',
+    '⊃': 'Superset',
+    '⊇': 'SupersetEqual',
+    '\uf75f': 'SystemEnterKey',
+    '\uf3af': 'SystemsModelDelay',
+    'ß': 'SZ',
+    '\uf7be': 'TabKey',
+    'τ': 'Tau',
+    '♉': 'TaurusSign',
+    '\uf3da': 'TensorProduct',
+    '\uf3db': 'TensorWedge',
+    'ť': 'THacek',
+    '∴': 'Therefore',
+    'θ': 'Theta',
+    '\u2005': 'ThickSpace',
+    '\u2009': 'ThinSpace',
+    'þ': 'Thorn',
+    '∼': 'Tilde',
+    '≃': 'TildeEqual',
+    '≅': 'TildeFullEqual',
+    '≈': 'TildeTilde',
+    '×': 'Times',
+    '™': 'Trademark',
+    '\uf3c7': 'Transpose',
+    '\uf758': 'TripleDot',
+    'ú': 'UAcute',
+    'ű': 'UDoubleAcute',
+    'ü': 'UDoubleDot',
+    'ù': 'UGrave',
+    'û': 'UHat',
+    '︸': 'UnderBrace',
+    '⎵': 'UnderBracket',
+    '︶': 'UnderParenthesis',
+    '\uf3d4': 'UndirectedEdge',
+    '⋃': 'Union',
+    '⊎': 'UnionPlus',
+    '↑': 'UpArrow',
+    '⤒': 'UpArrowBar',
+    '⇅': 'UpArrowDownArrow',
+    '↕': 'UpDownArrow',
+    '⥮': 'UpEquilibrium',
+    '↖': 'UpperLeftArrow',
+    '↗': 'UpperRightArrow',
+    '▴': 'UpPointer',
+    'υ': 'Upsilon',
+    '⊥': 'UpTee',
+    '↥': 'UpTeeArrow',
+    '♅': 'Uranus',
+    'ů': 'URing',
+    '⋁': 'Vee',
+    '♀': 'Venus',
+    '\uf3d0': 'VerticalBar',
+    '⋮': 'VerticalEllipsis',
+    '│': 'VerticalLine',
+    '\uf432': 'VerticalSeparator',
+    '≀': 'VerticalTilde',
+    '\u200a': 'VeryThinSpace',
+    '♍': 'VirgoSign',
+    '\uf725': 'WarningSign',
+    '⌚': 'WatchIcon',
+    '⋀': 'Wedge',
+    '℘': 'WeierstrassP',
+    '♗': 'WhiteBishop',
+    '♔': 'WhiteKing',
+    '♘': 'WhiteKnight',
+    '♙': 'WhitePawn',
+    '♕': 'WhiteQueen',
+    '♖': 'WhiteRook',
+    '\uf720': 'Wolf',
+    '\uf11e': 'WolframLanguageLogo',
+    '\uf11f': 'WolframLanguageLogoCircle',
+    'ξ': 'Xi',
+    '\uf4a2': 'Xnor',
+    '⊻': 'Xor',
+    'ý': 'YAcute',
+    'ÿ': 'YDoubleDot',
+    '¥': 'Yen',
+    'ζ': 'Zeta',
+    'ž': 'ZHacek',
+}
+
+# Combine the two dictionaries and exclude entries whose key maps to itself.
+# The exclusion of those entries is just an optmization (the regex patter 
+# matcher won't have to look for those).
+WL_TO_UNICODE_DICT = {k: f"\\[{v}]" for k, v in WL_TO_NAMED.items()}
+WL_TO_UNICODE_DICT.update(WL_TO_UNICODE)
+WL_TO_UNICODE_DICT = {re.escape(k): v 
+                     for k, v in WL_TO_UNICODE_DICT.items() if k != v}
+
+# The keys should be sorted to prevent shorter keys from obscuring longer keys 
+# when pattern matching
+WL_TO_UNICODE_RE = re.compile(
+    "|".join(sorted(WL_TO_UNICODE_DICT.keys(), key=lambda k: (-len(k), k)))
+)
+
+WL_TO_PLAIN_DICT = {re.escape(k): v for k, v in WL_TO_NAMED.items()}
+
+# The keys should be sorted to prevent shorter keys from obscuring longer keys 
+# when pattern matching
+WL_TO_PLAIN_RE = re.compile(
+    "|".join(sorted(WL_TO_PLAIN_DICT.keys(), key=lambda k: (-len(k), k)))
+)
+
+UNICODE_REPLACE_DICT = {re.escape(k): v for k, v in UNICODE_TO_WL.items()}
+
+# The keys should be sorted to prevent shorter keys from obscuring longer keys 
+# when pattern matching
+UNICODE_REPLACE_RE = re.compile(
+    "|".join(sorted(UNICODE_REPLACE_DICT.keys(), key=lambda k: (-len(k), k)))
+)
+# All supported named characters
+named_characters = {v: k for k, v in WL_TO_NAMED.items()}
+
+def replace_wl_with_plain_text(wl_input: str, use_unicode=True) -> str:
+    """WL uses some non-unicode character for various things.
+    Replace them with the unicode equivalent.
+    """
+    r = WL_TO_UNICODE_RE if use_unicode else WL_TO_PLAIN_RE
+    d = WL_TO_UNICODE_DICT if use_unicode else WL_TO_PLAIN_DICT
+
+    return r.sub(lambda m: d[re.escape(m.group(0))], wl_input)
+
+def replace_unicode_with_wl(unicode_input: str) -> str:
+    """WL uses some non-unicode character for various things.
+    Replace their unicode equivalent with them.
+    """
+    return UNICODE_REPLACE_RE.sub(
+        lambda m: UNICODE_REPLACE_DICT[re.escape(m.group(0))], unicode_input
+    )
+
+
